AC Optimal Power Flow Solution

For further information on this topic, please see the AC Optimal Power Flow section of the Manual.

API Index

Build Model

acOptimalPowerFlow

Build Model

JuliaGrid.acOptimalPowerFlow — Function

acOptimalPowerFlow(system::PowerSystem, optimizer; bridge, name, balance, limit,
    rating, capability)

The function takes the PowerSystem composite type as input to establish the structure for solving the AC optimal power flow. The optimizer argument is also required to create and solve the optimization problem. If the acModel field within the PowerSystem composite type has not been created, the function will initiate an update automatically.

Keywords

JuliaGrid offers the ability to manipulate the jump model based on the guidelines provided in the JuMP documentation. However, certain configurations may require different method calls, such as:

bridge: used to manage the bridging mechanism

name: used to manage the creation of string names.

Moreover, we have included keywords that regulate the usage of different types of constraints:

balance: controls the equality constraints that relate to the active and reactive power balance equations

limit: controls the inequality constraints that relate to the voltage magnitude and angle differences between buses

rating: controls the inequality constraints that relate to the long-term rating of branches

capability: controls the inequality constraints that relate to the active and reactive power generator outputs.

By default, all of these keywords are set to true and are of the Bool type.

JuMP

The JuliaGrid builds the AC optimal power flow around the JuMP package and supports commonly used solvers. For more information, refer to the JuMP documenatation.

Returns

The function returns an instance of the ACOptimalPowerFlow type, which includes the following fields:

voltage: the bus voltage magnitudes and angles

output: the output active and reactive powers of each generator

jump: the JuMP model

constraint: holds the constraint references to the JuMP model.

Examples

Create the complete DC optimal power flow model:

system = powerSystem("case14.h5")
dcModel!(system)

model = acOptimalPowerFlow(system, HiGHS.Optimizer)

Create the DC optimal power flow model without rating constraints:

system = powerSystem("case14.h5")
dcModel!(system)

model = acOptimalPowerFlow(system, HiGHS.Optimizer; rating = false)