AC Power Flow

For further information on this topic, please see the Power System Analysis section of the Manual.

API Index

Newton Raphson Method

newtonRaphson

mismatch!

solve!

Fast Newton Raphson Method

fastNewtonRaphsonBX

fastNewtonRaphsonXB

mismatch!

solve!

Gauss-Seidel Method

gaussSeidel

mismatch!

solve!

Power and Current Analysis

analysisBus

analysisBranch

analysisGenerator

Additional Functions

reactiveLimit!

adjustAngle!

Newton Raphson Method

JuliaGrid.newtonRaphson — Function

newtonRaphson(system::PowerSystem)

The function accepts the PowerSystem composite type as input and uses it to set up the Newton-Raphson method to solve AC power flow. Additionally, if the AC model was not created, the function will automatically initiate an update of the acModel field within the PowerSystem composite type.

Returns

The function returns an instance of the NewtonRaphson subtype of the abstract ACPowerFlow type, which includes the following fields:

voltage: the magnitudes and angles of bus voltages

jacobian: the Jacobian matrix

mismatch: the active and reactive power injection mismatches

increment: the magnitudes and angles of bus voltage increments

pq: indices of demand buses

pvpq: indices of demand and generator buses.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)

JuliaGrid.mismatch! — Method

mismatch!(system::PowerSystem, model::NewtonRaphson)

The function calculates both active and reactive power injection mismatches and returns their maximum absolute values. These values can be used to terminate the iteration loop of the Newton-Raphson method, which is utilized to solve the AC power flow problem.

This function updates the mismatch variables in the Newton-Raphson method and should be called during the iteration loop before invoking the solve! function.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
mismatch!(system, model)

JuliaGrid.solve! — Method

solve!(system::PowerSystem, model::NewtonRaphson)

The function solves the AC power flow problem and computes the magnitudes and angles of bus voltages using the Newton-Raphson method.

Before calling this function, the mismatch! function should be executed to update the mismatch variables. The computed voltages are stored in the voltage field of the respective struct type.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

Fast Newton Raphson Method

JuliaGrid.fastNewtonRaphsonBX — Function

fastNewtonRaphsonBX(system::PowerSystem)

The function accepts the PowerSystem composite type as input and uses it to set up the Fast Newton-Raphson method of version BX to solve AC power flow. Additionally, if the AC model was not created, the function will automatically initiate an update of the acModel field within the PowerSystem composite type.

Returns

The function returns an instance of the FastNewtonRaphson subtype of the abstract ACPowerFlow type, which includes the following fields:

voltage: the magnitudes and angles of bus voltages

active:

jacobian: the Jacobian matrix associated with active power equations

mismatch: the active power injection mismatches

increment: the angles of bus voltage increments

factorization: the factorized Jacobian matrix

reactive:

jacobian: the Jacobian matrix associated with reactive power equations

mismatch: the reative power injection mismatches

increment: the magnitudes of bus voltage increments

factorization: the factorized Jacobian matrix

pq: indices of demand buses

pvpq: indices of demand and generator buses.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = fastNewtonRaphsonBX(system)

JuliaGrid.fastNewtonRaphsonXB — Function

fastNewtonRaphsonXB(system::PowerSystem)

The function accepts the PowerSystem composite type as input and uses it to set up the Fast Newton-Raphson method of version XB to solve AC power flow. Additionally, if the AC model was not created, the function will automatically initiate an update of the acModel field within the PowerSystem composite type.

Returns

The function returns an instance of the FastNewtonRaphson subtype of the abstract ACPowerFlow type, which includes the following fields:

voltage: the magnitudes and angles of bus voltages

active:

jacobian: the Jacobian matrix associated with active power equations

mismatch: the active power injection mismatches

increment: the angles of bus voltage increments

factorization: the factorized Jacobian matrix

reactive:

jacobian: the Jacobian matrix associated with reactive power equations

mismatch: the reative power injection mismatches

increment: the magnitudes of bus voltage increments

factorization: the factorized Jacobian matrix

pq: indices of demand buses

pvpq: indices of demand and generator buses.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = fastNewtonRaphsonXB(system)

JuliaGrid.mismatch! — Method

mismatch!(system::PowerSystem, model::FastNewtonRaphson)

The function calculates both active and reactive power injection mismatches and returns their maximum absolute values. These values can be used to terminate the iteration loop of the fast Newton-Raphson method, which is utilized to solve the AC power flow problem.

This function updates the mismatch variables in the fast Newton-Raphson method and should be called during the iteration loop before invoking the solve! function.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = fastNewtonRaphsonXB(system)
mismatch!(system, model)

JuliaGrid.solve! — Method

solve!(system::PowerSystem, model::FastNewtonRaphson)

The function solves the AC power flow problem and computes the magnitudes and angles of bus voltages using the fast Newton-Raphson method.

Before calling this function, the mismatch! function should be executed to update the mismatch variables. The computed voltages are stored in the voltage field of the respective struct type.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = fastNewtonRaphsonXB(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

Gauss-Seidel Method

JuliaGrid.gaussSeidel — Function

gaussSeidel(system::PowerSystem)

The function accepts the PowerSystem composite type as input and uses it to set up the Gauss-Seidel method to solve AC power flow. Additionally, if the AC model was not created, the function will automatically initiate an update of the acModel field within the PowerSystem composite type.

Returns

The function returns an instance of the GaussSeidel subtype of the abstract ACPowerFlow type, which includes the following fields:

voltage: the magnitudes and angles of bus voltages

complex: the complex voltages

magnitude: the bus voltage magnitudes for corrections

pq: indices of demand buses

pv: indices of generator buses.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = gaussSeidel(system)

JuliaGrid.mismatch! — Method

mismatch!(system::PowerSystem, model::GaussSeidel)

The Gauss-Seidel method does not need mismatches to obtain bus voltages, but the maximum absolute values are commonly employed to stop the iteration loop. The function does not save any data and should be utilized during the iteration loop before invoking the solve! function.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = gaussSeidel(system)
mismatch!(system, model)

JuliaGrid.solve! — Method

solve!(system::PowerSystem, model::GaussSeidel)

The function solves the AC power flow problem and computes the magnitudes and angles of bus voltages using the Gauss-Seidel method.

The computed voltages are stored in the voltage field of the respective struct type.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = gaussSeidel(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

Power and Current Analysis

Missing docstring.

Missing docstring for analysisBus(::PowerSystem, ::ACPowerFlow). Check Documenter's build log for details.

Missing docstring.

Missing docstring for analysisBranch(::PowerSystem, ::ACPowerFlow). Check Documenter's build log for details.

Missing docstring.

Missing docstring for analysisGenerator(::PowerSystem, ::ACPowerFlow). Check Documenter's build log for details.

Additional Functions

JuliaGrid.reactiveLimit! — Function

reactiveLimit!(system::PowerSystem, model::ACPowerFlow, power::PowerGenerator)

The function verifies whether the generators in a power system exceed their reactive power limits. This is done by setting the reactive power of the generators to within the limits if they are violated, after determining the bus voltage magnitudes and angles. If the limits are violated, the corresponding generator buses or the slack bus are converted to demand buses.

Arguments

Initially, the analysisGenerator function must be executed. Afterward, the function uses the results from this function to assign values to the generator.output.active and bus.supply.active fields of the PowerSystem type.

At the end of the process, the function inspects the reactive powers of the generator and adjusts them to their maximum or minimum values if they violate the threshold. The generator.output.reactive field of the PowerSystem type is then modified accordingly. In light of this modification, the bus.supply.reactive field of the PowerSystem type is also updated, and the bus types in bus.layout.type are adjusted. If the slack bus is converted, the bus.layout.slack field is modified accordingly.

Returns

The function returns the violate variable to indicate which buses violate the limits, with -1 indicating a violation of the minimum limits and 1 indicating a violation of the maximum limits.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end
power = analysisGenerator(system, model)

violate = reactiveLimit!(system, model, power)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

JuliaGrid.adjustAngle! — Function

adjustAngle!(system::PowerSystem, model::ACPowerFlow; slack)

The function modifies the bus voltage angles based on a different slack bus than the one identified by the bus.layout.slack field. This function only updates the voltage.angle variable of the Model composite type.

For instance, if the reactive power of the generator exceeds the limit on the slack bus, the reactiveLimit! function will change that bus to the demand bus and designate the first generator bus in the sequence as the new slack bus. After obtaining the updated AC power flow solution based on the new slack bus, it is possible to adjust the voltage angles to align with the angle of the original slack bus. The slack keyword specifies the bus label of the original slack bus.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end
power = analysisGenerator(system, model)

reactiveLimit!(system, model, power)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

adjustAngle!(system, model; slack = 1)