Power and Current Analysis

In the following section, we have provided a list of functions that can be utilized for post-processing analysis. Once the voltage values are obtained through power flow analysis or optimal power flow analysis, these functions can be used to calculate power or current values. The specific procedures for computing these values depend on the chosen analysis, which are described in separate manuals for further information.

Power Analysis

power!

powerInjection

powerSupply

powerShunt

powerFrom

powerTo

powerCharging

powerSeries

powerGenerator

Current Analysis

current!

currentInjection

currentFrom

currentTo

currentSeries

Power Analysis

JuliaGrid.power! — Method

power!(system::PowerSystem, analysis::AC)

The function computes the active and reactive powers associated with buses, branches, and generators in the AC framework.

Updates

This function updates the power field of the AC abstract type by computing the following electrical quantities:

injection: active and reactive power bus injections;

supply: active and reactive power bus injections from the generators;

shunt: active and reactive power values associated with shunt element at each bus;

from: active and reactive power flows at the "from" end of each branch;

to: active and reactive power flows at the "to" end of each branch;

charging: active and reactive power values linked with branch charging admittances for each branch;

series active and reactive power losses through each branch series impedance;

generator: produced active and reactive power outputs of each generator.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the powers within the AC power flow;

ACOptimalPowerFlow: computes the powers within the AC optimal power flow.

Examples

Compute powers after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
power!(system, analysis)

Compute powers after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
power!(system, analysis)

JuliaGrid.power! — Method

power!(system::PowerSystem, analysis::DC)

The function calculates the active power values related to buses, branches, and generators within the DC analysis framework. It modifies the power field of the abstract type DC.

Updates

This function computes the following electrical quantities:

injection: active power injections at each bus;

supply: active power injections from the generators at each bus;

from: active power flows at each "from" bus end of the branch;

to: active power flows at each "to" bus end of the branch;

generator: output active powers of each generator.

Abstract type

The abstract type DC can have the following subtypes:

DCPowerFlow: computes the powers within the DC power flow;

DCOptimalPowerFlow: computes the powers within the DC optimal power flow.

Examples

Compute powers after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system)
solve!(system, analysis)
power!(system, analysis)

Compute powers after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcOptimalPowerFlow(system)
optimize!(system, analysis)
power!(system, analysis)

JuliaGrid.powerInjection — Method

powerInjection(system::PowerSystem, analysis::AC, label)

The function returns the active and reactive power injections associated with a specific bus in the AC framework. The label keyword argument must match an existing bus label.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the powers within the AC power flow;

ACOptimalPowerFlow: computes the powers within the AC optimal power flow.

Examples

Compute powers after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
injection = powerInjection(system, analysis; label = 1)

Compute powers after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
injection = powerInjection(system, analysis; label = 1)

JuliaGrid.powerInjection — Method

powerInjection(system::PowerSystem, analysis::DC; label)

The function returns the active power injection associated with a specific bus in the DC framework. The label keyword argument must match an existing bus label.

Abstract type

The abstract type DC can have the following subtypes:

DCPowerFlow: computes the power within the DC power flow;

DCOptimalPowerFlow: computes the power within the DC optimal power flow;

Examples

Compute the active power of a specific bus after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system)
solve!(system, analysis)
injection = powerInjection(system, analysis; label = 2)

Compute the active power of a specific bus after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcOptimalPowerFlow(system)
optimize!(system, analysis)
injection = powerInjection(system, analysis; label = 2)

JuliaGrid.powerSupply — Method

powerSupply(system::PowerSystem, analysis::AC, label)

The function returns the active and reactive power injections from the generators associated with a specific bus in the AC framework. The label keyword argument must match an existing bus label.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the powers within the AC power flow;

ACOptimalPowerFlow: computes the powers within the AC optimal power flow.

Examples

Compute powers after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
supply = powerSupply(system, analysis; label = 1)

Compute powers after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
supply = powerSupply(system, analysis; label = 1)

JuliaGrid.powerSupply — Method

powerSupply(system::PowerSystem, analysis::DC; label)

The function returns the active power injection from the generators associated with a specific bus in the DC framework. The label keyword argument must match an existing bus label.

Abstract type

The abstract type DC can have the following subtypes:

DCPowerFlow: computes the power within the DC power flow,

DCOptimalPowerFlow: computes the power within the DC optimal power flow.

Examples

Compute the active power of a specific bus after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system)
solve!(system, analysis)
supply = powerSupply(system, analysis; label = 2)

Compute the active power of a specific bus after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcOptimalPowerFlow(system)
optimize!(system, analysis)
supply = powerSupply(system, analysis; label = 2)

JuliaGrid.powerShunt — Method

powerShunt(system::PowerSystem, analysis::AC, label)

The function returns the active and reactive power values of the shunt element associated with a specific bus in the AC framework. The label keyword argument must match an existing bus label.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the powers within the AC power flow;

ACOptimalPowerFlow: computes the powers within the AC optimal power flow.

Examples

Compute powers after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
supply = powerShunt(system, analysis; label = 1)

Compute powers after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
supply = powerShunt(system, analysis; label = 1)

JuliaGrid.powerFrom — Method

powerFrom(system::PowerSystem, analysis::AC; label)

The function returns the active and reactive power flows at the "from" bus end associated with a specific branch in the AC framework. The label keyword argument must match an existing branch label.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the powers within the AC power flow;

ACOptimalPowerFlow: computes the powers within the AC optimal power flow.

Examples

Compute powers after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
from = powerFrom(system, analysis; label = 2)

Compute powers after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
from = powerFrom(system, analysis; label = 2)

JuliaGrid.powerFrom — Method

powerFrom(system::PowerSystem, analysis::DC; label)

The function returns the active power flow at the "from" bus end associated with a specific branch in the DC framework. The label keyword argument must match an existing branch label.

Abstract type

The abstract type DC can have the following subtypes:

DCPowerFlow: computes the power within the DC power flow;

DCOptimalPowerFlow: computes the power within the DC optimal power flow.

Examples

Compute the active power of a specific branch after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system)
solve!(system, analysis)
from = powerFrom(system, analysis; label = 2)

Compute the active power of a specific branch after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcOptimalPowerFlow(system)
optimize!(system, analysis)
from = powerFrom(system, analysis; label = 2)

JuliaGrid.powerTo — Method

powerTo(system::PowerSystem, analysis::AC; label)

The function returns the active and reactive power flows at the "to" bus end associated with a specific branch in the AC framework. The label keyword argument must match an existing branch label.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the powers within the AC power flow;

ACOptimalPowerFlow: computes the powers within the AC optimal power flow.

Examples

Compute powers after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
to = powerTo(system, analysis; label = 2)

Compute powers after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
to = powerTo(system, analysis; label = 2)

JuliaGrid.powerTo — Method

powerTo(system::PowerSystem, analysis::DC; label)

The function returns the active power flow at the "to" bus end associated with a specific branch in the DC framework. The label keyword argument must match an existing branch label.

Abstract type

The abstract type DC can have the following subtypes:

DCPowerFlow: computes the power within the DC power flow;

DCOptimalPowerFlow: computes the power within the DC optimal power flow.

Examples

Compute the active power of a specific branch after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system)
solve!(system, analysis)
to = powerTo(system, analysis; label = 2)

Compute the active power of a specific branch after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcOptimalPowerFlow(system)
optimize!(system, analysis)
to = powerTo(system, analysis; label = 2)

JuliaGrid.powerCharging — Method

powerCharging(system::PowerSystem, analysis::AC; label)

The function returns the active and reactive power values associated with the charging admittances of a specific branch in the AC framework. The label keyword argument must correspond to an existing branch label.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the power within the AC power flow;

ACOptimalPowerFlow: computes the power within the AC optimal power flow.

Examples

Compute the reactive power after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
charging = powerCharging(system, analysis; label = 2)

Compute the reactive power after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
charging = powerCharging(system, analysis; label = 2)

JuliaGrid.powerSeries — Method

powerSeries(system::PowerSystem, analysis::AC; label)

The function returns the active and reactive power losses across the series impedance of a specific branch within the AC framework. The label keyword argument should correspond to an existing branch label.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the power within the AC power flow;

ACOptimalPowerFlow: computes the power within the AC optimal power flow.

Examples

Compute the reactive power after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
series = powerSeries(system, analysis; label = 2)

Compute the reactive power after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
series = powerSeries(system, analysis; label = 2)

JuliaGrid.powerGenerator — Method

powerGenerator(system::PowerSystem, analysis::AC)

The function returns the active and reactive powers associated with a specific generator in the AC framework. The label keyword argument must match an existing generator label.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the powers within the AC power flow;

ACOptimalPowerFlow: computes the powers within the AC optimal power flow.

Examples

Compute powers after obtaining the AC power flow solution

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
output = powerGenerator(system, analysis; label = 1)

Compute powers after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
output = powerGenerator(system, analysis; label = 1)

JuliaGrid.powerGenerator — Method

powerGenerator(system::PowerSystem, analysis::DC; label)

This function returns the output active power associated with a specific generator in the DC framework. The label keyword argument must match an existing generator label.

Abstract type

The abstract type DC can have the following subtypes:

DCPowerFlow: computes the power within the DC power flow;

DCOptimalPowerFlow: computes the power within the DC optimal power flow.

Examples

Compute the active power of a specific generator after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system)
solve!(system, analysis)
generator = powerGenerator(system, analysis; label = 1)

Compute the active power of a specific generator after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcOptimalPowerFlow(system)
optimize!(system, analysis)
generator = powerGenerator(system, analysis; label = 1)

Current Analysis

JuliaGrid.current! — Method

current!(system::PowerSystem, analysis::AC)

The function computes the currents in the polar coordinate system associated with buses and branches in the AC framework.

Updates

This function calculates various electrical quantities in the polar coordinate system:

injection: current injections at each bus;

from: current flows at each "from" bus end of the branch;

to: current flows at each "to" bus end of the branch;

series: current flows through the series impedance of the branch in the direction from the "from" bus end to the "to" bus end of the branch.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the currents within the AC power flow;

ACOptimalPowerFlow: computes the currents within the AC optimal power flow.

Examples

Compute currents after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
current!(system, analysis)

Compute currents after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
current!(system, analysis)

JuliaGrid.currentInjection — Method

currentInjection(system::PowerSystem, analysis::AC; label)

The function returns the current in the polar coordinate system associated with a specific bus in the AC framework. The label keyword argument must match an existing bus label.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the current within the AC power flow;

ACOptimalPowerFlow: computes the current within the AC optimal power flow.

Examples

Compute the current after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
injection = currentInjection(system, analysis; label = 1)

Compute the current after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
injection = currentInjection(system, analysis; label = 1)

JuliaGrid.currentFrom — Method

currentFrom(system::PowerSystem, analysis::AC; label)

The function returns the current in the polar coordinate system at the "from" bus end associated with a specific branch in the AC framework. The label keyword argument must match an existing branch label.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the current within the AC power flow;

ACOptimalPowerFlow: computes the current within the AC optimal power flow.

Examples

Compute the current after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
from = currentFrom(system, analysis; label = 2)

Compute the current after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
from = currentFrom(system, analysis; label = 2)

JuliaGrid.currentTo — Method

currentTo(system::PowerSystem, analysis::AC; label)

The function returns the current in the polar coordinate system at the "to" bus end associated with a specific branch in the AC framework. The label keyword argument must match an existing branch label.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the current within the AC power flow;

ACOptimalPowerFlow: computes the current within the AC optimal power flow.

Examples

Compute the current after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
to = currentTo(system, analysis; label = 2)

Compute the current after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
to = currentTo(system, analysis; label = 2)

JuliaGrid.currentSeries — Method

currentSeries(system::PowerSystem, analysis::AC; label)

The function returns the current in the polar coordinate system through series impedance associated with a specific branch in the direction from the "from" bus end to the "to" bus end of the branch within the AC framework. The label keyword argument must match an existing branch label.

Abstract type

The abstract type AC can have the following subtypes:

ACPowerFlow: computes the current within the AC power flow;

ACOptimalPowerFlow: computes the current within the AC optimal power flow.

Examples

Compute the current after obtaining the AC power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
line = currentSeries(system, analysis; label = 2)

Compute the current after obtaining the AC optimal power flow solution:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
line = currentSeries(system, analysis; label = 2)