Configuration Setup

For further information on this topic, please see the Power System Model section of the Manual.

Enable or Disable Features

JuliaGrid.@enable — Macro

The macro allows for the activation of various features that are relevant to the analysis performed.

@enable(feature)

These features are associated with different aspects of the analysis, and can be specified using the feature parameter:

generatorVoltage: includes generator magnitude setpoints that are used during voltage initialization

piecewiseObjective: incorporates linear piecewise costs into the optimal power flow objective.

Examples

To activate a single feature, the macro can be called as follows:

@enable(generatorVoltage)

To activate multiple features, the macro can be called as follows:

@enable(generatorVoltage, piecewiseObjective)

JuliaGrid.@disable — Macro

The macro allows for the deactivation of various features that are relevant to the analysis performed.

@disable(feature)

These features are associated with different aspects of the analysis, and can be specified using the feature parameter:

generatorVoltage: excludes generator magnitude setpoints that are used during voltage initialization

piecewiseObjective: excludes linear piecewise costs from the optimal power flow objective.

Examples

To deactivate a single feature, the macro can be called as follows:

@disable(generatorVoltage)

To deactivate multiple features, the macro can be called as follows:

@disable(generatorVoltage, piecewiseObjective)