DC Power Analysis

For further information on this topic, please see the Power System Analysis section of the Manual.

API Index

Power Analysis

power

powerBus

powerBranch

powerGenerator

Power Analysis

JuliaGrid.power — Method

power(system::PowerSystem, model::DCAnalysis)

The function returns the active powers associated with buses, branches, and generators in the DC framework.

Abstract type

Subtypes of the abstract type DCAnalysis include:

DCPowerFlow: computes the powers within the DC power flow,

DCOptimalPowerFlow: computes the powers within the DC optimal power flow.

Returns

The function returns the instance of the DCPower type, which contains the following fields:

The bus field contains powers related to buses:

injection: the active power injection at buses,

supply: the active power injected at buses by the generators.

The branch field contains powers related to branches:

from: the active power flows at the "from" ends,

to: the active power flows at the "to" ends.

The generator field contains powers related to generators:

output: the active power output of the generators.

Examples

Compute powers after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

powers = power(system, model)

Compute powers after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system)
optimize!(system, model)

powers = power(system, model)

JuliaGrid.powerBus — Method

powerBus(system::PowerSystem, model::DCAnalysis; label)

This function calculates the active powers associated with a specific bus in the DC framework. The label keyword argument must match an existing bus label.

Abstract type

Subtypes of the abstract type DCAnalysis include:

DCPowerFlow: computes the powers within the DC power flow,

DCOptimalPowerFlow: computes the powers within the DC optimal power flow.

Returns

The function returns an instance of the DCPowerBus type, which contains the following fields:

injection: the active power injection at buses,

supply: the active power injected at buses by the generators.

Examples

Compute the powers of a specific bus after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

powers = powerBus(system, model; label = 2)

Compute the powers of a specific bus after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system)
optimize!(system, model)

powers = powerBus(system, model; label = 2)

JuliaGrid.powerBranch — Method

powerBranch(system::PowerSystem, model::DCAnalysis; label)

This function calculates the active powers associated with a specific branch in the DC framework. The label keyword argument must match an existing branch label.

Abstract type

Subtypes of the abstract type DCAnalysis include:

DCPowerFlow: computes the powers within the DC power flow,

DCOptimalPowerFlow: computes the powers within the DC optimal power flow.

Returns

The function returns an instance of the DCPowerBranch type, which contains the following fields:

from: the active power flows at the "from" ends,

to: the active power flows at the "to" ends.

Examples

Compute the powers of a specific branch after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

powers = powerBranch(system, model; label = 2)

Compute the powers of a specific branch after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system)
optimize!(system, model)

powers = powerBranch(system, model; label = 2)

JuliaGrid.powerGenerator — Method

powerGenerator(system::PowerSystem, model::DCAnalysis; label)

This function calculates the active powers associated with a specific generator in the DC framework. The label keyword argument must match an existing generator label.

Abstract type

Subtypes of the abstract type DCAnalysis include:

DCPowerFlow: computes the powers within the DC power flow,

DCOptimalPowerFlow: computes the powers within the DC optimal power flow.

Returns

The function returns an instance of the DCPowerGenerator type, which has the following field:

output: the active power output of the generators.

Examples

Compute the powers of a specific generator after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

powers = powerGenerator(system, model; label = 1)

Compute the powers of a specific generator after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system)
optimize!(system, model)

powers = powerGenerator(system, model; label = 1)