DC Optimal Power Flow Solution

For further information on this topic, please see the DC Optimal Power Flow section of the Manual.

API Index

Build Model

dcOptimalPowerFlow

Solve Optimal Power Flow

solve!

Delete Constraints

deleteBalance!

deleteLimit!

deleteRating!

deleteCapability!

Power Analysis

Please refer to the DC Analysis section of the API.

Build Model

JuliaGrid.dcOptimalPowerFlow — Function

dcOptimalPowerFlow(system::PowerSystem, optimizer; bridge, name, balance, limit,
    rating, capability)

The function takes the PowerSystem composite type as input to establish the structure for solving the DC optimal power flow. The optimizer argument is also required to create and solve the optimization problem. If the dcModel field within the PowerSystem composite type has not been created, the function will initiate an update automatically.

Keywords

JuliaGrid offers the ability to manipulate the jump model based on the guidelines provided in the JuMP documentation. However, certain configurations may require different method calls, such as:

bridge: used to manage the bridging mechanism

name: used to manage the creation of string names.

Moreover, we have included keywords that regulate the usage of different types of constraints:

balance: controls the equality constraints that relate to the active power balance equations

limit: controls the inequality constraints that relate to the voltage angle differences between buses

rating: controls the inequality constraints that relate to the long-term rating of branches

capability: controls the inequality constraints that relate to the active power generator outputs.

By default, all of these keywords are set to true and are of the Bool type.

JuMP

The JuliaGrid builds the DC optimal power flow around the JuMP package and supports commonly used solvers. For more information, refer to the JuMP documenatation.

Returns

The function returns an instance of the DCOptimalPowerFlow type, which includes the following fields:

voltage: the angles of bus voltages

output: the output active powers of the generators

jump: the JuMP model

constraint: holds the constraint references to the JuMP model.

Examples

Create the complete DC optimal power flow model:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system, HiGHS.Optimizer)

Create the DC optimal power flow model without rating constraints:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system, HiGHS.Optimizer; rating = false)

Solve Optimal Power Flow

JuliaGrid.solve! — Method

solve!(system::PowerSystem, model::DCOptimalPowerFlow)

The function finds the DC optimal power flow solution and calculate the angles of bus voltages and output active powers of the generators.

The calculated voltage angles and active powers are then stored in the angle variable of the voltage field and the active variable of the power field.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system, HiGHS.Optimizer)
solve!(system, model)

Delete Constraints

JuliaGrid.deleteBalance! — Method

deleteBalance!(system::PowerSystem, model::DCOptimalPowerFlow; label::Int64)

The function removes the active power balance constraint associated with the bus designated by the label in the DC optimal power flow framework.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system, HiGHS.Optimizer)

deleteBalance!(model; label = 1)

JuliaGrid.deleteLimit! — Method

deleteLimit!(system::PowerSystem, model::DCOptimalPowerFlow; label::Int64)

The function removes the voltage angle difference limit constraint associated with the branch designated by the label in the DC optimal power flow framework.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system, HiGHS.Optimizer)

deleteLimit!(model; label = 1)

JuliaGrid.deleteRating! — Method

deleteRating!(system::PowerSystem, model::DCOptimalPowerFlow; label::Int64)

The function removes the active power rating constraint associated with the branch designated by the label in the DC optimal power flow framework.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system, HiGHS.Optimizer)

deleteRating!(model; label = 1)

JuliaGrid.deleteCapability! — Method

deleteCapability!(system::PowerSystem, model::DCOptimalPowerFlow; label::Int64)

The function removes the active power capability constraint associated with the generator designated by the label in the DC optimal power flow framework.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system, HiGHS.Optimizer)

deleteCapability!(model; label = 1)