DC Power Flow

For further information on this topic, please see the Power System Analysis section of the Manual.

API Index

DC Power Flow

dcPowerFlow

solve!

Power Analysis

analysisBus

analysisBranch

analysisGenerator

DC Power Flow

JuliaGrid.dcPowerFlow — Function

dcPowerFlow(system::PowerSystem)

The function accepts the PowerSystem composite type as input, which is utilized to establish the structure for solving the DC power flow.

If the DC model was not created, the function will automatically initiate an update of the dcModel field within the PowerSystem composite type. Additionally, if the slack bus lacks an in-service generator, JuliaGrid considers it a mistake and defines a new slack bus as the first generator bus with an in-service generator in the bus type list.

Returns

The function returns an instance of the DCPowerFlow type, which includes the following filled fields:

voltage: the variable allocated to store the angles of bus voltages,

factorization: the factorized nodal matrix.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)

JuliaGrid.solve! — Method

solve!(system::PowerSystem, model::DCPowerFlow)

By computing the voltage angles for each bus, the function solves the DC power flow problem. The resulting voltage angles are stored in the voltage field of the DCPowerFlow type.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

Power Analysis

Missing docstring.

Missing docstring for analysisBus(::PowerSystem, ::DCPowerFlow). Check Documenter's build log for details.

Missing docstring.

Missing docstring for analysisBranch(::PowerSystem, ::DCPowerFlow). Check Documenter's build log for details.

Missing docstring.

Missing docstring for analysisGenerator(::PowerSystem, ::DCPowerFlow). Check Documenter's build log for details.