AC and DC Model

More information can be found in the Power System Model section of the manual.

AC Model

JuliaGrid.acModel! — Function

The function generates vectors and matrices based on the power system topology and parameters associated with AC analysis. We advise the reader to read the section AC Model, which explains all the data involved in the field acModel.

acModel!(system::PowerSystem)

The function updates the field acModel. Once formed, the field will be automatically updated when using functions addBranch!, shuntBus!, statusBranch!, parameterBranch!.

Example

system = powerSystem("case14.h5")
acModel!(system)

DC Model

JuliaGrid.dcModel! — Function

The function generates vectors and matrices based on the power system topology and parameters associated with DC analysis. We advise the reader to read the section DC Model, which explains all the data involved in the field dcModel.

dcModel!(system::PowerSystem)

The function updates the field dcModel. Once formed, the field will be automatically updated when using functions addBranch!, statusBranch!, parameterBranch!.

Example

system = powerSystem("case14.h5")
dcModel!(system)