Optimal Power Flow Solution

For further information on this topic, please see the Power System Analysis section of the Manual.

API Index

Build Model

dcOptimalPowerFlow

Solve Optimal Power Flow

optimize!

Build Model

JuliaGrid.dcOptimalPowerFlow — Function

dcOptimalPowerFlow(system::PowerSystem, optimizer; bridges, names, balance, limit,
    rating, capability)

The function takes the PowerSystem composite type as input to establish the structure for solving the DC optimal power flow. The optimizer argument is also required to create and solve the optimization problem. If the dcModel field within the PowerSystem composite type has not been created, the function will initiate an update automatically.

Keywords

JuliaGrid offers the ability to manipulate the jump model based on the guidelines provided in the JuMP documentation. However, certain configurations may require different method calls, such as:

bridges: used to manage the bridging mechanism

names: used to manage the creation of string names.

Moreover, we have included keywords that regulate the usage of different types of constraints:

balance: controls the equality constraints that relate to the active power balance equations

limit: controls the inequality constraints that relate to the voltage angle differences between buses

rating: controls the inequality constraints that relate to the long-term rating of branches

capability: controls the inequality constraints that relate to the active power generator outputs.

By default, all of these keywords are set to true and are of the Bool type.

JuMP

The JuliaGrid builds the DC optimal power flow around the JuMP package and supports commonly used solvers. For more information, refer to the JuMP documenatation.

Returns

The function returns an instance of the DCOptimalPowerFlow type, which includes the following fields:

voltage: the angles of bus voltages

output: the active power output of the generators

jump: the JuMP model

constraint: holds the constraint references to the JuMP model.

Examples

Create the complete DC optimal power flow model:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system, HiGHS.Optimizer)

Create the DC optimal power flow model without rating constraints:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system, HiGHS.Optimizer; rating = false)

Solve Optimal Power Flow

Missing docstring.

Missing docstring for optimize!. Check Documenter's build log for details.