Post-processing Analysis

For further information on this topic, please see the Power System Model section of the Manual.

API Index

bus!

branch!

generator!

Bus Function

JuliaGrid.bus! — Function

This function calculates the powers and currents associated with buses.

bus!(system::PowerSystem, result::Result)

After the function is executed, the bus field within the Result type gets updated.

AC Power Flow Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, result)
    if all(stopping .< 1e-8)
        break
    end
    solvePowerFlow!(system, result)
end

bus!(system, result)

DC Power Flow Example

system = powerSystem("case14.h5")
dcModel!(system)

result = solvePowerFlow(system)
bus!(system, result)

Branch Function

JuliaGrid.branch! — Function

The function is used to calculate the powers and currents associated with branches for AC power flow analysis and, in relation to DC power flow analysis, it only calculates active powers.

branch!(system::PowerSystem, result::Result)

The function is responsible for updating the branch field within the Result type after it has been executed.

AC Power Flow Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, result)
    if all(stopping .< 1e-8)
        break
    end
    solvePowerFlow!(system, result)
end

branch!(system, result)

DC Power Flow Example

system = powerSystem("case14.h5")
dcModel!(system)

result = solvePowerFlow(system)
branch!(system, result)

Generator Function

JuliaGrid.generator! — Function

The function computes powers related to generators.

generator!(system::PowerSystem, result::Result)

Once executed, the generator field within the Result type is updated accordingly.

AC Power Flow Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, result)
    if all(stopping .< 1e-8)
        break
    end
    solvePowerFlow!(system, result)
end

generator!(system, result)

DC Power Flow Example

system = powerSystem("case14.h5")
dcModel!(system)

result = solvePowerFlow(system)
generator!(system, result)