Power Flow

For further information on this topic, please see the AC Power Flow or DC Power Flow sections of the Manual. Below, we have provided a list of functions that can be utilized for power flow analysis.

Build Model

newtonRaphson

fastNewtonRaphsonBX

fastNewtonRaphsonXB

gaussSeidel

dcPowerFlow

Solve Power Flow

mismatch!

solve!

Power Analysis

power!

powerInjection

powerSupply

powerShunt

powerFrom

powerTo

powerCharging

powerLoss

powerGenerator

Current Analysis

current!

currentInjection

currentFrom

currentTo

currentLine

Additional Functions

reactiveLimit!

adjustAngle!

Build Model

JuliaGrid.newtonRaphson — Function

newtonRaphson(system::PowerSystem)

The function accepts the PowerSystem composite type as input and uses it to set up the Newton-Raphson method to solve AC power flow. Additionally, if the AC model was not created, the function will automatically initiate an update of the acModel field within the PowerSystem composite type.

Returns

The function returns an instance of the NewtonRaphson subtype of the abstract ACPowerFlow type, which includes the following fields:

voltage: the magnitudes and angles of bus voltages

power: the variable allocated to store the active and reactive powers,

current: the variable allocated to store the currents,

jacobian: the Jacobian matrix

mismatch: the active and reactive power injection mismatches

increment: the magnitudes and angles of bus voltage increments

pq: indices of demand buses

pvpq: indices of demand and generator buses.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)

JuliaGrid.fastNewtonRaphsonBX — Function

fastNewtonRaphsonBX(system::PowerSystem)

The function accepts the PowerSystem composite type as input and uses it to set up the Fast Newton-Raphson method of version BX to solve AC power flow. Additionally, if the AC model was not created, the function will automatically initiate an update of the acModel field within the PowerSystem composite type.

Returns

The function returns an instance of the FastNewtonRaphson subtype of the abstract ACPowerFlow type, which includes the following fields:

voltage: the magnitudes and angles of bus voltages

power: the variable allocated to store the active and reactive powers,

current: the variable allocated to store the currents,

active:

jacobian: the Jacobian matrix associated with active power equations

mismatch: the active power injection mismatches

increment: the angles of bus voltage increments

factorization: the factorized Jacobian matrix

reactive:

jacobian: the Jacobian matrix associated with reactive power equations

mismatch: the reative power injection mismatches

increment: the magnitudes of bus voltage increments

factorization: the factorized Jacobian matrix

pq: indices of demand buses

pvpq: indices of demand and generator buses.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = fastNewtonRaphsonBX(system)

JuliaGrid.fastNewtonRaphsonXB — Function

fastNewtonRaphsonXB(system::PowerSystem)

The function accepts the PowerSystem composite type as input and uses it to set up the Fast Newton-Raphson method of version XB to solve AC power flow. Additionally, if the AC model was not created, the function will automatically initiate an update of the acModel field within the PowerSystem composite type.

Returns

The function returns an instance of the FastNewtonRaphson subtype of the abstract ACPowerFlow type, which includes the following fields:

voltage: the magnitudes and angles of bus voltages

active:

jacobian: the Jacobian matrix associated with active power equations

mismatch: the active power injection mismatches

increment: the angles of bus voltage increments

factorization: the factorized Jacobian matrix

reactive:

jacobian: the Jacobian matrix associated with reactive power equations

mismatch: the reative power injection mismatches

increment: the magnitudes of bus voltage increments

factorization: the factorized Jacobian matrix

pq: indices of demand buses

pvpq: indices of demand and generator buses.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = fastNewtonRaphsonXB(system)

JuliaGrid.gaussSeidel — Function

gaussSeidel(system::PowerSystem)

The function accepts the PowerSystem composite type as input and uses it to set up the Gauss-Seidel method to solve AC power flow. Additionally, if the AC model was not created, the function will automatically initiate an update of the acModel field within the PowerSystem composite type.

Returns

The function returns an instance of the GaussSeidel subtype of the abstract ACPowerFlow type, which includes the following fields:

voltage: the magnitudes and angles of bus voltages

complex: the complex voltages

magnitude: the bus voltage magnitudes for corrections

pq: indices of demand buses

pv: indices of generator buses.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = gaussSeidel(system)

JuliaGrid.dcPowerFlow — Function

dcPowerFlow(system::PowerSystem)

The function accepts the PowerSystem composite type as input, which is utilized to establish the structure for solving the DC power flow.

If the DC model was not created, the function will automatically initiate an update of the dcModel field within the PowerSystem composite type. Additionally, if the slack bus lacks an in-service generator, JuliaGrid considers it a mistake and defines a new slack bus as the first generator bus with an in-service generator in the bus type list.

Returns

The function returns an instance of the DCPowerFlow type, which includes the following filled fields:

voltage: the variable allocated to store the angles of bus voltages,

power: the variable allocated to store the active powers,

factorization: the factorized nodal matrix.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)

Solve Power Flow

JuliaGrid.mismatch! — Method

mismatch!(system::PowerSystem, model::ACPowerFlow)

The function calculates both active and reactive power injection mismatches and returns their maximum absolute values, which can be utilized to terminate the iteration loop of methods employed to solve the AC power flow problem.

This function updates the mismatch variables in the Newton-Raphson and fast Newton-Raphson methods. It should be employed during the iteration loop before invoking the solve! function.

In contrast, the Gauss-Seidel method does not need mismatches to obtain bus voltages, but the maximum absolute values are commonly employed to stop the iteration loop. The function does not save any data and should be utilized during the iteration loop before invoking the solve! function.

Abstract type

The abstract type ACPowerFlow can have the following subtypes:

NewtonRaphson: computes the power mismatches within the Newton-Raphson method,

FastNewtonRaphson: computes the power mismatches within the fast Newton-Raphson method,

GaussSeidel: computes the power mismatches within the Gauss-Seidel method.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
mismatch!(system, model)

JuliaGrid.solve! — Method

solve!(system::PowerSystem, model::ACPowerFlow)

The function employs the Newton-Raphson, fast Newton-Raphson, or Gauss-Seidel method to solve the AC power flow problem and calculate the magnitudes and angles of bus voltages.

After the mismatch! function is called, solve! should be executed to perform a single iteration of the method. The calculated voltages are stored in the voltage field of the respective struct type.

Abstract type

The abstract type ACPowerFlow can have the following subtypes:

NewtonRaphson: computes the bus voltages within the Newton-Raphson method

FastNewtonRaphson: computes the bus voltages within the fast Newton-Raphson method

GaussSeidel: computes the bus voltages within the Gauss-Seidel method.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

JuliaGrid.solve! — Method

solve!(system::PowerSystem, model::DCPowerFlow)

By computing the voltage angles for each bus, the function solves the DC power flow problem. The resulting voltage angles are stored in the voltage field of the DCPowerFlow type.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

Additional Functions

JuliaGrid.reactiveLimit! — Function

reactiveLimit!(system::PowerSystem, model::ACPowerFlow)

The function verifies whether the generators in a power system exceed their reactive power limits. This is done by setting the reactive power of the generators to within the limits if they are violated, after determining the bus voltage magnitudes and angles. If the limits are violated, the corresponding generator buses or the slack bus are converted to demand buses.

The function assigns values to the generator.output.active and bus.supply.active variables of the PowerSystem type. Additionally, it examines the reactive powers of the generator and adjusts them to their maximum or minimum values if they exceed the specified threshold. Subsequently, the generator.output.reactive variable of the PowerSystem type is modified accordingly. As a result of this adjustment, the bus.supply.reactive variable of the PowerSystem type is also updated, and the bus types specified in bus.layout.type are modified. If the slack bus is converted, the bus.layout.slack field is correspondingly adjusted.

Abstract type

The abstract type ACPowerFlow can have the following subtypes:

NewtonRaphson: computes the bus voltages within the Newton-Raphson method

FastNewtonRaphson: computes the bus voltages within the fast Newton-Raphson method

GaussSeidel: computes the bus voltages within the Gauss-Seidel method.

Returns

The function returns the violate variable to indicate which buses violate the limits, with -1 indicating a violation of the minimum limits and 1 indicating a violation of the maximum limits.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

violate = reactiveLimit!(system, model)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

JuliaGrid.adjustAngle! — Function

adjustAngle!(system::PowerSystem, model::ACPowerFlow; slack)

The function modifies the bus voltage angles based on a different slack bus than the one identified by the bus.layout.slack field. This function only updates the voltage.angle variable of the Model composite type.

For instance, if the reactive power of the generator exceeds the limit on the slack bus, the reactiveLimit! function will change that bus to the demand bus and designate the first generator bus in the sequence as the new slack bus. After obtaining the updated AC power flow solution based on the new slack bus, it is possible to adjust the voltage angles to align with the angle of the original slack bus. The slack keyword specifies the bus label of the original slack bus.

Abstract type

The abstract type ACPowerFlow can have the following subtypes:

NewtonRaphson: computes the bus voltages within the Newton-Raphson method

FastNewtonRaphson: computes the bus voltages within the fast Newton-Raphson method

GaussSeidel: computes the bus voltages within the Gauss-Seidel method.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

reactiveLimit!(system, model)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

adjustAngle!(system, model; slack = 1)