Power Flow Solution

For further information on this topic, please see the Power System Analysis section of the Manual.

API Index

newtonRaphson

fastNewtonRaphsonBX

fastNewtonRaphsonXB

gaussSeidel

dcPowerFlow

mismatchNewtonRaphson!

mismatchFastNewtonRaphson!

mismatchGaussSeidel!

solveNewtonRaphson!

solveFastNewtonRaphson!

solveGaussSeidel!

solveDCPowerFlow!

reactivePowerLimit!

adjustVoltageAngle!

Newton-Raphson Method

JuliaGrid.newtonRaphson — Function

The function accepts the PowerSystem composite type as input and uses it to set up the Newton-Raphson method to solve AC power flow. Its output is the Result composite type.

newtonRaphson(system::PowerSystem)

The model field of the Result type is updated during the function's execution. Furthermore, if the AC model was not created, the function will automatically initiate an update of the acModel field within the PowerSystem composite type.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)

JuliaGrid.mismatchNewtonRaphson! — Function

The function calculates both active and reactive power injection mismatches and returns their maximum absolute values, which can be utilized to terminate the iteration loop of the Newton-Raphson method that is employed to solve the AC power flow problem.

mismatchNewtonRaphson!(system::PowerSystem, result::Result)

This function updates the mismatch field in the Result composite type and should be employed during the iteration loop before invoking the solveNewtonRaphson! function.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
mismatchNewtonRaphson!(system, result)

JuliaGrid.solveNewtonRaphson! — Function

The function employs the Newton-Raphson method to determine the magnitudes and angles of bus voltages, which is used to solve the AC power flow problem.

solveNewtonRaphson!(system, result)

This function updates the bus.voltage and model fields of the Result composite type during the iteration loop, which should follow the mismatchNewtonRaphson! function. Together, these functions collaborate to execute one iteration of the Newton-Raphson algorithm.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
for i = 1:10
    stopping = mismatchNewtonRaphson!(system, result)
    if all(stopping .< 1e-8)
        break
    end
    solveNewtonRaphson!(system, result)
end

Fast Newton-Raphson Method

JuliaGrid.fastNewtonRaphsonBX — Function

The function accepts the PowerSystem composite type as input and uses it to set up the Fast Newton-Raphson method of version BX to solve AC power flow. Its output is the Result composite type.

fastNewtonRaphsonBX(system::PowerSystem)

The model field of the Result type is updated during the function's execution. Furthermore, if the AC model was not created, the function will automatically initiate an update of the acModel field within the PowerSystem composite type.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = fastNewtonRaphsonBX(system)

JuliaGrid.fastNewtonRaphsonXB — Function

The function accepts the PowerSystem composite type as input and uses it to set up the Fast Newton-Raphson method of version XB to solve AC power flow. Its output is the Result composite type.

fastNewtonRaphsonXB(system::PowerSystem)

The model field of the Result type is updated during the function's execution. Furthermore, if the AC model was not created, the function will automatically initiate an update of the acModel field within the PowerSystem composite type.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = fastNewtonRaphsonXB(system)

JuliaGrid.mismatchFastNewtonRaphson! — Function

The function calculates both active and reactive power injection mismatches and returns their maximum absolute values, which can be utilized to terminate the iteration loop of the fast Newton-Raphson method that is employed to solve the AC power flow problem.

mismatchFastNewtonRaphson!(system::PowerSystem, result::Result)

This function updates the mismatch field in the Result composite type and should be employed during the iteration loop before invoking the solveFastNewtonRaphson! function.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = fastNewtonRaphson(system)
mismatchFastNewtonRaphson!(system, result)

JuliaGrid.solveFastNewtonRaphson! — Function

The function employs the fast Newton-Raphson method to determine the magnitudes and angles of bus voltages, which is used to solve the AC power flow problem.

solveFastNewtonRaphson!(system, result)

This function updates the bus.voltage and model fields of the Result composite type during the iteration loop, which should follow the mismatchFastNewtonRaphson! function. Together, these functions collaborate to execute one iteration of the fast Newton-Raphson algorithm.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = fastNewtonRaphsonBX(system)
for i = 1:10
    stopping = mismatchFastNewtonRaphson!(system, result)
    if all(stopping .< 1e-8)
        break
    end
    solveFastNewtonRaphson!(system, result)
end

Gauss-Seidel Method

JuliaGrid.gaussSeidel — Function

The function accepts the PowerSystem composite type as input and uses it to set up the Gauss-Seidel method to solve AC power flow. Its output is the Result composite type.

gaussSeidel(system::PowerSystem)

The model field of the Result type is updated during the function's execution. Furthermore, if the AC model was not created, the function will automatically initiate an update of the acModel field within the PowerSystem composite type.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = gaussSeidel(system)

JuliaGrid.mismatchGaussSeidel! — Function

The function calculates both active and reactive power injection mismatches and returns their maximum absolute values, which can be utilized to terminate the iteration loop of the Gauss-Seidel method that is employed to solve the AC power flow problem.

mismatchGaussSeidel!(system::PowerSystem, result::Result)

The Gauss-Seidel method does not need mismatches to obtain bus voltages, but the maximum absolute values are commonly employed to stop the iteration loop. As a result, the function does not save any data and should be utilized during the iteration loop before invoking the solveGaussSeidel! function.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = gaussSeidel(system)
mismatchGaussSeidel!(system, result)

JuliaGrid.solveGaussSeidel! — Function

The function employs the Gauss-Seidel method to determine the magnitudes and angles of bus voltages, which is used to solve the AC power flow problem.

solveGaussSeidel!(system, result)

This function updates the bus.voltage and model fields of the Result composite type during the iteration loop. The mismatchGaussSeidel! function can also be used in conjunction with this function to terminate the iteration loop.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = gaussSeidel(system)
for i = 1:10
    stopping = mismatchGaussSeidel!(system, result)
    if all(stopping .< 1e-8)
        break
    end
    solveGaussSeidel!(system, result)
end

DC Power Flow

JuliaGrid.dcPowerFlow — Function

The function accepts the PowerSystem composite type as input, which is utilized to establish the structure for solving the DC power flow. Its output is the Result composite type.

dcPowerFlow(system::PowerSystem)

The model field of the Result type is updated during the function's execution. Furthermore, if the DC model was not created, the function will automatically initiate an update of the dcModel field within the PowerSystem composite type.

Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)

JuliaGrid.solveDCPowerFlow! — Function

By computing the voltage angles for each bus, the function solves the DC power flow problem.

solveDCPowerFlow!(system, result)

The voltage.angle field of the composite type Result is modified by the function.

Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)
solveDCPowerFlow!(system, result)

Additional Functions

JuliaGrid.reactivePowerLimit! — Function

The function verifies whether the generators in a power system exceed their reactive power limits. This is done by setting the reactive power of the generators to within the limits if they are violated, after determining the bus voltage magnitudes and angles. If the limits are violated, the corresponding generator buses or the slack bus are converted to demand buses.

The function returns the violate variable to indicate which buses violate the limits, with -1 indicating a violation of the minimum limits and 1 indicating a violation of the maximum limits.

reactivePowerLimit!(system::PowerSystem, result::Result)

Initially, if the generator! function has not been run, it will be executed to update the generator field of the Result type.

Afterward, the function uses the results from the generator field to assign values to the generator.output.active and bus.supply.active fields of the System type.

At the end of the process, the function inspects the reactive powers of the generator and adjusts them to their maximum or minimum values if they violate the threshold. The generator.output.reactive field of the System type is then modified accordingly. In light of this modification, the bus.supply.reactive field of the System type is also updated, and the bus types in bus.layout.type are adjusted. If the slack bus is converted, the bus.layout.slack field is modified accordingly.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, result)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, result)
end

violate = reactivePowerLimit!(system, result)

result = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, result)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, result)
end

JuliaGrid.adjustVoltageAngle! — Function

The function modifies the bus voltage angles based on a different slack bus than the one identified by the bus.layout.slack field. This function only updates the bus.voltage.angle field of the Result type.

adjustVoltageAngle!(system::PowerSystem, result::Result; slack)

For instance, if the reactive power of the generator exceeds the limit on the slack bus, the reactivePowerLimit! function will change that bus to a demand bus and designate the first generator bus in the sequence as the new slack bus. After obtaining the updated AC power flow solution based on the new slack bus, it is possible to adjust the voltage angles to align with the angle of the original slack bus. The slack keyword specifies the bus label of the original slack bus.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, result)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, result)
end

reactivePowerLimit!(system, result)

result = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, result)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, result)
end

adjustVoltageAngle!(system, result; slack = 1)