Public Types

This section introduces the public types available in JuliaGrid. These types are designed to model and represent key components of power systems, measurements, and analyses, enabling users to efficiently simulate, optimize, and analyze various scenarios.

To load public types into the current scope, utilize the following command:

using JuliaGrid

Power System Model

PowerSystem

Bus

Branch

Generator

BaseData

Model

Measurement Model

Measurement

Voltmeter

Ammeter

Wattmeter

Varmeter

PMU

Analysis

Analysis

AC

DC

LDLt

LU

QR

Power Flow

ACPowerFlow

NewtonRaphson

FastNewtonRaphson

GaussSeidel

DCPowerFlow

Optimal Power Flow

ACOptimalPowerFlow

DCOptimalPowerFlow

Observability Analysis

Island

PMUPlacement

State Estimation

ACStateEstimation

PMUStateEstimation

DCStateEstimation

LWLS

NWLS

LAV

Normal

Orthogonal

Power System Model

JuliaGrid.PowerSystem — Type

PowerSystem

A composite type constructed using the powerSystem function to store power system data.

Fields

bus::Bus: Bus-related data.

branch::Branch: Branch-related data.

generator::Generator: Generator-related data.

base::BaseData: Base power and base voltages.

model::Model: Data related to AC and DC analyses.

JuliaGrid.Bus — Type

Bus

A composite type used in PowerSystem to store bus-related data.

Fields

label::Union{OrderedDict{String, Int64}, OrderedDict{Int64, Int64}}: Bus labels.

demand::BusDemand: Active and reactive power demands.

supply::BusSupply: Active and reactive power supplies from generators.

shunt::BusShunt: Active and reactive power injections or demands from shunt elements.

voltage::BusVoltage: Initial voltages and voltage magnitude constraints.

layout::BusLayout: Bus layout, including bus types.

number::Int64: Total number of buses.

JuliaGrid.Branch — Type

Branch

A composite type used in PowerSystem to store branch-related data.

Fields

label::Union{OrderedDict{String, Int64}, OrderedDict{Int64, Int64}}: Branch labels.

parameter::BranchParameter: Data related to the π-model of the branches.

flow::BranchFlow: Branch flow constraints.

voltage::BranchVoltage: Voltage angle difference constraints.

layout::BranchLayout: Branch layout, including operational statuses.

number::Int64: Total number of branches.

JuliaGrid.Generator — Type

Generator

A composite type used in PowerSystem to store generator-related data.

Fields

label::Union{OrderedDict{String, Int64}, OrderedDict{Int64, Int64}}: Generator labels.

output::GeneratorOutput: Active and reactive power outputs.

capability::GeneratorCapability: Power output constraints.

ramping::GeneratorRamping: Ramp rate limits.

voltage::GeneratorVoltage: Voltage magnitude setpoints.

cost::GeneratorCost: Costs associated with active and reactive power outputs.

layout::GeneratorLayout: Generator layout, including operational statuses.

number::Int64: Total number of generators.

JuliaGrid.BaseData — Type

BaseData

A composite type used in PowerSystem to store base data.

Fields

power::BasePower: Base power.

voltage::BaseVoltage: Base voltages.

JuliaGrid.Model — Type

Model

A composite type used in PowerSystem to store vectors and matrices related to the power system's topology and parameters.

Fields

ac::ACModel: AC model, including the nodal admittance matrix and Y-parameters of two-port branches.

dc::DCModel: DC model, including the nodal admittance matrix and branch admittances.

Measurement Model

JuliaGrid.Measurement — Type

Measurement

A composite type built using the measurement function to store measurement data.

Fields

voltmeter::Voltmeter: Data related to bus voltage magnitude measurements.

ammeter::Ammeter: Data related to branch current magnitude measurements.

wattmeter::Wattmeter: Data related to active power injection and active power flow measurements.

varmeter::Varmeter: Data related to reactive power injection and reactive power flow measurements.

pmu::PMU: Data related to bus voltage and branch current phasor measurements.

JuliaGrid.Voltmeter — Type

Voltmeter

A composite type used in Measurement to store voltmeter-related data.

Fields

label::Union{OrderedDict{String, Int64}, OrderedDict{Int64, Int64}}: Voltmeter labels.

magnitude::GaussMeter: Bus voltage magnitude measurements.

layout::VoltmeterLayout: Placement indices and indicators.

number::Int64: Total number of voltmeters.

JuliaGrid.Ammeter — Type

Ammeter

A composite type used in Measurement to store ammeter-related data.

Fields

label::Union{OrderedDict{String, Int64}, OrderedDict{Int64, Int64}}: Ammeter labels.

magnitude::GaussMeter: Branch current magnitude measurements.

layout::AmmeterLayout: Placement indices and indicators.

number::Int64: Total number of ammeters.

JuliaGrid.Wattmeter — Type

Wattmeter

A composite type used in Measurement to store wattmeter-related data.

Fields

label::Union{OrderedDict{String, Int64}, OrderedDict{Int64, Int64}}: Wattmeter labels.

active::GaussMeter: Active power injection and active power flow measurements.

layout::PowermeterLayout: Placement indices and indicators.

number::Int64: Total number of wattmeters.

JuliaGrid.Varmeter — Type

Varmeter

A composite type used in Measurement to store varmeter-related data.

Fields

label::Union{OrderedDict{String, Int64}, OrderedDict{Int64, Int64}}: Varmeter labels.

reactive::GaussMeter: Reactive power injection and reactive power flow measurements.

layout::PowermeterLayout: Placement indices and indicators.

number::Int64: Total number of varmeters.

JuliaGrid.PMU — Type

PMU

A composite type used in Measurement to store PMU-related data.

Fields

label::Union{OrderedDict{String, Int64}, OrderedDict{Int64, Int64}}: PMU labels.

magnitude::GaussMeter: Bus voltage and branch current magnitude measurements.

angle::GaussMeter: Bus voltage and branch current angle measurements.

layout::PmuLayout: Placement indices and indicators.

number::Int64: Total number of PMUs.

Analysis

JuliaGrid.Analysis — Type

Analysis

An abstract type used for representing both AC and DC analyses in JuliaGrid.

JuliaGrid.AC — Type

AC <: Analysis

An abstract type representing AC analyses in JuliaGrid.

JuliaGrid.DC — Type

DC <: Analysis

An abstract type representing DC analyses in JuliaGrid.

JuliaGrid.LDLt — Type

LDLt

An abstract type used for representing LDLt factorization in JuliaGrid.

JuliaGrid.LU — Type

LU

An abstract type used for representing LU factorization in JuliaGrid.

JuliaGrid.QR — Type

QR

An abstract type used for representing QR factorization in JuliaGrid.

Power Flow

JuliaGrid.ACPowerFlow — Type

ACPowerFlow{T} <: AC where T <: Union{NewtonRaphson, FastNewtonRaphson, GaussSeidel}

A composite type representing an AC power flow framework that uses a specified method for solving the AC power flow equations. The type parameter T defines the method used, which can be either NewtonRaphson, FastNewtonRaphson, or GaussSeidel.

Fields

voltage::Polar: Bus voltages represented in polar form.

power::ACPower: Active and reactive powers at the buses, branches, and generators.

current::ACCurrent: Currents at the buses and branches.

method::T: Vectors and matrices associated with the method used to solve the AC power flow.

JuliaGrid.NewtonRaphson — Type

NewtonRaphson

A composite type built using the newtonRaphson function to define the AC power flow framework, which will be solved using the Newton-Raphson method.

Fields

jacobian::SparseMatrixCSC{Float64, Int64}: Jacobian matrix.

mismatch::Vector{Float64}: Vector of mismatches.

increment::Vector{Float64}: Vector of state variable increments.

factorization::Factorization{Float64}: Factorization of the Jacobian matrix.

pq::Vector{Int64}: Indices related to demand buses.

pvpq::Vector{Int64}: Indices related to demand and generator buses.

pattern::Int64: Tracks pattern changes in entries of the Jacobian matrix.

JuliaGrid.FastNewtonRaphson — Type

FastNewtonRaphson

A composite type built using the fastNewtonRaphsonBX or fastNewtonRaphsonXB functions to define the AC power flow framework, which will be solved using the fast Newton-Raphson method.

Fields

active:FastNewtonRaphsonModel: Jacobian, mismatches, and incrementes for active power equations.

reactive:FastNewtonRaphsonModel: Jacobian, mismatches, and incrementes for active power equations.

pq::Vector{Int64}: Indices related to demand buses.

pvpq::Vector{Int64}: Indices related to demand and generator buses.

acmodel::Int64: Tracks values changing in entries of the Jacobian matrices.

pattern::Int64: Tracks pattern changes in entries of the Jacobian matrices.

bx::Bool: Version of the method, either BX or XB.

JuliaGrid.GaussSeidel — Type

GaussSeidel

A composite type built using the gaussSeidel function to define the AC power flow framework, which will be solved using the Gauss-Seidel method.

Fields

voltage::Vector{ComplexF64}: Vector of complex voltage values.

pq::Vector{Int64}: Indices related to demand buses.

pv::Vector{Int64}: Indices related to generator buses.

JuliaGrid.DCPowerFlow — Type

DCPowerFlow <: DC

A composite type built using the dcPowerFlow function to define the DC power flow framework.

Fields

voltage::PolarAngle: Bus voltage angles.

power::DCPower: Active powers at the buses, branches, and generators.

method::DCPowerFlowMethod: Factorization of the nodal admittance matrix.

Optimal Power Flow

JuliaGrid.ACOptimalPowerFlow — Type

ACOptimalPowerFlow <: AC

A composite type built using the acOptimalPowerFlow function to define the AC optimal power flow framework.

Fields

voltage::Polar: Bus voltages represented in polar form.

power::ACPower: Active and reactive powers at buses, branches, and generators.

current::ACCurrent: Currents at buses and branches.

method::ACOptimalPowerFlowMethod: The JuMP model, including variables, constraints, and objective.

JuliaGrid.DCOptimalPowerFlow — Type

DCOptimalPowerFlow <: DC

A composite type built using the dcOptimalPowerFlow function to define the DC optimal power flow framework.

Fields

voltage::PolarAngle: Bus voltage angles.

power::DCPower: Active powers at buses, branches, and generators.

method::DCOptimalPowerFlowMethod: The JuMP model, including variables, constraints, and objective.

Observability Analysis

JuliaGrid.Island — Type

Island

A composite type built using the islandTopologicalFlow and islandTopological functions, which holds data about observable islands.

Fields

island::Vector{Vector{Int64}}: List of observable islands, where each island is represented by a vector of bus indices.

bus::Vector{Int64}: Positions of buses in relation to each island.

tie::TieData: Tie data associated with buses and branches.

JuliaGrid.PMUPlacement — Type

PMUPlacement

A composite type built using the pmuPlacement function, which stores data on optimal PMU placement.

Fields

bus::Union{OrderedDict{String, Int64}, OrderedDict{Int64, Int64}}: PMU placement at buses.

from::Union{OrderedDict{String, Int64}, OrderedDict{Int64, Int64}}: PMU placement at from-buses.

to::Union{OrderedDict{String, Int64}, OrderedDict{Int64, Int64}}: PMU placement at to-buses.

State Estimation

JuliaGrid.ACStateEstimation — Type

ACStateEstimation{T} <: AC where T <: Union{NWLS{Normal}, NWLS{Orthogonal}, LAV}

A composite type built using the gaussNewton and acLavStateEstimation functions to define the AC state estimation framework. The type parameter T defines the estimation method, which can be either NWLS or LAV.

Fields

voltage::Polar: Bus voltages represented in polar form.

power::ACPower: Active and reactive powers at the buses and branches.

current::ACCurrent: Currents at the buses and branches.

method::T: The estimation model associated with the method used to solve the AC state estimation.

JuliaGrid.PMUStateEstimation — Type

PMUStateEstimation{T} <: AC where T <: Union{LWLS{Normal}, LWLS{Orthogonal}, LAV}

A composite type built using the pmuStateEstimation and pmuLavStateEstimation functions to define PMU state estimation framework. The type parameter T defines the estimation method, which can be either LWLS or LAV.

Fields

voltage::Polar: Bus voltages represented in polar form.

power::ACPower: Active and reactive powers at the buses and branches.

current::ACCurrent: Currents at the buses and branches.

method::T: The estimation model associated with the method used to solve the PMU state estimation.

JuliaGrid.DCStateEstimation — Type

DCStateEstimation{T} <: DC where T <: Union{LWLS{Normal}, LWLS{Orthogonal}, LAV}

A composite type built using the dcStateEstimation and dcLavStateEstimation functions to define the the DC state estimation framework. The type parameter T defines the estimation method, which can be either LWLS or LAV.

Fields

voltage::PolarAngle: Bus voltage angles.

power::DCPower: Active powers at the buses and generators.

method::T: The estimation model associated with the method used to solve the DC state estimation.

JuliaGrid.LWLS — Type

LWLS{T <: Union{Normal, Orthogonal}}

A composite type representing a linear weighted-least squares state estimation model.

Fields

coefficient::SparseMatrixCSC{Float64, Int64}: Coefficient matrix.

precision::SparseMatrixCSC{Float64, Int64}: Precision matrix.

mean::Vector{Float64}: Mean vector.

factorization::Factorization{Float64}: Factorization of the coefficient matrix.

number::Int64: Number of measurement devices.

pattern::Int64: Tracks pattern changes in the coefficient matrix.

run::Bool: Indicates whether factorization can be reused.

JuliaGrid.NWLS — Type

NWLS{T <: Union{Normal, Orthogonal}}

A composite type representing a nonlinear weighted-least squares state estimation model.

Fields

jacobian::SparseMatrixCSC{Float64, Int64}: Jacobian matrix.

precision::SparseMatrixCSC{Float64, Int64}: Precision matrix.

mean::Vector{Float64}: Mean vector.

residual::Vector{Float64}: Residual vector.

increment::Vector{Float64}: Increment vector.

factorization::Factorization{Float64}: Factorization of the Jacobian matrix.

objective::Float64: Value of the objective function.

type::Vector{Int8}: Indicators of measurement types.

index::Vector{Int64}: Indices of buses and branches where measurements are located.

range::Vector{Int64}: Range of measurement devices.

pattern::Int64: Tracks pattern changes in the Jacobian matrix.

JuliaGrid.LAV — Type

LAV

A composite type representing a least absolute value state estimation model.

Fields

jump::JuMP.Model: The JuMP model.

state::Union{StateAC, Nothing}: State variables data.

statex::Vector{VariableRef}: References to optimization variables for bus voltages.

statey::Vector{VariableRef}: References to optimization variables for bus voltages.

residualx::Vector{VariableRef}: References to optimization variables for residuals.

residualy::Vector{VariableRef}: References to optimization variables for residuals.

residual::Dict{Int64, ConstraintRef}: References to the residual constraints.

range::Vector{Int64}: Range of measurement devices.

number::Int64: Number of measurement devices.

JuliaGrid.Normal — Type

Normal

An abstract type representing weighted least squares state estimation methods where normal equation is solved. It is used as a type parameter in models such as LWLS and NWLS to indicate that the analysis is based on the normal equation.

JuliaGrid.Orthogonal — Type

Orthogonal

An abstract type representing orthogonal method used in weighted-least squares state estimation. It is used as a type parameter in models such as LWLS and NWLS to indicate that the analysis is based on the orthogonal method.