State Estimation

For further information on this topic, please see the AC State Estimation or DC State Estimation sections of the Manual. Below, we have provided a list of functions that can be utilized for power flow analysis.

Observability Analysis

islandTopologicalFlow

islandTopological

restorationGram!

PMU State Estimation

pmuStateEstimation

solve!

DC State Estimation

dcStateEstimation

solve!

Bad Data Analysis

residualTest!

Utility Function

pmuPlacement

Observability Analysis

JuliaGrid.islandTopologicalFlow — Method

islandTopologicalFlow(system::PowerSystem, meter::Union{Wattmeter, Varmeter})

The function employs a topological method to identify observable islands solely through active power flow measurements from the Wattmeter composite type or reactive power flow measurements from the Varmeter composite type. This results in the formation of disconnected and loop-free subgraphs. To be more specific, the function utilizes wattmeters or varmeters strategically placed on the branches to detect observable islands. To elaborate further, this approach detects islands based on the linear decoupled measurement model.

Arguments

To define flow observable islands, this function requires the composite types PowerSystem and either Wattmeter or Varmeter.

Returns

The function returns an abstract type Island, containing information about the islands:

island: a list enumerating observable islands with indices of buses;

bus: the positions of buses in relation to each island;

tie: tie data associated with buses and branches.

Examples

Find flow islands for the given set of measurement data using wattmeters:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")
statusWattmeter!(system, device; inservice = 15)

islands = islandTopologicalFlow(system, device.wattmeter)

Find flow islands for the given set of measurement data using varmeters:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")
statusVarmeter!(system, device; inservice = 15)

islands = islandTopologicalFlow(system, device.varmeter)

JuliaGrid.islandTopological — Method

islandTopological(system::PowerSystem, meter::Union{Wattmeter, Varmeter})

The function employs a topological method to identify maximal observable islands solely through active power measurements from the Wattmeter composite type or reactive power measurements from the Varmeter composite type. Specifically, it employs wattmeters or varmeters positioned on the branches to pinpoint flow observable islands. Subsequently, these islands are merged based on the available injection measurements obtained from the wattmeters or varmeters. To elaborate further, this approach detects islands based on the linear decoupled measurement model.

Arguments

To define maximal flow observable islands, this function requires the composite types PowerSystem and either Wattmeter or Varmeter.

Returns

The function returns an abstract type Island, containing information about the islands:

island: a list enumerating observable islands with indices of buses;

bus: the positions of buses in relation to each island;

tie: tie data associated with buses and branches.

Examples

Find maximal islands for the given set of measurement data using wattmeters:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")
statusWattmeter!(system, device; inservice = 12)

islands = islandTopological(system, device.wattmeter)

Find maximal islands for the given set of measurement data using varmeters:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")
statusVarmeter!(system, device; inservice = 12)

islands = islandTopological(system, device.varmeter)

JuliaGrid.restorationGram! — Method

restorationGram!(system::PowerSystem, device::Measurement, pseudo::Measurement, 
    islands::Island; threshold)

Upon identifying the islands, the function incorporates measurements from the available pseudo-measurements in the pseudo variable into the device variable to reinstate observability. If the abstract type Island is derived from wattmeters, candidates for restoring observability include active power measurements and bus voltage angle measurements from the pseudo variable. Conversely, if the abstract type Island is derived from varmeters, candidates for restoring observability encompass reactive power measurements and bus voltage magnitude measurements from the pseudo variable. This method relies on reduced coefficient matrices and the Gram matrix.

It is important to note that the device labels in the device and pseudo variables must be different to enable the function to successfully incorporate measurements from pseudo into the device set of measurements.

Arguments

This function requires the composite types PowerSystem and device, which holds measurements from which the islands variable is obtained. To restore observability, the function uses measurements from the pseudo variable and adds a non-redundant set from it to the device variable.

Keyword

The keyword threshold defines the zero pivot threshold value with a default value of 1e-5. More precisely, all computed pivots less than this value will be treated as zero pivots.

Updates

The function updates the device variable of the Measurement composite type.

Example

Restore observability for DC state estimation:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")
statusWattmeter!(system, device; inservice = 10)
statusPmu!(system, device; inservice = 0)

pseudo = measurement("pseudomeasurement14.h5")
islands = islandTopological(system, device.wattmeter)
restorationGram!(system, device, pseudo, islands)

analysis = dcStateEstimation(system, device)
solve!(system, analysis)

PMU State Estimation

JuliaGrid.pmuStateEstimation — Function

pmuStateEstimation(system::PowerSystem, device::Measurement, method; correlated)

The function sets up the framework to solve the linear state estimation model with PMUs only, where the vector of state variables is given in rectangular coordinates.

Arguments

This function establishes the framework by requiring the PowerSystem and Measurement composite types.

Moreover, the presence of the method parameter is not mandatory. It provides various approaches for addressing PMU state estimation. To address the WLS state estimation method, users can opt to utilize factorization techniques to decompose the gain matrix, such as LU, QR, or LDLt, especially when the gain matrix is symmetric. Opting for the Orthogonal method is advisable for a more robust solution in scenarios involving ill-conditioned data, particularly when substantial variations in variances are present.

Alternatively, instead of solving the WLS state estimation problem, users can utilize the LAV method to find an estimator. This can be achieved by selecting one of the optimization solvers, where typically Ipopt.Optimizer suffices for common scenarios.

If the method parameter is not provided, the default method for solving the PMU estimation will be LU factorization using the WLS framework.

Keyword

The boolean keyword correlated defines the correlation between measurement errors of a single PMU and is only relevant for WLS state estimation.

When correlated = false, which is the default setting, the measurement errors of a single PMU are not correlated. This results in the covariance matrix maintaining a diagonal form. In this case, users can also run the Orthogonal method to find the estimate of state variables.

On the other hand, when correlated = true, the covariance matrix does not maintain a diagonal form. Instead, it becomes a block diagonal matrix due to the correlation between measurement errors, and in this case, the Orthogonal method is not allowed. Users can then use LU, QR, or LDLt factorization over the gain matrix to obtain the solution.

Updates

If the AC model has not been created, the function will automatically trigger an update of the ac field within the PowerSystem composite type.

Returns

The function returns an instance of the DCStateEstimation abstract type, which includes the following fields:

voltage: the variable allocated to store the bus voltage magnitudes and angles;

power: the variable allocated to store the active and reactive powers;

method: the system model vectors and matrices, or alternatively, the optimization model;

bad: the variable linked to identifying bad data within the measurement set.

Examples

Set up the PMU state estimation WLS framework to be solved using the default LU factorization method:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = pmuStateEstimation(system, device)

Set up the PMU state estimation WLS framework to be solved using the default LU factorization method, where measurement errors are correlated:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = pmuStateEstimation(system, device; correlated = true)

Set up the PMU state estimation WLS framework to be solved using the orthogonal method:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = pmuStateEstimation(system, device, Orthogonal)

Set up the PMU state estimation LAV framework:

using Ipopt

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = pmuStateEstimation(system, device, Ipopt.Optimizer)

JuliaGrid.solve! — Method

solve!(system::PowerSystem, analysis::PMUStateEstimation)

This function computes the bus voltage magnitudes and angles to solve the state estimation problem exclusively using PMU data.

Updates

The resulting bus voltage magnitudes and angles are stored in the voltage field of the PMUStateEstimation type.

Examples

Solving the PMU state estimation model with WLS:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = pmuStateEstimation(system, device)
solve!(system, analysis)

Solving the PMU state estimation model with LAV:

using Ipopt

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = pmuStateEstimation(system, device, Ipopt.Optimizer)
solve!(system, analysis)

DC State Estimation

JuliaGrid.dcStateEstimation — Function

dcStateEstimation(system::PowerSystem, device::Measurement, method)

The function sets up the framework to solve the DC state estimation.

Arguments

This function requires the PowerSystem and Measurement composite types to establish the framework.

Moreover, the presence of the method parameter is not mandatory. It provides various approaches for addressing DC state estimation. To address the WLS state estimation method, users can opt to utilize factorization techniques to decompose the gain matrix, such as LU, QR, or LDLt, especially when the gain matrix is symmetric. Opting for the Orthogonal method is advisable for a more robust solution in scenarios involving ill-conditioned data, particularly when substantial variations in variances are present.

Alternatively, instead of solving the WLS state estimation problem, users can utilize the LAV method to find an estimator. This can be achieved by selecting one of the optimization solvers, where typically Ipopt.Optimizer suffices for common scenarios.

If the user does not provide the method, the default method for solving the DC estimation will be LU factorization.

Updates

If the DC model was not created, the function will automatically initiate an update of the dc field within the PowerSystem composite type. Additionally, if the slack bus lacks an in-service generator, JuliaGrid considers it a mistake and defines a new slack bus as the first generator bus with an in-service generator in the bus type list.

Returns

The function returns an instance of the DCStateEstimation abstract type, which includes the following fields:

voltage: the variable allocated to store the bus voltage angles;

power: the variable allocated to store the active powers;

method: the system model vectors and matrices, or alternatively, the optimization model;

bad: the variable linked to identifying bad data within the measurement set.

Examples

Set up the DC state estimation WLS framework to be solved using the default LU factorization method:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device)

Set up the DC state estimation WLS framework to be solved using the orthogonal method:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device, Orthogonal)

Set up the DC state estimation LAV framework:

using Ipopt

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device, Ipopt.Optimizer)

JuliaGrid.solve! — Method

solve!(system::PowerSystem, analysis::DCStateEstimation)

By computing the bus voltage angles, the function solves the DC state estimation problem.

Updates

The resulting bus voltage angles are stored in the voltage field of the DCStateEstimation type.

Examples

Solving the DC state estimation model with WLS:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device)
solve!(system, analysis)

Solving the DC state estimation model with LAV:

using Ipopt

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device, Ipopt.Optimizer)
solve!(system, analysis)

Bad Data Analysis

JuliaGrid.residualTest! — Function

residualTest!(system::PowerSystem, device::Measurement, analysis::StateEstimation; 
    threshold)

The function conducts bad data detection and identification using the largest normalized residual test, subsequently removing measurement outliers from the measurement set. It can be executed after obtaining estimation solutions.

Arguments

This function requires the composite types PowerSystem and Measurement, along with an abstract type. The abstract type StateEstimation can have the following subtypes:

DCStateEstimation: conducts bad data analysis within DC state estimation;

PMUStateEstimation: conducts bad data analysis within PMU state estimation.

Keyword

The keyword threshold establishes the identification threshold. If the largest normalized residual surpasses this threshold, the measurement is flagged as bad data. The default threshold value is set to threshold = 3.0.

Updates

In case bad data is detected, the function removes measurements from the coefficient and precision matrices, and mean vector within the DCStateEstimation type. Additionally, it marks the respective measurement within the Measurement type as out-of-service.

Furthermore, the variable outlier within the StateEstimation type stores information regarding bad data detection and identification:

detect: returns true after the function's execution if bad data is detected;

maxNormalizedResidual: denotes the value of the largest normalized residual;

label: signifies the label of the bad data;

index: represents the index of the bad data.

Example

Obtaining the solution after detecting and removing bad data:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device)
solve!(system, analysis)

residualTest!(system, device, analysis; threshold = 4.0)
solve!(system, analysis)

Obtaining the solution while bad data is detected:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device)

while analysis.bad.detect
    solve!(system, analysis)
    residualTest!(system, device, analysis; threshold = 4.0)
end

Utility Function

JuliaGrid.pmuPlacement — Function

pmuPlacement(system::PowerSystem, optimizer; bridge)

The function determines the optimal placement of PMUs through integer linear programming. Specifically, it identifies the minimum set of PMU locations required for effective power system state estimation, ensuring observability with the least number of PMUs.

The function accepts a PowerSystem composite type as input to establish the framework for finding the optimal PMU placement. If the ac field within the PowerSystem composite type is not yet created, the function automatically initiates an update process.

Additionally, the optimizer argument is a crucial component for formulating and solving the optimization problem. Typically, using the GLPK or HiGHS solver is sufficient. For more detailed information, please refer to the JuMP JuMP documenatation.

Keyword

The bridge keyword enables users to manage the bridging mechanism within the JuMP package.

Returns

The function returns an instance of the PlacementPMU type, containing variables such as:

bus: bus labels with indices marking the positions of PMUs at buses;

from: branch labels with indices marking the positions of PMUs at "from" bus ends;

to: branch labels with indices marking the positions of PMUs at "to" bus ends.

Note that if the conventional understanding of a PMU involves a device measuring the bus voltage phasor and all branch current phasors incident to the bus, the result is saved only in the bus variable. However, if we consider that a PMU measures individual phasors, each described with magnitude and angle, then measurements are needed at each bus in the bus variable, and each branch with positions given according to from and to variables.

Example

using GLPK, Ipopt

system = powerSystem("case14.h5")
analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)
current!(system, analysis)

V = analysis.voltage.magnitude
θ = analysis.voltage.angle
If = analysis.current.from.magnitude
ψf = analysis.current.from.angle
It = analysis.current.to.magnitude
ψt = analysis.current.to.angle

placement = pmuPlacement(system, GLPK.Optimizer)
device = measurement()

@pmu(label = "PMU ?: !")
for (label, i) in placement.bus
    addPmu!(system, device; bus = label, magnitude = V[i], angle = θ[i])
end
for (label, i) in placement.from
    addPmu!(system, device; from = label, magnitude = If[i], angle = ψf[i])
end
for (label, i) in placement.to
    addPmu!(system, device; to = label, magnitude = It[i], angle = ψt[i])
end