Unit System

For further information on this topic, please see the Power System Model section of the Manual.

Base

JuliaGrid.@base — Macro

By default, the units for base power and base voltages are set to volt-ampere (VA) and volt (V), but you can modify the prefixes using the macro:

@base(system::PowerSystem, power, voltage)

Prefixes must be specified according to the SI prefixes and should be included with the unit of power (VA) or unit of voltage (V). Keep in mind that the macro must be used after creating the composite type PowerSystem.

Example

system = powerSystem("case14.h5")
@base(system, MVA, kV)

Power

JuliaGrid.@power — Macro

JuliaGrid stores all data related with powers in per-units, and these cannot be altered. However, the power units of the built-in functions used to add or modified power system elements can be modified using the macro:

@power(active, reactive, apparent)

Prefixes must be specified according to the SI prefixes and should be included with the unit of active power (W), reactive power (VAr), or apparent power (VA). Also it is a possible to combine SI units with/without prefixes with per-units (pu).

Changing the unit of active power is reflected in the following quantities:

addBus!: active, conductance

shuntBus!: conductance

addGenerator!: active, minActive, maxActive, lowActive, upActive, loadFollowing, reserve10min, reserve30min

addActiveCost!: piecewise, polynomial

outputGenerator!: active

addBranch!: longTerm, shortTerm, emergency if rating type = 2.

Changing the unit of reactive power unit is reflected in the following quantities:

addBus!: reactive, susceptance

shuntBus!: susceptance

addGenerator!: reactive, minReactive, maxReactive, minLowReactive, maxLowReactive, minUpReactive, maxUpReactive, reactiveTimescale

addReactiveCost!: piecewise, polynomial

outputGenerator!: reactive.

Changing the unit of apparent power unit is reflected in the following quantities:

addBranch!: longTerm, shortTerm, emergency if rating type = 1 or type = 3.

Example

@power(MW, kVAr, VA)

Voltage

JuliaGrid.@voltage — Macro

JuliaGrid stores all data related with voltages in per-units and radians, and these cannot be altered. However, the voltage magnitude and angle units of the built-in functions used to add or modified power system elements can be modified using the macro:

@voltage(magnitude, angle)

Prefixes must be specified according to the SI prefixes and should be included with the unit of voltage magnitude (V). The second option is to define the unit of voltage magnitude in per-unit (pu). The unit of the voltage angle should be given in radian (rad) or degree (deg).

Changing the unit of voltage magnitude is reflected in the following quantities:

addBus!: magnitude, minMagnitude, maxMagnitude

addGenerator!: magnitude.

Changing the unit of voltage angle is reflected in the following quantities:

addBus!: angle

addBranch!: shiftAngle, minDiffAngle, maxDiffAngle

parameterBranch!: shiftAngle.

Example

@voltage(kV, deg)

Parameter

JuliaGrid.@parameter — Macro

JuliaGrid stores all data related with impedances and admittancies in per-units, and these cannot be altered. However, units of impedance and admittance of the built-in functions used to add or modified power system elements can be modified using the macro:

@parameter(impedance, admittance)

Prefixes must be specified according to the SI prefixes and should be included with the unit of impedance (Ω) or unit of admittance (S). The second option is to define the units in per-unit (pu).

In the case where impedance and admittance are being used in SI units (Ω and S) and these units are related to the transformer, the assignment must be based on the primary side of the transformer.

Changing the units of impedance is reflected in the following quantities in specific functions:

addBranch!: resistance, reactance

parameterBranch!: resistance, reactance.

Changing the units of admittance is reflected in the following quantities:

addBranch!: susceptance

parameterBranch!: susceptance.

Example

@parameter(Ω, pu)

Default

JuliaGrid.@default — Macro

We have a macro that resets the units of the built-in functions to their default settings of per-units (pu) and radians (rad).

@default(mode)

The mode argument can take on the following values:

all: resets all units to their default settings

power: sets active, reactive, and apparent power to per-units (pu)

voltage: sets voltage magnitude to per-unit (pu) and voltage angle to radian (rad)

parameter: sets impedance and admittance to per-units (pu).

Example

@default(all)