Wrapper Functions

For further information on this topic, please see the Power System Model section of the Manual.

API Index

mismatch!

solve!

AC Power Flow Mismatches

JuliaGrid.mismatch! — Function

The function calculates both active and reactive power injection mismatches and returns their maximum absolute values. These maximum values are useful for stopping the iteration loop of the AC power flow problem-solving method.

mismatch!(system::PowerSystem, analysis::Analysis)

This function is designed to be used within the iteration loop before calling the solve! function.

The function acts as a wrapper and relies on the method chosen for solving the AC power flow problem. Depending on the selected method, the function calls one of these functions:

mismatchNewtonRaphson!

mismatchFastNewtonRaphson!

mismatchGaussSeidel!

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
mismatch!(system, analysis)

Solve

JuliaGrid.solve! — Function

The function serves as a wrapper that integrates multiple functions to solve AC or DC power flow problems.

solve!(system::PowerSystem, analysis::Analysis)

It then updates the bus.voltage and model fields of the Analysis composite type accordingly.

To use this function in the AC power flow framework, it should be employed within the iteration loop following the mismatch! function. Together, these functions execute a single iteration of the Newton-Raphson or fast Newton-Raphson method. If the Gauss-Seidel method is used, this function alone is sufficient to calculate the bus voltages.

Depending on the selected method or type of the power flow, the function calls one of these functions:

solveNewtonRaphson!

solveFastNewtonRaphson!

solveGaussSeidel!

solveDCPowerFlow!

Examples

AC Power Flow:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

DC Power Flow:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system)
solve!(system, analysis)