Build Power System Model

The main composite type PowerSystem with fields bus, branch, generator, acModel, dcModel, and basePower can be created using a method:

powerSystem().

Once the model is created, it is possible to add buses, branches and generators using functions:

addBus!()

addBranch!()

addGenerator!().

The final step is the formation and saving of vectors and matrices obtained based on the power system topology and parameters using functions:

acModel!()

dcModel!().

Note that, once the field acModel and dcModel are formed, using function addBranch!(), will automatically trigger the update of these fields. In contrast, adding a new bus, using addBus!(), requires executing the functions acModel!() and dcModel!() again.

Then, it is possible to manipulate the parameters of buses, branches and generators using functions:

shuntBus!()

statusBranch!()

parameterBranch!()

statusGenerator!()

outputGenerator!().

The execution of these functions will automatically trigger the update of all subtypes affected by these functions.

Build Model

The method powerSystem() builds the main composite type PowerSystem and populate fields bus, branch, generator and basePower.

JuliaGrid.powerSystem — Method

The path to the HDF5 file with .h5 extension should be passed to the method:

powerSystem("pathToExternalData/name.h5")

Similarly, the path to the Matpower file with .m extension should be passed to the same method:

powerSystem("pathToExternalData/name.m")

Ignoring the method argument initializes the main composite type PowerSystem, which enables building the model from scratch:

powerSystem()

Once the main composite type PowerSystem is created it is possible to add new buses, branches or generators, and also change the parameters of the existing ones.

Bus Functions

Functions receives the main composite type PowerSystem and arguments by keyword to set or change bus parameters, and affect field bus.

JuliaGrid.addBus! — Function

The function add a new bus. Names, descriptions and units of keywords are given in the table bus group.

addBus!(system::PowerSystem; label, slackLabel = 0, area = 1, lossZone = 1,
    active = 0.0, reactive = 0.0, conductance = 0.0, susceptance = 0.0,
    magnitude: = 1.0, angle: = 0.0, minMagnitude = 0.9, maxMagnitude = 1.1,
    base = 0.0)

The keyword label is mandatory. The slack bus, using the keyword slackLabel, can be specified in each function call with the label of the bus being defined or already existing. If the bus is not defined as the slack, the function addBus!() automatically defines the bus as the demand bus (PQ). If a generator is connected to a bus, using the function addGenerator!(), the bus becomes a generator bus (PV).

JuliaGrid.shuntBus! — Function

The function allows changing conductance and susceptance parameters of the shunt element connected to the bus.

shuntBus!(system::PowerSystem; label, conductance, susceptance)

The keyword label should correspond to the already defined bus label. Keywords conductance or susceptance can be omitted, then the value of the omitted parameter remains unchanged. The function also updates the field acModel, if field exist.

Branch Functions

Functions receives the main composite type PowerSystem and arguments by keyword to set or change branch parameters. Further, functions affect field branch, but also fields acModel and dcModel. More precisely, once acModel and dcModel are created, the execution of functions will automatically trigger the update of these fields.

JuliaGrid.addBranch! — Function

The function add a new branch. Names, descriptions and units of keywords are given in the table branch group. A branch can be added between already defined buses.

addBranch!(system::PowerSystem; label, from, to, status = 1,
    resistance = 0.0, reactance = 0.0, susceptance = 0.0, turnsRatio = 0.0,
    shiftAngle = 0.0, longTerm = 0.0, shortTerm = 0.0, emergency = 0.0,
    minAngleDifference = -2*pi, maxAngleDifference = 2*pi)

The keywords label, from, to, and one of the parameters resistance or reactance are mandatory.

JuliaGrid.statusBranch! — Function

The function allows changing the operating status of the branch, from in-service to out-of-service, and vice versa.

statusBranch!(system::PowerSystem; label, status = 0)

The keywords label should correspond to the already defined branch label.

JuliaGrid.parameterBranch! — Function

The function parameterBranch! allows changing resistance, reactance, susceptance, turnsRatio and shiftAngle parameters of the branch. parameterBranch!(system; label, resistance, reactance, susceptance, turnsRatio, shiftAngle)

The keywords label should correspond to the already defined branch label. Keywords resistance, reactance, susceptance, turnsRatio or shiftAngle can be omitted, then the value of the omitted parameter remains unchanged.

Generator Functions

Functions receives the main composite type PowerSystem and arguments by keyword to set or change generator parameters. Further, functions affect fields generator and bus.

JuliaGrid.addGenerator! — Function

The function add a new generator. Names, descriptions and units of keywords are given in the table generator group. A generator can be added at already defined bus.

addGenerator!(system::PowerSystem; label, bus, area::Float64 = 0.0, status = 1,
    active::Float64 = 0.0, reactive = 0.0, magnitude = 1.0,
    minActive = 0.0, maxActive = Inf64, minReactive = -Inf64, maxReactive = Inf64,
    lowerActive = 0.0, minReactiveLower = 0.0, maxReactiveLower = 0.0,
    upperActive = 0.0, minReactiveUpper = 0.0, maxReactiveUpper = 0.0,
    loadFollowing = 0.0, reserve10minute = 0.0, reserve30minute = 0.0, reactiveTimescale = 0.0,
    activeModel = 2, activeStartup = 0.0, activeShutdown = 0.0, activeDataPoint = 3,
    activeCoefficient = Float64[],
    reactiveModel = 2, reactiveStartup = 0.0, reactiveShutdown = 0.0, reactiveDataPoint = 3,
    reactiveCoefficient = Float64[])

The keywords label and bus are mandatory.

JuliaGrid.statusGenerator! — Function

The function allows changing the operating status of the generator, from in-service to out-of-service, and vice versa.

statusGenerator!(system; label, status = 0)

The keywords label should correspond to the already defined generator label.

JuliaGrid.outputGenerator! — Function

The function allows changing active and reactive output power of the generator. outputGenerator!(system; label, active, reactive)

The keywords label should correspond to the already defined generator label. Keywords active or reactive can be omitted, then the value of the omitted parameter remains unchanged.

Build AC or DC Model

The functions receives the main composite type PowerSystem and forms vectors and matrices related with AC or DC simulations.

JuliaGrid.acModel! — Function

We advise the reader to read the section in-depth AC Model, that explains all the data involved in the field acModel.

acModel!(system::PowerSystem)

The function affects field acModel. Once formed, the field will be automatically updated when using functions addBranch!(), shuntBus!(), statusBranch!() parameterBranch!().

JuliaGrid.dcModel! — Function

We advise the reader to read the section in-depth DC Model, that explains all the data involved in the field dcModel.

dcModel!(system::PowerSystem)

The function affects field dcModel. Once formed, the field will be automatically updated when using functions addBranch!(), statusBranch!(), parameterBranch!().

Build DC Model

The function dcModel!() receives the main composite type PowerSystem and forms vectors and matrices related with DC simulations:

dcModel!(system)

The function affects field acModel. Once formed, the field will be automatically updated when using functions addBranch!(), statusBranch!() parameterBranch!(). We advise the reader to read the section in-depth DC Model, that explains all the data involved in the field dcModel.

Modifying Other Parameters

Changing other parameters of the power system can be done by changing variables by accessing their values in fields bus, branch and generator of the main type powerSystem.