Operating State of Branches

Once the bus voltages are obtained from one of the analyses, it is possible to determine powers nad currents related to the branches. The function receives composite types PowerSystem and Result.

JuliaGrid.branch! — Function

The function computes powers and currents related to branches for the AC power flow analysis. For the DC power flow analysis, the function computes only active powers.

branch!(system::PowerSystem, result::Result)

The function updates the field result.branch of the composite type Result.

AC Power Flow Example

system = powerSystem("case14.h5")
acModel!(system)

result = gaussSeidel(system)
stopping = result.algorithm.iteration.stopping
for i = 1:200
    gaussSeidel!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

branch!(system, result)

DC Power Flow Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)
branch!(system, result)