Power Flow Analysis

The calculation of bus voltages is essential to solving the power flow problem. The composite type PowerSystem, which includes bus, branch, and generator fields, is required to obtain a solution. Additionally, depending on the type of power flow used, either acModel or dcModel must be used.

After creating the composite type PowerSystem, the next step is to create the composite type Result, which has fields bus, branch, generator, and algorithm. In the DC power flow, Result is created when determining the bus voltage angles using the dcPowerFlow() function. On the other hand, the AC power flow requires the iterative method to be initialized, which is when the composite type Result is created using one of the following functions:

newtonRaphson()

fastNewtonRaphsonBX()

fastNewtonRaphsonXB()

gaussSeidel().

To calculate bus voltages, the appropriate function can be used depending on the type of power flow and method selected. The following functions are available:

newtonRaphson!()

fastNewtonRaphson!()

gaussSeidel!()

dcPowerFlow().

Note that when solving the AC power flow problem, the methods should be called inside a loop to simulate an iterative process.

JuliaGrid offers a set of post-processing analysis functions for calculating powers, losses, and currents associated with buses, branches, or generators after obtaining AC or DC power flow solutions:

bus!()

branch!()

generator!().

It's important to note that complex currents are stored in the polar coordinate system, while complex powers are stored in the rectangular coordinate system within JuliaGrid.

The JuliaGrid package provides two additional functions. One function validates the reactive power limits of generators once the AC power flow solution has been computed. The other function adjusts the voltage angles to match the angle of an arbitrary slack bus:

reactivePowerLimit!()

adjustVoltageAngle!().

Newton-Raphson Method

JuliaGrid.newtonRaphson — Function

The function accepts the PowerSystem composite type as input, uses it to set up the Newton-Raphson method, and then produces the Result composite type as output.

newtonRaphson(system::PowerSystem)

The algorithm field of the Result type is updated during the function's execution.

JuliaGrid.newtonRaphson! — Function

The function updates the bus.voltage and algorithm fields of the Result composite type by computing the magnitudes and angles of bus voltages using the Newton-Raphson method.

newtonRaphson!(system::PowerSystem, result::Result)

It is intended to be used within a for loop as it performs only one iteration of the Newton-Raphson method. It is recommended that the reader peruses the section on the Newton-Raphson Method to gain a comprehensive understanding of its implementation, including all relevant data.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:10
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

Fast Newton-Raphson Method

JuliaGrid.fastNewtonRaphsonBX — Function

The function accepts the PowerSystem composite type as input, uses it to set up the fast Newton-Raphson method of version BX, and then produces the Result composite type as output.

fastNewtonRaphsonBX(system::PowerSystem)

The algorithm field of the Result type is updated during the function's execution.

JuliaGrid.fastNewtonRaphsonXB — Function

The function accepts the PowerSystem composite type as input, uses it to set up the fast Newton-Raphson method of version XB, and then produces the Result composite type as output.

fastNewtonRaphsonXB(system::PowerSystem)

The algorithm field of the Result type is updated during the function's execution.

JuliaGrid.fastNewtonRaphson! — Function

The function updates the bus.voltage and algorithm fields of the Result composite type by computing the magnitudes and angles of bus voltages using the fast Newton-Raphson method.

fastNewtonRaphson!(system::PowerSystem, result::Result)

It is intended to be used within a for loop as it performs only one iteration of the fast Newton-Raphson method. It is recommended that the reader peruses the section on the Fast Newton-Raphson Method to gain a comprehensive understanding of its implementation, including all relevant data.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = fastNewtonRaphsonBX(system)
stopping = result.algorithm.iteration.stopping
for i = 1:100
    fastNewtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

Gauss-Seidel Method

JuliaGrid.gaussSeidel — Function

The function accepts the PowerSystem composite type as input, uses it to set up the Gauss-Seidel method, and then produces the Result composite type as output.

gaussSeidel(system::PowerSystem)

The algorithm field of the Result type is updated during the function's execution.

JuliaGrid.gaussSeidel! — Function

The function updates the bus.voltage and algorithm fields of the Result composite type by computing the magnitudes and angles of bus voltages using the Gauss-Seidel method.

gaussSeidel!(system::PowerSystem, result::Result)

It is intended to be used within a for loop as it performs only one iteration of the Gauss-Seidel method. It is recommended that the reader peruses the section on the Gauss-Seidel Method to gain a comprehensive understanding of its implementation, including all relevant data.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = gaussSeidel(system)
stopping = result.algorithm.iteration.stopping
for i = 1:1000
    gaussSeidel!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

DC Power Flow Solution

JuliaGrid.dcPowerFlow — Function

The function takes a PowerSystem composite type as input and uses it to solve the DC power flow problem by calculating the voltage angles for each bus.

dcPowerFlow(system::PowerSystem)

The function returns a composite type Result as output, which includes updated bus.voltage.angle and algorithm fields. These fields are modified during the execution of the function. It is recommended that the reader peruses the section on the DC Power Flow Solution to gain a comprehensive understanding of its implementation, including all relevant data.

Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)

Post-processing Analysis

JuliaGrid.bus! — Function

This function calculates the powers and currents associated with buses.

bus!(system::PowerSystem, result::Result)

After the function is executed, the bus field within the Result type gets updated. To gain a comprehensive understanding of the equations used, along with all relevant data, it is recommended that the reader examines either the section on AC Bus Post-processing Analysis or the section on DC Bus Post-processing Analysis.

AC Power Flow Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:10
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

bus!(system, result)

DC Power Flow Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)
bus!(system, result)

JuliaGrid.branch! — Function

The function is used to calculate the powers and currents associated with branches for AC power flow analysis and, in relation to DC power flow analysis, it only calculates active powers.

branch!(system::PowerSystem, result::Result)

The function is responsible for updating the branch field within the Result type after it has been executed. To gain a comprehensive understanding of the equations used, along with all relevant data, it is recommended that the reader examines either the section on AC Branch Post-processing Analysis or the section on DC Branch Post-processing Analysis.

AC Power Flow Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:10
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

branch!(system, result)

DC Power Flow Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)
branch!(system, result)

JuliaGrid.generator! — Function

The function computes powers related to generators.

generator!(system::PowerSystem, result::Result)

Once executed, the generator field within the Result type is updated accordingly. To gain a comprehensive understanding of the equations used, along with all relevant data, it is recommended that the reader examines either the section on AC Generator Post-processing Analysis or the section on DC Generator Post-processing Analysis.

AC Power Flow Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:10
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

generator!(system, result)

DC Power Flow Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)
generator!(system, result)

Additional Functions

JuliaGrid.reactivePowerLimit! — Function

The function verifies whether the generators in a power system exceed their reactive power limits. This is done by setting the reactive power of the generators to within the limits if they are violated, after determining the bus voltage magnitudes and angles. If the limits are violated, the corresponding PV buses or the slack bus are converted to PQ buses.

The function returns the violate variable to indicate which buses violate the limits, with -1 indicating a violation of the minimum limits and 1 indicating a violation of the maximum limits.

reactivePowerLimit!(system::PowerSystem, result::Result)

First, if the generator!() function has not been executed, reactivePowerLimit!() will execute it and update the generator field of the Result type.

Afterward, the function uses the results from generator!() to assign values to the generator.output.active and bus.supply.active fields of the System type.

At the end of the process, the function inspects the reactive powers of the generator and adjusts them to their maximum or minimum values if they violate the threshold. The generator.output.reactive field of the System type is then modified accordingly. In light of this modification, the bus.supply.reactive field of the System type is also updated, and the bus types in bus.layout.type are adjusted. If the slack bus is converted, the bus.layout.slack field is modified accordingly.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:200
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

violate = reactivePowerLimit!(system, result)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:200
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

JuliaGrid.adjustVoltageAngle! — Function

The function modifies the bus voltage angles based on a different slack bus than the one identified by the bus.layout.slack field. This function only updates the bus.voltage.angle field of the Result type.

adjustVoltageAngle!(system::PowerSystem, result::Result; slack)

For instance, if the reactive power of the generator exceeds the limit on the slack bus, the reactivePowerLimit!() function will change that bus to a PQ bus and designate the first PV bus in the sequence as the new slack bus. After obtaining the updated AC power flow solution based on the new slack bus, it is possible to adjust the voltage angles to align with the angle of the original slack bus. The slack keyword specifies the bus label of the original slack bus.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:200
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

reactivePowerLimit!(system, result)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:200
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

adjustVoltageAngle!(system, result; slack = 1)