Generator Reactive Power Limits

By default, the AC power flow methods solve the system of non-linear equations and reveal bus voltage magnitudes and angles ignoring any limits. However, JuliaGrid provides a function reactivePowerLimit!() that checks reactive power limits.

JuliaGrid.reactivePowerLimit! — Function

The function checks reactive power limits of the generators. More precisely, after the bus voltage magnitudes and angles have been determined, the function sets the reactive power of the generators within the limits if they are violated, and corresponding PV buses or slack bus will be converted to PQ buses.

reactivePowerLimit!(system::PowerSystem, result::Result)

The function updates the field system.generator.layout.violate to indicate on which buses the limits are violated. If the minimum limits are violated, the mark -1 will appear in the appropriate place, the violation of the maximum limits is marked with 1.

Further, the function updates fields system.generator.output.reactive, system.bus.supply.reactive,and system.bus.layout.type. For the case when the slack bus is converted the function updates the field system.bus.layout.slackIndex.

In case the function generator!() is not executed, the function will trigger the execution of this function and will update the field result.generator.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:200
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

reactivePowerLimit!(system, result)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:200
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

Furthermore, if there is a slack bus conversion, it is possible to adjust the voltage angles according to the original slack bus.

JuliaGrid.adjustVoltageAngle! — Function

The function adjust bus voltage angles according to the original slack bus. More precisely, in the case when the reactive power of the generator is beyond the limit on the slack bus, the function reactivePowerLimit!() will convert that bus into PQ bus, and declare the first PV bus in the list as the new slack bus. Then, after the new AC power flow solution is determined, it is possible to correct the voltage angles, so that their values are in accordance with the angle of the original slack bus.

adjustVoltageAngle!(system::PowerSystem, result::Result)

The function updates field result.bus.voltage.angle.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:200
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

reactivePowerLimit!(system, result)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:200
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

adjustVoltageAngle!(system, result)