Power Flow Solution

The calculation of bus voltages is essential to solving the power flow problem. The composite type PowerSystem, which includes bus, branch, and generator fields, is required to obtain a solution. Additionally, depending on the type of power flow used, either acModel or dcModel must be used.

After creating the composite type PowerSystem, the next step is to create the composite type Result, which has fields bus, branch, generator, and algorithm. In the DC power flow, Result is created when determining the bus voltage angles using the dcPowerFlow() function. On the other hand, the AC power flow requires the iterative method to be initialized, which is when the composite type Result is created using one of the following functions:

newtonRaphson()

fastNewtonRaphsonBX()

fastNewtonRaphsonXB()

gaussSeidel().

To calculate bus voltages, the appropriate function can be used depending on the type of power flow and method selected. The following functions are available:

newtonRaphson!()

fastNewtonRaphson!()

gaussSeidel!()

dcPowerFlow().

Note that when solving the AC power flow problem, the methods should be called inside a loop to simulate an iterative process.

Additionally, JuliaGrid offers a range of functions for conducting power flow analysis, which involves calculating the powers and currents associated with buses, branches, or generators. These functions include:

bus!()

branch!()

generator!().

The JuliaGrid package also includes a function for verifying the reactive power limits of generators after the AC power flow solution has been obtained:

reactivePowerLimit!()

Newton-Raphson Method

JuliaGrid.newtonRaphson — Function

The function accepts the PowerSystem composite type as input, uses it to set up the Newton-Raphson method, and then produces the Result composite type as output.

newtonRaphson(system::PowerSystem)

The algorithm field of the Result type is updated during the function's execution.

JuliaGrid.newtonRaphson! — Function

The function updates the bus.voltage and algorithm fields of the Result composite type by computing the magnitudes and angles of bus voltages using the Newton-Raphson method.

newtonRaphson!(system::PowerSystem, result::Result)

It is intended to be used within a for loop as it performs only one iteration of the Newton-Raphson method.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:10
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

Fast Newton-Raphson Method

JuliaGrid.fastNewtonRaphsonBX — Function

The function accepts the PowerSystem composite type as input, uses it to set up the fast Newton-Raphson method of version BX, and then produces the Result composite type as output.

fastNewtonRaphsonBX(system::PowerSystem)

The algorithm field of the Result type is updated during the function's execution.

JuliaGrid.fastNewtonRaphsonXB — Function

The function accepts the PowerSystem composite type as input, uses it to set up the fast Newton-Raphson method of version XB, and then produces the Result composite type as output.

fastNewtonRaphsonXB(system::PowerSystem)

The algorithm field of the Result type is updated during the function's execution.

JuliaGrid.fastNewtonRaphson! — Function

The function updates the bus.voltage and algorithm fields of the Result composite type by computing the magnitudes and angles of bus voltages using the fast Newton-Raphson method.

fastNewtonRaphson!(system::PowerSystem, result::Result)

It is intended to be used within a for loop as it performs only one iteration of the fast Newton-Raphson method.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = fastNewtonRaphsonBX(system)
stopping = result.algorithm.iteration.stopping
for i = 1:100
    fastNewtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

Gauss-Seidel Method

JuliaGrid.gaussSeidel — Function

The function accepts the PowerSystem composite type as input, uses it to set up the Gauss-Seidel method, and then produces the Result composite type as output.

gaussSeidel(system::PowerSystem)

The algorithm field of the Result type is updated during the function's execution.

JuliaGrid.gaussSeidel! — Function

The function updates the bus.voltage and algorithm fields of the Result composite type by computing the magnitudes and angles of bus voltages using the Gauss-Seidel method.

gaussSeidel!(system::PowerSystem, result::Result)

It is intended to be used within a for loop as it performs only one iteration of the Gauss-Seidel method.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = gaussSeidel(system)
stopping = result.algorithm.iteration.stopping
for i = 1:1000
    gaussSeidel!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

DC Power Flow Solution

JuliaGrid.dcPowerFlow — Function

The function takes a PowerSystem composite type as input and uses it to solve the DC power flow problem by calculating the voltage angles for each bus.

dcPowerFlow(system::PowerSystem)

The function returns a composite type Result as output, which includes updated bus.voltage.angle and algorithm fields. These fields are modified during the execution of the function.

Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)