Build Power System Model

The composite type PowerSystem with fields bus, branch, generator, acModel, dcModel, and basePower can be created using a function:

powerSystem().

Once the model is created, it is possible to add buses, branches and generators using the functions:

addBus!()

addBranch!()

addGenerator!().

In addition, it is possible to manipulate the parameters of buses, branches and generators using functions:

shuntBus!()

statusBranch!()

parameterBranch!()

statusGenerator!()

outputGenerator!().

The execution of these functions will automatically trigger the update of all fields affected by these functions. Changing other parameters of the power system can be done by changing variables by accessing their values in fields bus, branch, generator, and basePower of the composite type PowerSystem.

At any moment after the composite type PowerSystem is formed, it is possible to save the data in HDF5 format that describe the power system:

savePowerSystem().

To create vectors and matrices obtained based on the power system topology and parameters, the following functions are used:

acModel!()

dcModel!().

Note that, functions can be executed at any time, and when all power system buses are defined. More precisely, once the field acModel and dcModel are formed, using function addBranch!(), will automatically trigger the update of these fields. In contrast, adding a new bus, using addBus!(), requires executing the functions acModel!() and dcModel!() again. Further, the execution of the functions related to parameter manipulation of buses and branches will automatically trigger the update of the fields acModel and dcModel.

Build Model

The function builds the composite type PowerSystem and populates fields bus, branch, generator and basePower.

JuliaGrid.powerSystem — Function

To use the powerSystem() function, the path to the HDF5 file with the .h5 extension can be passed as arguments. For example:

powerSystem("pathToExternalData/name.h5")

Similarly, the path to the Matpower file with the .m extension can be passed to the same function:

powerSystem("pathToExternalData/name.m")

Alternatively, the PowerSystem composite type can be initialized by calling the function without any arguments:

powerSystem()

This allows the model to be built from scratch and modified as needed.

In general, once the composite type PowerSystem has been created, it is possible to add new buses, branches, or generators, or modify the parameters of existing ones.

Example

system = powerSystem("case14.h5")

Bus Functions

Functions receive the composite type PowerSystem and arguments by keyword to set or change bus parameters and update the field bus.

JuliaGrid.addBus! — Function

The function adds a new bus and updates the field system.bus.

addBus!(system::PowerSystem; label::Int64,
    active::Float64, reactive::Float64,
    conductance::Float64, susceptance::Float64,
    magnitude::Float64, angle::Float64,
    minMagnitude::Float64, maxMagnitude::Float64,
    base::Float64, area::Int64, lossZone::Int64)

Descriptions, types and units of keywords are given below:

label - unique bus label

active - active power demand

reactive - reactive power demand

conductance - active power demanded of the shunt element

susceptance - reactive power injected of the shunt element

magnitude - initial value of the voltage magnitude

angle - initial value of the voltage angle

minMagnitude - minimum voltage magnitude value

maxMagnitude - maximum voltage magnitude value

base - base value of the voltage magnitude

area - area number

lossZone - loss zone

The function automatically defines the bus as the demand bus. The defined bus can become a generator bus by creating a generator using the function addGenerator!(). That is, the slack bus using the function slackBus().

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)

JuliaGrid.shuntBus! — Function

The function allows changing conductance and susceptance parameters of the shunt element connected to the bus.

shuntBus!(system::PowerSystem; label::Int64, conductance::Float64, susceptance::Float64)

The keyword label should correspond to the already defined bus label. Keywords conductance or susceptance can be omitted, then the value of the omitted parameter remains unchanged.

The usefulness of the function is that its execution automatically updates the field acModel. That is, when changing these parameters, it is not necessary to create this model from scratch.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
shuntBus!(system; label = 1, conductance = 0.04)

Branch Functions

Functions receive the composite type PowerSystem and arguments by keyword to set or change branch parameters. Further, functions update the field branch, but also fields acModel and dcModel. More precisely, once acModel and dcModel are created, the execution of functions will automatically trigger the update of these fields.

JuliaGrid.addBranch! — Function

The function adds a new branch and updates the field system.branch. A branch can be added between already defined buses.

addBranch!(system::PowerSystem; label::Int64, from::Int64, to::Int64,
    resistance::Float64 = 0.0, reactance::Float64 = 0.0, susceptance::Float64 = 0.0,
    turnsRatio::Float64 = 0.0, shiftAngle::Float64 = 0.0,
    longTerm::Float64 = 0.0, shortTerm::Float64 = 0.0, emergency::Float64 = 0.0,
    minAngleDifference::Float64 = -2*pi, maxAngleDifference::Float64 = 2*pi,
    status::Int64 = 1)

Descriptions, types and units of keywords are given below:

label - unique branch label (positive integer)

from - from bus label, corresponds to the bus label

to - to bus label, corresponds to the bus label

resistance - branch resistance (per-unit)

reactance - branch reactance (per-unit)

susceptance - total line charging susceptance (per-unit)

turnsRatio - transformer off-nominal turns ratio, equal to zero for a line

shiftAngle - transformer phase shift angle, where positive value defines delay (radian)

longTerm - short-term rating (equal to zero for unlimited)

shortTerm - long-term rating (equal to zero for unlimited)

emergency - emergency rating (equal to zero for unlimited)

minAngleDifference - minimum voltage angle difference value between from and to bus (radian)

maxAngleDifference - maximum voltage angle difference value between from and to bus (radian)

status - operating status of the branch, in-service = 1, out-of-service = 0

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
addBus!(system; label = 2, active = 0.15, reactive = 0.08)

addBranch!(system; label = 1, from = 1, to = 2, resistance = 0.05, reactance = 0.12)

JuliaGrid.statusBranch! — Function

The function allows changing the operating status of the branch, from in-service to out-of-service, and vice versa.

statusBranch!(system::PowerSystem; label::Int64, status::Int64)

The keywords label should correspond to the already defined branch label.

The usefulness of the function is that its execution automatically updates the fields acModel and dcModel. That is, when changing the operating status of the branch, it is not necessary to create models from scratch.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
addBus!(system; label = 2, active = 0.15, reactive = 0.08)

addBranch!(system; label = 1, from = 1, to = 2, resistance = 0.05, reactance = 0.12)
statusBranch!(system; label = 1, status = 0)

JuliaGrid.parameterBranch! — Function

The function allows changing resistance, reactance, susceptance, turnsRatio and shiftAngle parameters of the branch.

parameterBranch!(system::PowerSystem; label::Int64,
    resistance::Float64, reactance::Float64, susceptance::Float64,
    turnsRatio::Float64, shiftAngle::Float64)

The keywords label should correspond to the already defined branch label. Keywords resistance, reactance, susceptance, turnsRatio or shiftAngle can be omitted, and then the value of the omitted parameter remains unchanged.

The usefulness of the function is that its execution automatically updates the fields acModel and dcModel. That is, when changing these parameters, it is not necessary to create models from scratch.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
addBus!(system; label = 2, active = 0.15, reactive = 0.08)

addBranch!(system; label = 1, from = 1, to = 2, resistance = 0.05, reactance = 0.12)
parameterBranch!(system; label = 1, susceptance = 0.062)

Generator Functions

Functions receive the composite type PowerSystem and arguments by keyword to set or change generator parameters. Further, functions update fields generator and bus.

JuliaGrid.addGenerator! — Function

The function adds a new generator and updates the field system.generator. A generator can be added to an already defined bus.

addGenerator!(system::PowerSystem; label::Int64, bus::Int64,
    active::Float64 = 0.0, reactive::Float64 = 0.0, magnitude::Float64 = 0.0,
    minActive::Float64 = 0.0, maxActive::Float64 = Inf,
    minReactive::Float64 = -Inf, maxReactive::Float64 = Inf,
    lowerActive::Float64 = 0.0,
    minReactiveLower::Float64 = 0.0, maxReactiveLower::Float64 = 0.0,
    upperActive::Float64 = 0.0,
    minReactiveUpper::Float64 = 0.0, maxReactiveUpper::Float64 = 0.0,
    loadFollowing::Float64 = 0.0, reactiveTimescale::Float64 = 0.0
    reserve10minute::Float64 = 0.0, reserve30minute::Float64 = 0.0,
    area::Float64 = 0.0, status::Int64 = 1)

Descriptions, types and units of keywords are given below:

label - unique generator label (positive integer)

bus - bus label to which the generator is connected

active - output active power (per-unit)

reactive - output reactive power (per-unit)

magnitude - voltage magnitude setpoint (per-unit)

minActive - minimum allowed output active power value (per-unit)

maxActive - maximum allowed output active power value (per-unit)

minReactive - minimum allowed output reactive power value (per-unit)

maxReactive - maximum allowed output reactive power value (per-unit)

lowerActive - lower allowed active power output value of PQ capability curve (per-unit)

minReactiveLower - minimum allowed reactive power output value at lowerActive value (per-unit)

maxReactiveLower - maximum allowed reactive power output value at lowerActive value (per-unit)

upperActive - upper allowed active power output value of PQ capability curve (per-unit)

minReactiveUpper - minimum allowed reactive power output value at upperActive value (per-unit)

maxReactiveUpper - maximum allowed reactive power output value at upperActive value (per-unit)

loadFollowing - ramp rate for load following/AG (per-unit/minute)

reserve10minute - ramp rate for 10-minute reserves (per-unit)

reserve30minute - ramp rate for 30-minute reserves (per-unit)

reactiveTimescale - ramp rate for reactive power, two seconds timescale (per-unit/minute)

area - area participation factor

status - operating status, in-service = 1, out-of-service = 0

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)

addGenerator!(system; label = 1, bus = 1, active = 0.5, reactive = 0.1)

Missing docstring.

Missing docstring for statusGenerator!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for outputGenerator!. Check Documenter's build log for details.

Save Model

The function receives the composite type PowerSystem and arguments by keyword to set the path, reference and note of the data set.

JuliaGrid.savePowerSystem — Function

The function saves power system data in the HDF5 file using fields bus, branch, generator and basePower of the composite type PowerSystem.

savePowerSystem(system::PowerSystem; path, reference, note)

The keyword path::String is mandatory in the form "path/name.h5", while keywords reference::String and note::String are optional.

Example

system = powerSystem("case14.m")
savePowerSystem(system; path = "D:/case14.h5")

Build AC and DC Model

Functions receive the composite type PowerSystem and form vectors and matrices related to AC or DC simulations.

Missing docstring.

Missing docstring for acModel!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dcModel!. Check Documenter's build log for details.