Power System Model

The JuliaGrid supports the composite type PowerSystem to preserve power system data, with the following fields:

bus

branch

generator

base

acModel

dcModel.

The fields bus, branch, and generator hold data related to buses, branches, and generators, respectively. The base field holds base values for power and voltages, with three-phase power as the default for the base power and line-to-line voltages for base voltages. The acModel and dcModel fields store vectors and matrices calculated based on the power system's topology and parameters.

The composite type PowerSystem can be created using a function:

powerSystem().

JuliaGrid supports three modes for populating the PowerSystem type: using built-in functions, using HDF5 file format, and using Matpower case files. It is recommended to use the HDF5 format for large-scale systems. To facilitate this, JuliaGrid has the function:

savePowerSystem().

This function allows to save power systems that were either loaded from Matpower case files or created using built-in functions in the HDF5 format.

Once the PowerSystem type is created, you can add buses, branches, and generators using the following functions:

addBus!()

addBranch!()

addGenerator!().

The input data should be entered in per-units or radians, but this default setting can be changed using the @unit macro.

In addition, it is possible to manipulate the parameters of buses, branches, and generators using the following functions:

shuntBus!()

statusBranch!()

parameterBranch!()

statusGenerator!()

outputGenerator!().

Executing these functions will automatically update all fields affected by them. You can also change other parameters of the power system by accessing and modifying the values in the bus, branch, generator, and base fields of the PowerSystem composite type.

To create vectors and matrices based on the power system topology and parameters, you can use the following functions:

acModel!()

dcModel!().

Note that these functions can be executed at any time once all power system buses are defined. Specifically, using the addBranch!() function to add a new branch will automatically update the acModel and dcModel fields. However, adding a new bus using addBus!() requires executing the acModel!() and dcModel!() functions again. In addition, executing functions related to parameter manipulation of buses and branches will also automatically update the acModel and dcModel fields.

Build Model

The function builds the composite type PowerSystem and populates bus, branch, generator and base fields.

JuliaGrid.powerSystem — Function

To use the powerSystem() function, the path to the HDF5 file with the .h5 extension can be passed as arguments. For example:

powerSystem("pathToExternalData/name.h5")

Similarly, the path to the Matpower file with the .m extension can be passed to the same function:

powerSystem("pathToExternalData/name.m")

Alternatively, the PowerSystem composite type can be initialized by calling the function without any arguments:

powerSystem()

This allows the model to be built from scratch and modified as needed.

In general, once the composite type PowerSystem has been created, it is possible to add new buses, branches, or generators, or modify the parameters of existing ones.

Example

system = powerSystem("case14.h5")

The function modifies the units for base power and base voltages and updates the base field.

JuliaGrid.baseUnit! — Function

By default, the units for base power and base voltage are set to volt-amperes (VA) and volts (V). The function can be used to change these units by defining the appropriate prefixes for the aforementioned units:

baseUnit!(system::PowerSystem; power::Symbol, voltage::Symbol)

Note that the prefixes are defined as specified in the International System of Units.

Example

system = powerSystem("case14.h5")
baseUnit!(system; power = :MVA, voltage = :kV)