Power and Current Analysis

In the following section, we have provided a list of functions that can be utilized for post-processing analysis. Once the voltage values are obtained through power flow analysis or optimal power flow analysis, these functions can be used to calculate power or current values. The specific procedures for computing these values depend on the chosen analysis, which are described in separate manuals for further information.

Power Analysis

power

powerBus

powerBranch

powerGenerator

Current Analysis

current

currentBus

currentBranch

Power Analysis

JuliaGrid.power — Method

power(system::PowerSystem, model::ACAnalysis)

The function returns the active and reactive powers associated with buses, branches, and generators in the AC framework.

Abstract type

The abstract type ACAnalysis can have the following subtypes:

ACPowerFlow: computes the powers within the AC power flow,

ACOptimalPowerFlow: computes the powers within the AC optimal power flow.

Returns

The function returns the instance of the Power type, which contains the following fields:

The bus field contains powers related to buses:

injection: active and reactive power injections,

supply: active and reactive power injections from the generators,

shunt: active and reactive powers associated with shunt elements.

The branch field contains powers related to branches:

from: active and reactive power flows at each "from" bus end,

to: active and reactive power flows at each "to" bus end,

shunt: reactive power injections by each branch,

loss: active and reactive power losses.

The generator field contains powers related to generators:

output: output active and reactive powers.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

powers = power(system, model)

JuliaGrid.power! — Method

power!(system::PowerSystem, model::DCAnalysis)

The function calculates the active power values related to buses, branches, and generators within the DC analysis framework. It modifies the power field of the abstract type DCAnalysis.

This function computes the following electrical quantities:

injection: active power injections at each bus,

supply: active power injections from the generators at each bus,

from: active power flows at each "from" bus end of the branch,

to: active power flows at each "to" bus end of the branch,

generator: output active powers of each generator.

Abstract type

The abstract type DCAnalysis can have the following subtypes:

DCPowerFlow: computes the powers within the DC power flow,

DCOptimalPowerFlow: computes the powers within the DC optimal power flow.

Examples

Compute powers after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

power!(system, model)

Compute powers after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system)
optimize!(system, model)

power!(system, model)

JuliaGrid.powerBus — Method

powerBus(system::PowerSystem, model::ACAnalysis, label)

The function returns the active and reactive powers associated associated with a specific bus in the AC framework. The label keyword argument must match an existing bus label.

Abstract type

The abstract type ACAnalysis can have the following subtypes:

ACPowerFlow: computes the powers within the AC power flow,

ACOptimalPowerFlow: computes the powers within the AC optimal power flow.

Returns

The function returns the instance of the PowerBus type, which contains the following fields:

injection: active and reactive power injections at the bus,

supply: active and reactive power injections from the generators at the bus,

shunt: active and reactive powers associated with shunt element at the bus.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

powers = powerBus(system, model; label = 1)

JuliaGrid.powerInjection — Method

powerInjection(system::PowerSystem, model::DCAnalysis; label)

The function returns the active power injection associated with a specific bus in the DC framework. The label keyword argument must match an existing bus label.

Abstract type

The abstract type DCAnalysis can have the following subtypes:

DCPowerFlow: computes the powers within the DC power flow,

DCOptimalPowerFlow: computes the powers within the DC optimal power flow.

Examples

Compute the active power of a specific bus after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

injection = powerInjection(system, model; label = 2)

Compute the active power of a specific bus after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system)
optimize!(system, model)

injection = powerInjection(system, model; label = 2)

JuliaGrid.powerSupply — Method

powerSupply(system::PowerSystem, model::DCAnalysis; label)

The function returns the active power injection from the generators associated with a specific bus in the DC framework. The label keyword argument must match an existing bus label.

Abstract type

The abstract type DCAnalysis can have the following subtypes:

DCPowerFlow: computes the powers within the DC power flow,

DCOptimalPowerFlow: computes the powers within the DC optimal power flow.

Examples

Compute the active power of a specific bus after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

supply = powerSupply(system, model; label = 2)

Compute the active power of a specific bus after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system)
optimize!(system, model)

supply = powerSupply(system, model; label = 2)

JuliaGrid.powerBranch — Method

powerBranch(system::PowerSystem, model::ACAnalysis; label)

The function returns the active and reactive powers associated with a specific branch in the AC framework. The label keyword argument must match an existing branch label.

Abstract type

The abstract type ACAnalysis can have the following subtypes:

ACPowerFlow: computes the powers within the AC power flow,

ACOptimalPowerFlow: computes the powers within the AC optimal power flow.

Returns

The function returns the instance of the PowerBranch type, which contains the following fields:

from: active and reactive power flows at the "from" bus end of the branch,

to: active and reactive power flows at the "to" bus end of the branch,

shunt: reactive power injection by the branch,

loss: active and reactive power losses at the branch.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

powers = powerBranch(system, model; label = 2)

JuliaGrid.powerFrom — Method

powerFrom(system::PowerSystem, model::DCAnalysis; label)

The function returns the active power flow at the "from" bus end associated with a specific branch in the DC framework. The label keyword argument must match an existing branch label.

Abstract type

The abstract type DCAnalysis can have the following subtypes:

DCPowerFlow: computes the powers within the DC power flow,

DCOptimalPowerFlow: computes the powers within the DC optimal power flow.

Examples

Compute the active power of a specific branch after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

from = powerFrom(system, model; label = 2)

Compute the active power of a specific branch after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system)
optimize!(system, model)

from = powerFrom(system, model; label = 2)

JuliaGrid.powerTo — Method

powerTo(system::PowerSystem, model::DCAnalysis; label)

The function returns the active power flow at the "to" bus end associated with a specific branch in the DC framework. The label keyword argument must match an existing branch label.

Abstract type

The abstract type DCAnalysis can have the following subtypes:

DCPowerFlow: computes the powers within the DC power flow,

DCOptimalPowerFlow: computes the powers within the DC optimal power flow.

Examples

Compute the active power of a specific branch after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

to = powerTo(system, model; label = 2)

Compute the active power of a specific branch after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system)
optimize!(system, model)

to = powerTo(system, model; label = 2)

JuliaGrid.powerGenerator — Method

powerGenerator(system::PowerSystem, model::ACAnalysis)

The function returns the active and reactive powers associated with a specific generator in the AC framework. The label keyword argument must match an existing generator label.

Abstract type

The abstract type ACAnalysis can have the following subtypes:

ACPowerFlow: computes the powers within the AC power flow,

ACOptimalPowerFlow: computes the powers within the AC optimal power flow.

Returns

The function returns the instance of the PowerGenerator type, which contains the following field:

output: output active and reactive powers of the generator.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

powers = powerGenerator(system, model; label = 1)

JuliaGrid.powerGenerator — Method

powerGenerator(system::PowerSystem, model::DCAnalysis; label)

This function returns the output active power associated with a specific generator in the DC framework. The label keyword argument must match an existing generator label.

Abstract type

The abstract type DCAnalysis can have the following subtypes:

DCPowerFlow: computes the powers within the DC power flow,

DCOptimalPowerFlow: computes the powers within the DC optimal power flow.

Examples

Compute the active power of a specific generator after obtaining the DC power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

generator = powerGenerator(system, model; label = 1)

Compute the active power of a specific generator after obtaining the DC optimal power flow solution:

system = powerSystem("case14.h5")
dcModel!(system)

model = dcOptimalPowerFlow(system)
optimize!(system, model)

generator = powerGenerator(system, model; label = 1)

Current Analysis

JuliaGrid.current — Method

current(system::PowerSystem, model::ACAnalysis)

The function returns the currents in the polar coordinate system associated with buses and branches in the AC framework.

Abstract type

The abstract type ACAnalysis can have the following subtypes:

ACPowerFlow: computes the currents within the AC power flow,

ACOptimalPowerFlow: computes the currents within the AC optimal power flow.

Returns

The function returns the instance of the Current type, which contains the following fields:

The bus field contains currents related to buses:

injection: current injection magnitudes and angles.

The branch field contains powers related to branches:

from: current flow magnitudes and angles at each "from" bus end,

to: current flow magnitudes and angles at each "to" bus end,

branch: current flow magnitudes and angles through series impedances.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

currents = current(system, model)

JuliaGrid.currentBus — Method

currentBus(system::PowerSystem, model::ACAnalysis; label)

The function returns the currents in the polar coordinate system associated with a specific bus in the AC framework. The label keyword argument must match an existing bus label.

Abstract type

The abstract type ACAnalysis can have the following subtypes:

ACPowerFlow: computes the currents within the AC power flow,

ACOptimalPowerFlow: computes the currents within the AC optimal power flow.

Returns

The function returns the instance of the CurrentBus type, which contains the following field:

injection: current injection magnitude and angle at the bus.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

currents = currentBus(system, model; label = 1)

JuliaGrid.currentBranch — Method

currentBranch(system::PowerSystem, model::ACAnalysis; label)

The function returns the currents in the polar coordinate system associated with a specific branch in the AC framework. The label keyword argument must match an existing branch label.

Abstract type

The abstract type ACAnalysis can have the following subtypes:

ACPowerFlow: computes the currents within the AC power flow,

ACOptimalPowerFlow: computes the currents within the AC optimal power flow.

Returns

The function returns the instance of the CurrentBranch type, which contains the following fields:

from: current flow magnitude and angle at the "from" bus end of the branch,

to: current flow magnitudes and angles at the "to" bus end of the branch,

branch: current flow magnitude and angle through series impedance of the branch.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

currents = currentBranch(system, model; label = 2)