Measurement Model

For further information on this topic, please see the Measurement Model section of the Manual. Below, we have provided a list of functions that can be used to create, save, and manipulate with measurement devices.

To load measurement model API functionalities into the current scope, utilize the following command:

using JuliaGrid

Measurement Data

measurement

ems

saveMeasurement

status!

Voltmeter

addVoltmeter!

updateVoltmeter!

statusVoltmeter!

@voltmeter

Ammeter

addAmmeter!

updateAmmeter!

statusAmmeter!

@ammeter

Wattmeter

addWattmeter!

updateWattmeter!

statusWattmeter!

@wattmeter

Varmeter

addVarmeter!

updateVarmeter!

statusVarmeter!

@varmeter

PMU

addPmu!

updatePmu!

statusPmu!

@pmu

Measurement Data

JuliaGrid.measurement — Function

measurement(system::PowerSystem, file::String)

The function builds a Measurement composite type for the provided PowerSystem, using data stored in the specified file. It populates the voltmeter, ammeter, wattmeter, varmeter, and pmu fields. Once the Measurement type is created, additional measurement devices can be added, or the parameters of existing ones can be modified.

Argument

The function requires an existing PowerSystem instance and a string specifying the path to an HDF5 file with a .h5 extension.

Returns

The Measurement type with the following fields:

voltmeter: Bus voltage magnitude measurements.

ammeter: Branch current magnitude measurements.

wattmeter: Active power injection and active power flow measurements.

varmeter: Reactive power injection and reactive power flow measurements.

pmu: Bus voltage and branch current phasor measurements.

system: The reference to the power system.

Units

JuliaGrid stores all data in per-units and radians format.

Example

system = powerSystem("case14.h5")
monitoring = measurement(system, "monitoring.h5")

measurement(system::PowerSystem)

Alternatively, the Measurement composite type can be initialized without any data for the specified PowerSystem type. This allows the model to be built from scratch and modified as needed.

Example

system = powerSystem("case14.h5")
monitoring = measurement(system)

JuliaGrid.ems — Function

ems(system::String, monitoring::Vararg{String})

This function builds the PowerSystem and Measurement composite types using data from the provided files. It acts as a wrapper that includes both powerSystem and measurement functions.

Arguments

The function requires a string path to the power system data, which can either be stored in an HDF5 file with the .h5 extension or a Matpower file with the .m extension. It also needs a string path to the monitoring data, which should be in an HDF5 file with the .h5 extension.

Additionally, users can ignore the path to the monitoring data to build and populate only the PowerSystem type, while initializing an empty Measurement type.

Returns

The function returns the PowerSystem and Measurement types.

Examples

Build the power system and the measurement model:

system, monitoring = ems("case14.m", "monitoring.h5")

Build the power system and multiple measurement models:

system, monitoring, pseudo = ems("case14.m", "monitoring.h5", "pseudo.h5")

ems()

Alternatively, the PowerSystem and Measurement composite types can be initialized without any data. This provides the flexibility to build and modify the models from the ground up.

Example

system, monitoring = ems()

JuliaGrid.saveMeasurement — Function

saveMeasurement(monitoring::Measurement; path::String, reference::String, note::String)

The function saves the measurement's data in the HDF5 file using the fields voltmeter, ammeter, wattmeter, varmeter, and pmu from the Measurement type.

Keywords

The location and file name of the HDF5 file is specified by the mandatory keyword path in the format of "path/name.h5". Additional information can be provided by the optional keywords reference and note, which can be saved along with the power system data.

View HDF5 File

To view the saved HDF5 file, you can use the HDFView software.

Example

using Ipopt

system, monitoring = ems("case14.m")

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
powerFlow!(analysis; power = true)

addVoltmeter!(monitoring, analysis)
addWattmeter!(monitoring, analysis)

saveMeasurement(monitoring; path = "D:/monitoring.h5")

JuliaGrid.status! — Function

status!(monitoring::Measurement; inservice, outservice, redundancy)

The function generates a set of measurements, assigning measurement devices randomly to either in-service or out-of-service states based on specified keywords.

Keywords

Only one of the following keywords can be used at a time to configure the measurement set:

inservice: Sets the number of in-service devices.

outservice: Sets the number of out-of-service devices.

redundancy: Determines in-service devices based on redundancy.

Updates

The function updates all the status fields within the Measurement type.

Examples

Creating a measurement set with a specific number of in-service devices:

system, monitoring = ems("case14.h5")

analysis = newtonRaphson(system)
powerFlow!(analysis; power = true)

addVoltmeter!(monitoring, analysis)
addWattmeter!(monitoring, analysis)

status!(monitoring; inservice = 30)

Creating a measurement set using redundancy:

system, monitoring = ems("case14.h5")

analysis = newtonRaphson(system)
powerFlow!(analysis; power = true)

addVoltmeter!(monitoring, analysis)
addWattmeter!(monitoring, analysis)
addVarmeter!(monitoring, analysis)

status!(monitoring; redundancy = 2.5)

Voltmeter

JuliaGrid.addVoltmeter! — Method

addVoltmeter!(monitoring::Measurement; label, bus, magnitude, variance, noise, status)

The function adds a voltmeter that measures bus voltage magnitude to the Measurement type. The voltmeter can be added to an already defined bus.

Keywords

The voltmeter is defined with the following keywords:

label: Unique label for the voltmeter.

bus: Label of the bus to which the voltmeter is connected.

magnitude (pu or V): Bus voltage magnitude value.

variance (pu or V): Variance of the bus voltage magnitude measurement.

noise: Specifies how to generate the measurement mean:

noise = true: adds white Gaussian noise with the variance to the magnitude,

noise = false: uses the magnitude value only.

status: Operating status of the voltmeter:

status = 1: in-service,

status = 0: out-of-service.

Note that all voltage values are referenced to line-to-neutral voltages.

Updates

The function updates the voltmeter field of the Measurement type.

Default Settings

Default settings for certain keywords are as follows: variance = 1e-4, noise = false, status = 1, and users can modify these default settings using the @voltmeter macro.

Units

The default units for the magnitude and variance keywords are per-units. However, users can choose to use volts as the units by applying the @voltage macro.

Examples

Adding voltmeters using the default unit system:

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132e3)

addVoltmeter!(monitoring; label = "Voltmeter 1", bus = "Bus 1", magnitude = 1.1)
addVoltmeter!(monitoring; label = "Voltmeter 2", bus = "Bus 1", magnitude = 1.0)

Adding voltmeters using a custom unit system:

@voltage(kV, rad, kV)

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132.0)

addVoltmeter!(monitoring; label = "Voltmeter 1", bus = "Bus 1", magnitude = 145.2)
addVoltmeter!(monitoring; label = "Voltmeter 2", bus = "Bus 1", magnitude = 132.0)

JuliaGrid.addVoltmeter! — Method

addVoltmeter!(monitoring::Measurement, analysis::AC; variance, noise, status)

The function incorporates voltmeters into the Measurement type for every bus within the PowerSystem type from which Measurement was created. These measurements are derived from the exact bus voltage magnitudes defined in the AC type.

Keywords

Voltmeters can be configured using:

variance (pu or V): Measurements variance.

noise: Defines the method for generating the measurement means:

noise = true: adds white Gaussian noise to the voltage magnitudes using the defined variance,

noise = false: uses the exact voltage magnitude values without adding noise.

status: Operating status:

status = 1: in-service,

status = 0: out-of-service.

Updates

The function updates the voltmeter field of the Measurement type.

Default Settings

Default settings for keywords are as follows: variance = 1e-4, noise = false, and status = 1, and users can modify these default settings using the @voltmeter macro.

Units

By default, the unit for variance is per-unit. However, users can choose to use volts as the units by applying the @voltage macro.

Example

@voltmeter(label = "Voltmeter ?")

system, monitoring = ems("case14.h5")

analysis = newtonRaphson(system)
powerFlow!(analysis)

addVoltmeter!(monitoring, analysis; variance = 1e-3, noise = true)

JuliaGrid.updateVoltmeter! — Function

updateVoltmeter!(monitoring::Measurement; kwargs...)

The function allows for the alteration of parameters for a voltmeter.

Keywords

To update a specific voltmeter, provide the necessary kwargs input arguments in accordance with the keywords specified in the addVoltmeter! function, along with their respective values. Ensure that the label keyword matches the label of the existing voltmeter. If any keywords are omitted, their corresponding values will remain unchanged.

Updates

The function updates the voltmeter field within the Measurement type.

Units

Units for input parameters can be changed using the same method as described for the addVoltmeter! function.

Example

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132e3)

addVoltmeter!(monitoring; label = "Voltmeter 1", bus = "Bus 1", magnitude = 1.1)
updateVoltmeter!(monitoring; label = "Voltmeter 1", magnitude = 0.9)

updateVoltmeter!(analysis::Analysis; kwargs...)

The function extends the updateVoltmeter! function. By passing the Analysis type, the function first updates the specific voltmeter within the Measurement type using the provided kwargs, and then updates the Analysis type with all parameters associated with that voltmeter.

A key feature of this function is that any prior modifications made to the specified voltmeter are preserved and applied to the Analysis type when the function is executed, ensuring consistency throughout the update process.

Example

system, monitoring = ems("case14.h5", "monitoring.h5")

analysis = gaussNewton(monitoring)
stateEstimation!(analysis)

updateVoltmeter!(analysis; label = 2, magnitude = 0.9)

JuliaGrid.statusVoltmeter! — Function

statusVoltmeter!(monitoring::Measurement; inservice, outservice, redundancy)

The function generates a set of voltmeters, assigning voltmeters randomly to either in-service or out-of-service states based on specified keywords.

Keywords

Only one of the following keywords can be used at a time to configure the measurement set:

inservice: Sets the number of in-service voltmeters.

outservice: Sets the number of out-of-service voltmeters.

redundancy: Determines in-service voltmeters based on redundancy.

Updates

The function updates the status field within the Voltmeter type.

Example

system, monitoring = ems("case14.h5")

analysis = newtonRaphson(system)
powerFlow!(analysis)

addVoltmeter!(monitoring, analysis)

statusVoltmeter!(monitoring; inservice = 10)

JuliaGrid.@voltmeter — Macro

@voltmeter(label, variance, noise, status)

The macro generates a template for a voltmeter.

Keywords

To establish the voltmeter template, users can specify default values for the variance, noise, and status keywords, along with pattern for labels using the label keyword.

Units

By default, the unit for variance is per-unit. However, users can choose to use volts as the units by applying the @voltage macro.

Examples

Adding a voltmeter using the default unit system:

@voltmeter(label = "Voltmeter ?", variance = 1e-5)

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132e3)

addVoltmeter!(monitoring; bus = "Bus 1", magnitude = 1.1)

Adding a voltmeter using a custom unit system:

@voltage(kV, rad, kV)
@voltmeter(label = "Voltmeter ?", variance = 0.00132)

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132.0)

addVoltmeter!(monitoring; bus = "Bus 1", magnitude = 145.2)

Ammeter

JuliaGrid.addAmmeter! — Method

addAmmeter!(monitoring::Measurement;
    label, from, to, magnitude, variance, noise, square, status)

The function adds an ammeter that measures branch current magnitude to the Measurement type. The ammeter can be added to an already defined branch.

Keywords

The ammeter is defined with the following keywords:

label: Unique label for the ammeter.

from: Label of the branch if the ammeter is located at the from-bus end.

to: Label of the branch if the ammeter is located at the to-bus end.

magnitude (pu or A): Branch current magnitude value.

variance (pu or A): Variance of the branch current magnitude measurement.

noise: Specifies how to generate the measurement mean:

noise = true: adds white Gaussian noise with the variance to the magnitude,

noise = false: uses the magnitude value only.

square: Specifies how the measurement is included in the state estimation model:

square = true: included in squared form,

square = false: included in its original form.

status: Operating status of the ammeter:

status = 1: in-service,

status = 0: out-of-service.

Updates

The function updates the ammeter field of the Measurement type.

Default Settings

Default settings for certain keywords are as follows: variance = 1e-4, noise = false, square = false, and status = 1, which apply to ammeters located at both the from-bus and to-bus ends. Users can fine-tune these settings by explicitly specifying the variance and status for ammeters positioned on either the from-bus or to-bus ends of branches using the @ammeter macro.

Units

The default units for the magnitude and variance keywords are per-units. However, users can choose to use amperes as the units by applying the @current macro.

Examples

Adding ammeters using the default unit system:

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132e3)
addBus!(system; label = "Bus 2", base = 132e3)
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addAmmeter!(monitoring; label = "Ammeter 1", from = "Branch 1", magnitude = 1.1)
addAmmeter!(monitoring; label = "Ammeter 2", to = "Branch 1", magnitude = 1.0)

Adding ammeters using a custom unit system:

@current(A, rad)

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132e3)
addBus!(system; label = "Bus 2", base = 132e3)
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addAmmeter!(monitoring; label = "Ammeter 1", from = "Branch 1", magnitude = 481.125)
addAmmeter!(monitoring; label = "Ammeter 2", to = "Branch 1", magnitude = 437.386)

JuliaGrid.addAmmeter! — Method

addAmmeter!(monitoring::Measurement, analysis::AC;
    varianceFrom, statusFrom, varianceTo, statusTo, noise, square)

The function incorporates ammeters into the Measurement type for every branch within the PowerSystem type from which Measurement was created. These measurements are derived from the exact branch current magnitudes defined in the AC type.

Keywords

Ammeters at the from-bus ends of the branches can be configured using:

varianceFrom (pu or A): Measurement variance.

statusFrom: Operating status:

statusFrom = 1: in-service,

statusFrom = 0: out-of-service,

statusFrom = -1: not included in the Measurement type.

Ammeters at the to-bus ends of the branches can be configured using:

varianceTo (pu or A): Measurement variance.

statusTo: Operating status:

statusTo = 1: in-service,

statusTo = 0: out-of-service,

statusTo = -1: not included in the Measurement type.

Settings for generating measurements include:

noise: Defines the method for generating the measurement means:

noise = true: adds white Gaussian noise to the current magnitudes using defined variances,

noise = false: uses the exact current magnitude values without adding noise.

Settings for including measurements:

square: Specifies how measurements are included in the state estimation model:

square = true: included in squared form,

square = false: included in its original form.

Updates

The function updates the ammeter field of the Measurement type.

Default Settings

Default settings for keywords are as follows: varianceFrom = 1e-4, statusFrom = 1, varianceTo = 1e-4, statusTo = 1, noise = false, and square = false. Users can change these default settings using the @ammeter macro.

Units

The default units for the varianceFrom and varianceTo keywords are per-units. However, users can choose to use amperes as the units by applying the @current macro.

Example

@ammeter(label = "Ammeter ?")

system, monitoring = ems("case14.h5")

analysis = newtonRaphson(system)
powerFlow!(analysis; current = true)

addAmmeter!(monitoring, analysis; varianceFrom = 1e-3, statusTo = 0, square = true)

JuliaGrid.updateAmmeter! — Function

updateAmmeter!(monitoring::Measurement; kwargs...)

The function allows for the alteration of parameters for an ammeter.

Keywords

To update a specific ammeter, provide the necessary kwargs input arguments in accordance with the keywords specified in the addAmmeter! function, along with their respective values. Ensure that the label keyword matches the label of the existing ammeter you want to modify. If any keywords are omitted, their corresponding values will remain unchanged.

Updates

The function updates the ammeter field within the Measurement type.

Units

Units for input parameters can be changed using the same method as described for the addAmmeter! function.

Example

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132e3)
addBus!(system; label = "Bus 2", base = 132e3)
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addAmmeter!(monitoring; label = "Ammeter 1", from = "Branch 1", magnitude = 1.1)
updateAmmeter!(monitoring; label = "Ammeter 1", magnitude = 1.2, variance = 1e-3)

updateAmmeter!(analysis::Analysis; kwargs...)

The function extends the updateAmmeter! function. By passing the Analysis type, the function first updates the specific ammeter within the Measurement type using the provided kwargs, and then updates the Analysis type with all parameters associated with that ammeter.

A key feature of this function is that any prior modifications made to the specified ammeter are preserved and applied to the Analysis type when the function is executed, ensuring consistency throughout the update process.

Example

system, monitoring = ems("case14.h5", "monitoring.h5")

analysis = gaussNewton(monitoring)
stateEstimation!(analysis)

updateAmmeter!(analysis; label = "From 1", magnitude = 0.9, variance = 1e-5)

JuliaGrid.statusAmmeter! — Function

statusAmmeter!(monitoring::Measurement;
    inservice, inserviceFrom, inserviceTo,
    outservice, outserviceFrom, outserviceTo,
    redundancy, redundancyFrom, redundancyTo)

The function generates a set of ammeters, assigning ammeters randomly to either in-service or out-of-service states based on specified keywords.

Keywords

Users can use one main keyword or two fine-tuning keywords to specify distinct locations per function call:

inservice: Sets the number of in-service ammeters or allows fine-tuning:

inserviceFrom: sets only ammeters loacted at the from-bus end,

inserviceTo: sets only ammeters loacted at the to-bus end.

outservice: Sets the number of out-of-service ammeters or allows fine-tuning:

outserviceFrom: sets only ammeters loacted at the from-bus end,

outserviceTo: sets only ammeters loacted at the to-bus end.

redundancy: Determines in-service ammeters based on redundancy or allows fine-tuning:

redundancyFrom: determines only ammeters loacted at the from-bus end,

redundancyTo: determines only ammeters loacted at the to-bus end.

Updates

The function updates the status field within the Ammeter type.

Example

system, monitoring = ems("case14.h5")

analysis = newtonRaphson(system)
powerFlow!(analysis; current = true)

addAmmeter!(monitoring, analysis)

statusAmmeter!(monitoring; inserviceFrom = 5, inserviceTo = 10)

JuliaGrid.@ammeter — Macro

@ammeter(label, varianceFrom, statusFrom, varianceTo, statusTo, noise, square)

The macro generates a template for an ammeter.

Keywords

To establish the ammeter template, users can set default variance and status values for ammeters at both the from-bus and to-bus ends of branches, using varianceFrom and statusFrom for the former and varianceTo and statusTo for the latter. Users can also configure label patterns with the label keyword, as well as specify the noise type. Finally, the square keyword enables including the measurement in squared form for the state estimation model.

Units

The default units for the varianceFrom and varianceTo keywords are per-units. However, users can choose to use amperes as the units by applying the @current macro.

Examples

Adding an ammeter using the default unit system:

@ammeter(label = "Ammeter ?", varianceTo = 1e-3, statusTo = 0, square = true)

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132e3)
addBus!(system; label = "Bus 2", base = 132e3)
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addAmmeter!(monitoring; to = "Branch 1", magnitude = 1.1)

Adding an ammeter using a custom unit system:

@current(A, rad)
@ammeter(label = "Ammeter ?", varianceTo = 0.004374, statusTo = 0, square = true)

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132e3)
addBus!(system; label = "Bus 2", base = 132e3)
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addAmmeter!(monitoring; label = "Ammeter 1", to = "Branch 1", magnitude = 481.125)

Wattmeter

JuliaGrid.addWattmeter! — Method

addWattmeter!(monitoring::Measurement;
    label, bus, from, to, active, variance, noise, status)

The function adds a wattmeter that measures active power injection or active power flow to the Measurement type. The wattmeter can be added to an already defined bus or branch.

Keywords

The wattmeter is defined with the following keywords:

label: Unique label for the wattmeter.

bus: Label of the bus if the wattmeter is located at the bus.

from: Label of the branch if the wattmeter is located at the from-bus end.

to: Label of the branch if the wattmeter is located at the to-bus end.

active (pu or W): Active power value.

variance (pu or W): Variance of the active power measurement.

noise: Specifies how to generate the measurement mean:

noise = true: adds white Gaussian noise with the variance to the active,

noise = false: uses the active value only.

status: Operating status of the wattmeter:

status = 1: in-service,

status = 0: out-of-service.

Note that when powers are given in SI units, they correspond to three-phase power.

Updates

The function updates the wattmeter field of the Measurement type.

Default Settings

Default settings for certain keywords are as follows: variance = 1e-4, noise = false, and status = 1, which apply to wattmeters located at the bus, as well as at both the from-bus and to-bus ends. Users can fine-tune these settings by explicitly specifying the variance and status for wattmeters positioned at the buses, from-bus ends, or to-bus ends of branches using the @wattmeter macro.

Units

The default units for the active and variance keywords are per-units. However, users can choose to use watts as the units by applying the @power macro.

Examples

Adding wattmeters using the default unit system:

system, monitoring = ems()

addBus!(system; label = "Bus 1")
addBus!(system; label = "Bus 2")
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addWattmeter!(monitoring; label = "Wattmeter 1", bus = "Bus 2", active = 0.4)
addWattmeter!(monitoring; label = "Wattmeter 2", from = "Branch 1", active = 0.1)

Adding wattmeters using a custom unit system:

@power(MW, pu)

system, monitoring = ems()

addBus!(system; label = "Bus 1")
addBus!(system; label = "Bus 2")
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addWattmeter!(monitoring; label = "Wattmeter 1", bus = "Bus 2", active = 40.0)
addWattmeter!(monitoring; label = "Wattmeter 2", from = "Branch 1", active = 10.0)

JuliaGrid.addWattmeter! — Method

addWattmeter!(monitoring::Measurement, analysis::AC;
    varianceBus, statusBus, varianceFrom, statusFrom, varianceTo, statusTo, noise)

The function incorporates wattmeters into the Measurement type for every branch within the PowerSystem type from which Measurement was created. These measurements are derived from the exact active power injections at buses and active power flows in branches defined in the AC type.

Keywords

Wattmeters at the buses can be configured using:

varianceBus (pu or W): Measurement variance.

statusBus: Operating status:

statusBus = 1: in-service,

statusBus = 0: out-of-service,

statusBus = -1: not included in the Measurement type.

Wattmeters at the from-bus ends of the branches can be configured using:

varianceFrom (pu or W): Measurement variance.

statusFrom: Operating status:

statusFrom = 1: in-service,

statusFrom = 0: out-of-service,

statusFrom = -1: not included in the Measurement type.

Wattmeters at the to-bus ends of the branches can be configured using:

varianceTo (pu or W): Measurement variance.

statusTo: Operating status:

statusTo = 1: in-service,

statusTo = 0: out-of-service,

statusTo = -1: not included in the Measurement type.

Settings for generating measurements include:

noise: Defines the method for generating the measurement means:

noise = true: adds white Gaussian noise to the active power values using defined variances,

noise = false: uses the exact active power values without adding noise.

Updates

The function updates the wattmeter field of the Measurement composite type.

Default Settings

Default settings for keywords are as follows: varianceBus = 1e-4, statusBus = 1, varianceFrom = 1e-4, statusFrom = 1, varianceTo = 1e-4, statusTo = 1, and noise = false. Users can change these default settings using the @wattmeter macro.

Units

The default units for the varianceBus, varianceFrom, and varianceTo keywords are per-units. However, users can choose to use watts as the units by applying the @power macro.

Example

@wattmeter(label = "Wattmeter ?")

system, monitoring = ems("case14.h5")

analysis = newtonRaphson(system)
powerFlow!(system, analysis; power = true)

addWattmeter!(monitoring, analysis; varianceBus = 1e-3, statusFrom = 0)

JuliaGrid.updateWattmeter! — Function

updateWattmeter!(monitoring::Measurement; kwargs...)

The function allows for the alteration of parameters for a wattmeter.

Keywords

To update a specific wattmeter, provide the necessary kwargs input arguments in accordance with the keywords specified in the addWattmeter! function, along with their respective values. Ensure that the label keyword matches the label of the existing wattmeter you want to modify. If any keywords are omitted, their corresponding values will remain unchanged.

Updates

The function updates the wattmeter field within the Measurement composite type.

Units

Units for input parameters can be changed using the same method as described for the addWattmeter! function.

Example

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132e3)
addBus!(system; label = "Bus 2", base = 132e3)
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addWattmeter!(monitoring; label = "Wattmeter 1", from = "Branch 1", active = 1.1)
updateWattmeter!(monitoring; label = "Wattmeter 1", active = 1.2, variance = 1e-4)

updateWattmeter!(analysis::Analysis; kwargs...)

The function extends the updateWattmeter! function. By passing the Analysis type, the function first updates the specific wattmeter within the Measurement type using the provided kwargs, and then updates the Analysis type with all parameters associated with that wattmeter.

A key feature of this function is that any prior modifications made to the specified wattmeter are preserved and applied to the Analysis type when the function is executed, ensuring consistency throughout the update process.

Example

system, monitoring = ems("case14.h5", "monitoring.h5")

analysis = gaussNewton(monitoring)
stateEstimation!(analysis)

updateWattmeter!(analysis; label = 4, active = 0.5, variance = 1e-4)

JuliaGrid.statusWattmeter! — Function

statusWattmeter!(monitoring::Measurement;
    inservice, inserviceBus, inserviceFrom, inserviceTo,
    outservice, outserviceBus outserviceFrom, outserviceTo,
    redundancy, redundancyBus, redundancyFrom, redundancyTo)

The function generates a set of wattmeters, assigning wattmeters randomly to either in-service or out-of-service states based on specified keywords.

Keywords

Users can use one main keyword or three fine-tuning keywords to specify distinct locations per function call:

inservice: Sets the number of in-service wattmeters or allows fine-tuning:

inserviceBus: sets only wattmeters loacted at the bus,

inserviceFrom: sets only wattmeters loacted at the from-bus end,

inserviceTo: sets only wattmeters loacted at the to-bus end.

outservice: Sets the number of out-of-service wattmeters or allows fine-tuning:

outserviceBus: sets only wattmeters loacted at the bus,

outserviceFrom: sets only wattmeters loacted at the from-bus end,

outserviceTo: sets only wattmeters loacted at the to-bus end.

redundancy: Determines in-service wattmeters based on redundancy or allows fine-tuning:

redundancyBus: determines only wattmeters loacted at the bus,

redundancyFrom: determines only wattmeters loacted at the from-bus end,

redundancyTo: determines only wattmeters loacted at the to-bus end.

Updates

The function updates the status field within the Wattmeter type.

Example

system, monitoring = ems("case14.h5")

analysis = newtonRaphson(system)
powerFlow!(analysis; power = true)

addWattmeter!(monitoring, analysis)

statusWattmeter!(monitoring; outserviceBus = 14, inserviceFrom = 10, outserviceTo = 2)

JuliaGrid.@wattmeter — Macro

@wattmeter(label, varianceBus, statusBus, varianceFrom, statusFrom,
    varianceTo, statusTo, noise)

The macro generates a template for a wattmeter.

Keywords

To establish the wattmeter template, users can set default variance and status values for wattmeters at buses using varianceBus and statusBus, and at both the from-bus and to-bus ends of branches using varianceFrom and statusFrom for the former and varianceTo and statusTo for the latter. Users can also configure label patterns with the label keyword, as well as specify the noise type.

Units

The default units for the varianceBus, varianceFrom, and varianceTo keywords are per-units. However, users can choose to use watts as the units by applying the @power macro.

Examples

Adding wattmeters using the default unit system:

system, monitoring = ems()

addBus!(system; label = "Bus 1")
addBus!(system; label = "Bus 2")
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

@wattmeter(label = "Wattmeter ?", varianceBus = 1e-3, varianceFrom = 1e-4)
addWattmeter!(monitoring; bus = "Bus 2", active = 0.4)
addWattmeter!(monitoring; from = "Branch 1", active = 0.1)

Adding wattmeters using a custom unit system:

@power(MW, pu)
system, monitoring = ems()

addBus!(system; label = "Bus 1")
addBus!(system; label = "Bus 2")
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

@wattmeter(label = "Wattmeter ?", varianceBus = 1e-1, varianceFrom = 1e-4)
addWattmeter!(monitoring; bus = "Bus 2", active = 40.0)
addWattmeter!(monitoring; from = "Branch 1", active = 10.0)

Varmeter

JuliaGrid.addVarmeter! — Method

addVarmeter!(monitoring::Measurement;
    label, bus, from, to, reactive, variance, noise, status)

The function adds a varmeter that measures reactive power injection or reactive power flow to the Measurement type. The varmeter can be added to an already defined bus or branch.

Keywords

The varmeter is defined with the following keywords:

label: Unique label for the varmeter.

bus: Label of the bus if the varmeter is located at the bus.

from: Label of the branch if the varmeter is located at the from-bus end.

to: Label of the branch if the varmeter is located at the to-bus end.

reactive (pu or VAr): Reactive power value.

variance (pu or VAr): Variance of the reactive power measurement.

noise: Specifies how to generate the measurement mean:

noise = true: adds white Gaussian noise with the variance to the reactive,

noise = false: uses the reactive value only.

status: Operating status of the varmeter:

status = 1: in-service,

status = 0: out-of-service.

Note that when powers are given in SI units, they correspond to three-phase power.

Updates

The function updates the varmeter field of the Measurement type.

Default Settings

Default settings for certain keywords are as follows: variance = 1e-4, noise = false, and status = 1, which apply to varmeters located at the bus, as well as at both the from-bus and to-bus ends. Users can fine-tune these settings by explicitly specifying the variance and status for varmeters positioned at the buses, from-bus ends, or to-bus ends of branches using the @varmeter macro.

Units

The default units for the reactive and variance keywords are per-units. However, users can choose to use volt-amperes reactive as the units by applying the @power macro.

Examples

Adding varmeters using the default unit system:

system, monitoring = ems()

addBus!(system; label = "Bus 1")
addBus!(system; label = "Bus 2")
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addVarmeter!(monitoring; label = "Varmeter 1", bus = "Bus 2", reactive = 0.4)
addVarmeter!(monitoring; label = "Varmeter 2", from = "Branch 1", reactive = 0.1)

Adding varmeters using a custom unit system:

@power(pu, MVAr)

system, monitoring = ems()

addBus!(system; label = "Bus 1")
addBus!(system; label = "Bus 2")
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addVarmeter!(monitoring; label = "Varmeter 1", bus = "Bus 2", reactive = 40.0)
addVarmeter!(monitoring; label = "Varmeter 2", from = "Branch 1", reactive = 10.0)

JuliaGrid.addVarmeter! — Method

addVarmeter!(monitoring::Measurement, analysis::AC;
    varianceBus, statusBus, varianceFrom, statusFrom, varianceTo, statusTo, noise)

The function incorporates varmeters into the Measurement type for every branch within the PowerSystem type from which Measurement was created. These measurements are derived from the exact reactive power injections at buses and reactive power flows in branches defined in the AC type.

Keywords

Varmeters at the buses can be configured using:

varianceBus (pu or VAr): Measurement variance.

statusBus: Operating status:

statusBus = 1: in-service,

statusBus = 0: out-of-service,

statusBus = -1: not included in the Measurement type.

Varmeters at the from-bus ends of the branches can be configured using:

varianceFrom (pu or VAr): Measurement variance.

statusFrom: Operating status:

statusFrom = 1: in-service,

statusFrom = 0: out-of-service,

statusFrom = -1: not included in the Measurement type.

Varmeters at the to-bus ends of the branches can be configured using:

varianceTo (pu or VAr): Measurement variance.

statusTo: Operating status:

statusTo = 1: in-service,

statusTo = 0: out-of-service,

statusTo = -1: not included in the Measurement type.

Settings for generating measurements include:

noise: Defines the method for generating the measurement means:

noise = true: adds white Gaussian noise to the reactive power values using defined variances,

noise = false: uses the exact reactive power values without adding noise.

Updates

The function updates the varmeter field of the Measurement composite type.

Default Settings

Default settings for keywords are as follows: varianceBus = 1e-4, statusBus = 1, varianceFrom = 1e-4, statusFrom = 1, varianceTo = 1e-4, statusTo = 1, and noise = false. Users can change these default settings using the @varmeter macro.

Units

The default units for the varianceBus, varianceFrom, and varianceTo keywords are per-units. However, users can choose to use volt-amperes reactive as the units by applying the @power macro.

Example

@varmeter(label = "Varmeter ?")

system, monitoring = ems("case14.h5")

analysis = newtonRaphson(system)
powerFlow!(system, analysis; power = true)

addVarmeter!(monitoring, analysis; varianceFrom = 1e-3, statusBus = 0)

JuliaGrid.updateVarmeter! — Function

updateVarmeter!(monitoring::Measurement; kwargs...)

The function allows for the alteration of parameters for a varmeter.

Keywords

To update a specific varmeter, provide the necessary kwargs input arguments in accordance with the keywords specified in the addVarmeter! function, along with their respective values. Ensure that the label keyword matches the label of the existing varmeter you want to modify. If any keywords are omitted, their corresponding values will remain unchanged.

Updates

The function updates the varmeter field within the Measurement type.

Units

Units for input parameters can be changed using the same method as described for the addVarmeter! function.

Example

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132e3)
addBus!(system; label = "Bus 2", base = 132e3)
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addVarmeter!(monitoring; label = "Varmeter 1", from = "Branch 1", reactive = 1.1)
updateVarmeter!(monitoring; label = "Varmeter 1", reactive = 1.2, variance = 1e-4)

updateVarmeter!(analysis::Analysis; kwargs...)

The function extends the updateVarmeter! function. By passing the Analysis type, the function first updates the specific varmeter within the Measurement type using the provided kwargs, and then updates the Analysis type with all parameters associated with that varmeter.

A key feature of this function is that any prior modifications made to the specified varmeter are preserved and applied to the Analysis type when the function is executed, ensuring consistency throughout the update process.

Example

system, monitoring = ems("case14.h5", "monitoring.h5")

analysis = gaussNewton(monitoring)
stateEstimation!(analysis)

updateVarmeter!(analysis; label = 4, reactive = 0.3, variance = 1e-3)

JuliaGrid.statusVarmeter! — Function

statusVarmeter!(monitoring::Measurement;
    inservice, inserviceBus, inserviceFrom, inserviceTo,
    outservice, outserviceBus outserviceFrom, outserviceTo,
    redundancy, redundancyBus, redundancyFrom, redundancyTo)

The function generates a set of varmeters, assigning varmeters randomly to either in-service or out-of-service states based on specified keywords.

Keywords

Users can use one main keyword or three fine-tuning keywords to specify distinct locations per function call:

inservice: Sets the number of in-service varmeters or allows fine-tuning:

inserviceBus: sets only varmeters loacted at the bus,

inserviceFrom: sets only varmeters loacted at the from-bus end,

inserviceTo: sets only varmeters loacted at the to-bus end.

outservice: Sets the number of out-of-service varmeters or allows fine-tuning:

outserviceBus: sets only varmeters loacted at the bus,

outserviceFrom: sets only varmeters loacted at the from-bus end,

outserviceTo: sets only varmeters loacted at the to-bus end.

redundancy: Determines in-service varmeters based on redundancy or allows fine-tuning:

redundancyBus: determines only varmeters loacted at the bus,

redundancyFrom: determines only varmeters loacted at the from-bus end,

redundancyTo: determines only varmeters loacted at the to-bus end.

Updates

The function updates the status field within the Varmeter type.

Example

system, monitoring = ems("case14.h5")

analysis = newtonRaphson(system)
powerFlow!(analysis; power = true)

addVarmeter!(monitoring, analysis)

statusVarmeter!(monitoring; inserviceFrom = 20)

JuliaGrid.@varmeter — Macro

@varmeter(label, varinaceBus, statusBus, varianceFrom, statusFrom,
    varianceTo, statusTo, noise)

The macro generates a template for a varmeter.

Keywords

To establish the varmeter template, users can set default variance and status values for varmeters at buses using varianceBus and statusBus, and at both the from-bus and to-bus ends of branches using varianceFrom and statusFrom for the former and varianceTo and statusTo for the latter. Users can also configure label patterns with the label keyword, as well as specify the noise type.

Units

The default units for the varianceBus, varianceFrom, and varianceTo keywords are per-units. However, users can choose to usevolt-amperes reactive as the units by applying the @power macro.

Examples

Adding varmeters using the default unit system:

system, monitoring = ems()

addBus!(system; label = "Bus 1")
addBus!(system; label = "Bus 2")
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

@varmeter(label = "Varmeter ?", varianceBus = 1e-3, varianceFrom = 1e-4)
addVarmeter!(monitoring; bus = "Bus 2", reactive = 0.4)
addVarmeter!(monitoring; from = "Branch 1", reactive = 0.1)

Adding varmeters using a custom unit system:

@power(pu, MVAr)
system, monitoring = ems()

addBus!(system; label = "Bus 1")
addBus!(system; label = "Bus 2")
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

@varmeter(label = "Varmeter ?", varianceBus = 1e-1, varianceFrom = 1e-4)
addVarmeter!(monitoring; bus = "Bus 2", reactive = 40.0)
addVarmeter!(monitoring; from = "Branch 1", reactive = 10.0)

PMU

JuliaGrid.addPmu! — Method

addPmu!(monitoring::Measurement;
    label, bus, from, to, magnitude, varianceMagnitude, angle, varianceAngle,
    noise, correlated, polar, square, status)

The function adds a PMU to the Measurement type. The PMU can be added to an already defined bus or branch. When defining the PMU, it is essential to provide the bus voltage magnitude and angle if the PMU is located at a bus or the branch current magnitude and angle if the PMU is located at a branch.

Keywords

The PMU is defined with the following keywords:

label: Unique label for the PMU.

bus: Label of the bus if the PMU is located at the bus.

from: Label of the branch if the PMU is located at the from-bus end.

to: Label of the branch if the PMU is located at the to-bus end.

magnitude (pu or V, A): Bus voltage or branch current magnitude value.

varianceMagnitude (pu or V, A): Magnitude measurement variance.

angle (rad or deg): Bus voltage or branch current angle value.

varianceAngle (rad or deg): Angle measurement variance.

noise: Specifies how to generate the measurement means:

noise = true: adds white Gaussian noises with variances to the magnitude and angle,

noise = false: uses the magnitude and angle values only.

correlated: Specifies error correlation for PMUs for algorithms utilizing rectangular coordinates:

correlated = true: considers correlated errors,

correlated = false: disregards correlations between errors.

polar: Chooses the coordinate system for including phasor measurements in AC state estimation:

polar = true: adopts the polar coordinate system,

polar = false: adopts the rectangular coordinate system.

square: Specifies how the current magnitude is included in the model when using the polar system:

square = true: included in squared form,

square = false: included in its original form.

status: Operating status of the phasor measurement:

status = 1: in-service,

status = 0: out-of-service.

Note that all voltage values are referenced to line-to-neutral voltages.

Updates

The function updates the pmu field of the Measurement type.

Default Settings

Default settings for certain keywords are as follows: varianceMagnitude = 1e-8, varianceAngle = 1e-8, noise = false, correlated = false, polar = false, square = false, and status = 1 which apply to PMUs located at the bus, as well as at both the from-bus and to-bus ends. Users can fine-tune these settings by explicitly specifying the variance and status for PMUs positioned at the buses, from-bus ends, or to-bus ends of branches using the @pmu macro.

Units

The default units for the magnitude, varianceMagnitude, and angle, varianceAngle keywords are per-units and radians. However, users have the option to switch to volts and degrees when the PMU is located at a bus using the @voltage macro, or amperes and degrees when the PMU is located at a branch through the use of the @current macro.

Examples

Adding PMUs using the default unit system:

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 131.8e3)
addBus!(system; label = "Bus 2", base = 131.8e3)
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addPmu!(monitoring; label = "PMU 1", bus = "Bus 1", magnitude = 1.1, angle = -0.1)
addPmu!(monitoring; label = "PMU 2", from = "Branch 1", magnitude = 1.1, angle = 0.1)

Adding PMUs using a custom unit system:

@voltage(kV, deg, kV)
@current(A, deg)

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 131.8)
addBus!(system; label = "Bus 2", base = 131.8)
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addPmu!(monitoring; label = "PMU 1", bus = "Bus 1", magnitude = 145, angle = -5.7)
addPmu!(monitoring; label = "PMU 2", from = "Branch 1", magnitude = 481, angle = 5.7)

JuliaGrid.addPmu! — Method

addPmu!(monitoring::Measurement, analysis::AC;
    varianceMagnitudeBus, varianceAngleBus, statusBus,
    varianceMagnitudeFrom, varianceAngleFrom, statusFrom,
    varianceMagnitudeTo, varianceAngleTo, statusTo,
    noise, correlated, polar, square)

The function incorporates PMUs into the Measurement type for every branch within the PowerSystem type from which Measurement was created. These measurements are derived from the exact bus voltage magnitudes and angles, as well as branch current magnitudes and angles defined in the AC type.

Keywords

PMUs at the buses can be configured using:

varianceMagnitudeBus (pu or V): Variance of bus voltage magnitude measurements.

varianceAngleBus (rad or deg): Variance of bus voltage angle measurements.

statuseBus: Operating status:

statusBus = 1: in-service,

statusBus = 0: out-of-service,

statusBus = -1: not included in the Measurement type.

PMUs at the from-bus ends of the branches can be configured using:

varianceMagnitudeFrom (pu or A): Variance of current magnitude measurements.

varianceAngleFrom (rad or deg): Variance of current angle measurements.

statusFrom: Operating status:

statusFrom = 1: in-service,

statusFrom = 0: out-of-service,

statusFrom = -1: not included in the Measurement type.

PMUs at the to-bus ends of the branches can be configured using:

varianceMagnitudeTo (pu or A): Variance of current magnitude measurements.

varianceAngleTo (rad or deg): Variance of current angle measurements.

statusTo: Operating status:

statusTo = 1: in-service,

statusTo = 0: out-of-service,

statusTo = -1: not included in the Measurement type.

Settings for generating measurements include:

noise: Defines the method for generating the measurement means:

noise = true: adds white Gaussian noise to the phasor values using defined variances,

noise = false: uses the exact phasor values without adding noise.

Settings for handling phasor measurements include:

correlated: Specifies error correlation for PMUs for algorithms utilizing rectangular coordinates:

correlated = true: considers correlated errors,

correlated = false: disregards correlations between errors.

polar: Chooses the coordinate system for including phasor measurements in AC state estimation:

polar = true: adopts the polar coordinate system,

polar = false: adopts the rectangular coordinate system.

square: Specifies how current magnitudes are included in the model when using the polar system:

square = true: included in squared form,

square = false: included in its original form.

Updates

The function updates the pmu field of the Measurement type.

Default Settings

Default settings for variance keywords are established at 1e-8, with all statuses set to 1, polar = false, correlated = false, noise = false, and square = false. Users can change these default settings using the @pmu macro.

Units

The default units for the variance keywords are in per-units and radians. However, users have the option to switch to volts and degrees when the PMU is located at a bus using the @voltage macro, or amperes and degrees when the PMU is located at a branch through the use of the @current macro.

Example

@pmu(label = "PMU ?")

system, monitoring = ems("case14.h5")

analysis = newtonRaphson(system)
powerFlow!(analysis; current = true)

addPmu!(monitoring, analysis; varianceMagnitudeBus = 1e-3)

JuliaGrid.updatePmu! — Function

updatePmu!(monitoring::Measurement; kwargs...)

The function allows for the alteration of parameters for a PMU.

Keywords

To update a specific PMU, provide the necessary kwargs input arguments in accordance with the keywords specified in the addPmu! function, along with their respective values. Ensure that the label keyword matches the label of the existing PMU you want to modify. If any keywords are omitted, their corresponding values will remain unchanged.

Updates

The function updates the pmu field within the Measurement type.

Units

Units for input parameters can be changed using the same method as described for the addPmu! function.

Example

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132e3)

addPmu!(monitoring; label = "PMU 1", bus = "Bus 1", magnitude = 1.1, angle = -0.1)
updatePmu!(monitoring; label = "PMU 1", magnitude = 1.05)

updatePmu!(analysis::Analysis; kwargs...)

The function extends the updatePmu! function. By passing the Analysis type, the function first updates the specific PMU within the Measurement type using the provided kwargs, and then updates the Analysis type with all parameters associated with that PMU.

A key feature of this function is that any prior modifications made to the specified PMU are preserved and applied to the Analysis type when the function is executed, ensuring consistency throughout the update process.

Example

system, monitoring = ems("case14.h5", "monitoring.h5")

analysis = gaussNewton(monitoring)
stateEstimation!(analysis)

updatePmu!(analysis; label = 4, magnitude = 0.95, angle = -0.1)

JuliaGrid.statusPmu! — Function

statusPmu!(monitoring::Measurement;
    inservice, inserviceBus, inserviceFrom, inserviceTo,
    outservice, outserviceBus outserviceFrom, outserviceTo,
    redundancy, redundancyBus, redundancyFrom, redundancyTo)

The function generates a set of PMUs, assigning PMUs randomly to either in-service or out-of-service states based on specified keywords. It is important to note that when we refer to PMU, we encompass both magnitude and angle measurements.

Keywords

Users may use either one main keyword or three fine-tuning keywords that specify distinct locations per function call:

inservice: Sets the number of in-service PMUs or allows fine-tuning:

inserviceBus: sets only PMUs loacted at the bus,

inserviceFrom: sets only PMUs loacted at the from-bus end,

inserviceTo: sets only PMUs loacted at the to-bus end.

outservice: Sets the number of out-of-service PMUs or allows fine-tuning:

outserviceBus: sets only PMUs loacted at the bus,

outserviceFrom: sets only PMUs loacted at the from-bus end,

outserviceTo: sets only PMUs loacted at the to-bus end.

redundancy: Determines in-service PMUs based on redundancy or allows fine-tuning:

redundancyBus: determines only PMUs loacted at the bus,

redundancyFrom: determines only PMUs loacted at the from-bus end,

redundancyTo: determines only PMUs loacted at the to-bus end.

Updates

The function updates the status fields within the PMU type.

Example

system, monitoring = ems("case14.h5")

analysis = newtonRaphson(system)
powerFlow!(analysis; current = true)

addPmu!(monitoring, analysis)

statusPmu!(monitoring; inserviceBus = 10)

JuliaGrid.@pmu — Macro

@pmu(label, noise, correlated, polar, square,
    varianceMagnitudeBus, varianceAngleBus, statusBus,
    varianceMagnitudeFrom, varianceAngleFrom, statusFrom,
    varianceMagnitudeTo, varianceAngleTo, statusTo)

The macro generates a template for a PMU.

Keywords

To establish the PMU template, users can configure the pattern for labels using the label keyword, specify the type of noise, and indicate the correlated,polar, andsquare` system utilized for managing phasors during state estimation.

Users have the option to set default values for magnitude and angle variances, as well as statuses. This can be done for PMUs located at the buses using the varianceMagnitudeBus, varianceAngleBus, and statusBus keywords.

The same configuration can be applied at both the from-bus ends of the branches using the varianceMagnitudeFrom, varianceAngleFrom, and statusFrom keywords.

For PMUs located at the to-bus ends of the branches, users can use the varianceMagnitudeTo, varianceAngleTo, and statusTo keywords.

Units

By default, the units for variances are per-units and radians. However, users have the option to switch to volts and degrees as the units for PMUs located at the buses by using the @voltage macro, or they can switch to amperes and degrees as the units for PMUs located at the branches by using the @current macro.

Examples

Adding PMUs using the default unit system:

@pmu(label = "PMU ?", varianceAngleBus = 1e-6, varianceMagnitudeFrom = 1e-4)

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132e3)
addBus!(system; label = "Bus 2", base = 132e3)
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addPmu!(monitoring; bus = "Bus 1", magnitude = 1.1, angle = -0.1)
addPmu!(monitoring; from = "Branch 1", magnitude = 1.1, angle = -0.2)

Adding PMUs using a custom unit system:

@voltage(kV, deg, kV)
@current(A, deg)
@pmu(label = "PMU ?", varianceAngleBus = 5.73e-5, varianceMagnitudeFrom = 0.0481)

system, monitoring = ems()

addBus!(system; label = "Bus 1", base = 132.0)
addBus!(system; label = "Bus 2", base = 132.0)
addBranch!(system; label = "Branch 1", from = "Bus 1", to = "Bus 2", reactance = 0.2)

addPmu!(monitoring; bus = "Bus 1", magnitude = 145.2, angle = -5.73)
addPmu!(monitoring; from = "Branch 1", magnitude = 481.125, angle = -11.46)