Optimal Power Flow

For further information on this topic, please see the AC Optimal Power Flow or DC Optimal Power Flow sections of the Manual. Below, we have provided a list of functions that can be utilized for optimal power flow analysis.

Build AC Optimal Power Flow Model

acOptimalPowerFlow

Solve AC Optimal Power Flow

solve!

Build DC Optimal Power Flow Model

dcOptimalPowerFlow

Solve DC Optimal Power Flow

solve!

Additional Function

startingPrimal!

Build AC Optimal Power Flow Model

JuliaGrid.acOptimalPowerFlow — Function

acOptimalPowerFlow(system::PowerSystem, optimizer; bridge, name)

The function takes the PowerSystem composite type as input to establish the structure for solving the AC optimal power flow. The optimizer argument is also required to create and solve the optimization problem. If the ac field within the PowerSystem composite type has not been created, the function will initiate an update automatically.

Additionally, the optimizer argument is a necessary component for formulating and solving the optimization problem. Specifically, JuliaGrid constructs the AC optimal power flow using the JuMP package and provides support for commonly employed solvers. For more detailed information, please consult the JuMP documenatation.

Keywords

JuliaGrid offers the ability to manipulate the jump model based on the guidelines provided in the JuMP documentation. However, certain configurations may require different method calls, such as:

bridge: used to manage the bridging mechanism;

name: used to manage the creation of string names.

By default, these keyword settings are configured as true.

Returns

The function returns an instance of the ACOptimalPowerFlow type, which includes the following fields:

voltage: the bus voltage magnitudes and angles;

power: the variable allocated to store the active and reactive powers;

current: the variable allocated to store the currents;

jump: the JuMP model;

variable: holds the variable references to the JuMP model;

constraint: holds the constraint references to the JuMP model;

objective: holds the objective expression of the JuMP model.

Examples

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)

Solve AC Optimal Power Flow

JuliaGrid.solve! — Method

solve!(system::PowerSystem, analysis::ACOptimalPowerFlow)

The function determines the optimal power flow for AC systems, computing the magnitudes and angles of bus voltages, as well as generating active and reactive power values for each generator.

Updates

The calculated voltage magnitudes and angles and active and reactive powers are then stored in the variables of the voltage and power.generator fields of the ACOptimalPowerFlow composite type.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)

Build DC Optimal Power Flow Model

JuliaGrid.dcOptimalPowerFlow — Function

dcOptimalPowerFlow(system::PowerSystem, optimizer; bridge, name)

The function takes the PowerSystem composite type as input to establish the structure for solving the DC optimal power flow. If the dc field within the PowerSystem composite type has not been created, the function will automatically initiate an update process.

Additionally, the optimizer argument is a necessary component for formulating and solving the optimization problem. Specifically, JuliaGrid constructs the DC optimal power flow using the JuMP package and provides support for commonly employed solvers. For more detailed information, please consult the JuMP documenatation.

Keywords

JuliaGrid offers the ability to manipulate the jump model based on the guidelines provided in the JuMP documentation. However, certain configurations may require different method calls, such as:

bridge: used to manage the bridging mechanism;

name: used to manage the creation of string names.

By default, these keyword settings are configured as true.

Returns

The function returns an instance of the DCOptimalPowerFlow type, which includes the following fields:

voltage: the variable allocated to store the bus voltage angle;

power: the variable allocated to store the active powers;

jump: the JuMP model;

variable: holds the variable references to the JuMP model;

constraint: holds the constraint references to the JuMP model;

objective: holds the objective expression of the JuMP model.

Examples

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcOptimalPowerFlow(system, HiGHS.Optimizer)

Solve DC Optimal Power Flow

JuliaGrid.solve! — Method

solve!(system::PowerSystem, analysis::DCOptimalPowerFlow)

The function determines the optimal power flow for DC systems, computing the angles of bus voltages, as well as generating active power values for each generator.

Updates

The calculated voltage angles and active powers are then stored in the variables of the voltage and power.generator fields of the DCOptimalPowerFlow composite type.

Example

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcOptimalPowerFlow(system, HiGHS.Optimizer)
solve!(system, analysis)

Additional Function

JuliaGrid.startingPrimal! — Method

startingPrimal!(system::PowerSystem, analysis::OptimalPowerFlow)

In the context of the ACOptimalPowerFlow composite type, this function retrieves the active and reactive power outputs of the generators, as well as the voltage magnitudes and angles from the PowerSystem composite type. It then assigns these values to the ACOptimalPowerFlow type, allowing users to initialize starting primal values as needed.

For the DCOptimalPowerFlow composite type, this function retrieves the active power outputs of the generators and the bus voltage angles from the PowerSystem composite type. These values are then assigned to the DCOptimalPowerFlow type, enabling users to initialize starting primal values according to their requirements.

Updates

This function only updates the voltage and generator fields of the OptimalPowerFlow abstract type.

Abstract type

The abstract type OptimalPowerFlow can have the following subtypes:

ACOptimalPowerFlow: employed to initializing starting primal values within the AC optimal power flow;

DCOptimalPowerFlow: employed to initialize starting primal values within the DC optimal power flow.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)

updateBus!(system, analysis; label = 14, reactive = 0.13, magnitude = 1.2, angle = -0.17)

startingPrimal!(system, analysis)
solve!(system, analysis)