Optimal Power Flow

For further information on this topic, please see the AC Optimal Power Flow or DC Optimal Power Flow sections of the Manual. Below, we have provided a list of functions that can be utilized for optimal power flow analysis.

Build Model

acOptimalPowerFlow

dcOptimalPowerFlow

Solve Optimal Power Flow

solve!

Power Analysis

power!

powerInjection

powerSupply

powerShunt

powerFrom

powerTo

powerCharging

powerSeries

Current Analysis

current!

currentInjection

currentFrom

currentTo

currentSeries

Build Model

JuliaGrid.acOptimalPowerFlow — Function

acOptimalPowerFlow(system::PowerSystem, optimizer; bridge, name, balance, limit,
    rating, capability)

The function takes the PowerSystem composite type as input to establish the structure for solving the AC optimal power flow. The optimizer argument is also required to create and solve the optimization problem. If the ac field within the PowerSystem composite type has not been created, the function will initiate an update automatically.

Keywords

JuliaGrid offers the ability to manipulate the jump model based on the guidelines provided in the JuMP documentation. However, certain configurations may require different method calls, such as:

bridge: used to manage the bridging mechanism;

name: used to manage the creation of string names.

Moreover, we have included keywords that regulate the usage of different types of constraints:

balance: controls the equality constraints that relate to the active and reactive power balance equations;

limit: controls the inequality constraints that relate to the voltage magnitude and angle differences between buses;

rating: controls the inequality constraints that relate to the long-term rating of branches;

capability: controls the inequality constraints that relate to the active and reactive power generator outputs.

By default, all of these keywords are set to true and are of the Bool type.

JuMP

The JuliaGrid builds the AC optimal power flow around the JuMP package and supports commonly used solvers. For more information, refer to the JuMP documenatation.

Returns

The function returns an instance of the ACOptimalPowerFlow type, which includes the following fields:

voltage: the bus voltage magnitudes and angles;

power: the variable allocated to store the active and reactive powers;

current: the variable allocated to store the currents;

jump: the JuMP model;

constraint: holds the constraint references to the JuMP model.

Examples

Create the complete AC optimal power flow model:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)

Create the AC optimal power flow model without rating constraints:

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer; rating = false)

JuliaGrid.dcOptimalPowerFlow — Function

dcOptimalPowerFlow(system::PowerSystem, optimizer; bridge, name, balance, limit,
    rating, capability)

The function takes the PowerSystem composite type as input to establish the structure for solving the DC optimal power flow. The optimizer argument is also required to create and solve the optimization problem. If the dc field within the PowerSystem composite type has not been created, the function will initiate an update automatically.

Keywords

JuliaGrid offers the ability to manipulate the jump model based on the guidelines provided in the JuMP documentation. However, certain configurations may require different method calls, such as:

bridge: used to manage the bridging mechanism;

name: used to manage the creation of string names.

Moreover, we have included keywords that regulate the usage of different types of constraints:

balance: controls the equality constraints that relate to the active power balance equations;

limit: controls the inequality constraints that relate to the voltage angle differences between buses;

rating: controls the inequality constraints that relate to the long-term rating of branches;

capability: controls the inequality constraints that relate to the active power generator outputs.

By default, all of these keywords are set to true and are of the Bool type.

JuMP

The JuliaGrid builds the DC optimal power flow around the JuMP package and supports commonly used solvers. For more information, refer to the JuMP documenatation.

Returns

The function returns an instance of the DCOptimalPowerFlow type, which includes the following fields:

voltage: the variable allocated to store the bus voltage angle,

power: the variable allocated to store the active powers,

jump: the JuMP model,

constraint: holds the constraint references to the JuMP model.

Examples

Create the complete DC optimal power flow model:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcOptimalPowerFlow(system, HiGHS.Optimizer)

Create the DC optimal power flow model without rating constraints:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcOptimalPowerFlow(system, HiGHS.Optimizer; rating = false)

Solve Optimal Power Flow

JuliaGrid.solve! — Method

solve!(system::PowerSystem, analysis::ACOptimalPowerFlow)

The function finds the AC optimal power flow solution and calculate the bus voltage magnitudes and angles, and output active and reactive powers of each generators.

The calculated voltage magnitudes and angles and active and reactive powers are then stored in the variables of the voltage and power fields of the ACOptimalPowerFlow composite type.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = acOptimalPowerFlow(system, Ipopt.Optimizer)
solve!(system, analysis)

JuliaGrid.solve! — Method

solve!(system::PowerSystem, analysis::DCOptimalPowerFlow)

The function finds the DC optimal power flow solution and calculate the bus voltage angles and output active powers of the generators.

The calculated voltage angles and active powers are then stored in the angle variable of the voltage field and the generator variable of the power field.

Example

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcOptimalPowerFlow(system, HiGHS.Optimizer)
solve!(system, analysis)