Postprocessing Analysis

For further information on this topic, please see the Power System Model section of the Manual.

API Index

analysisBus

analysisBranch

analysisGenerator

Bus Function

JuliaGrid.analysisBus — Function

analysisBus(system::PowerSystem, model::ACPowerFlow)

The function computes the powers and currents associated with buses in the AC power flow framework.

Returns

The function returns two instances of types.

The PowerBus type contains the following fields:

injection: the active and reactive power injections

supply: the active and reactive power injected by the generators

shunt: the active and reactive power associated with shunt elements.

The CurrentBus type contains the following field:

injection: the magnitude and angle of current injections.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

power, current = analysisBus(system, model)

analysisBus(system::PowerSystem, model::DCPowerFlow)

The function returns the active powers associated with buses in the DC power flow framework.

Returns

The DCPowerBus type contains the following fields:

injection: the active power injections

supply: the active power injected by the generators.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

power = analysisBus(system, model)

Branch Function

JuliaGrid.analysisBranch — Function

analysisBranch(system::PowerSystem, model::ACPowerFlow)

The function returns the powers and currents associated with branches in the AC power flow framework.

Returns

The function returns two instances of types.

The PowerBranch type contains the following fields:

from: the active and reactive power flows at the from ends

to: the active and reactive power flows at the to ends

shunt: the reactive power injections

loss: the active and reactive power losses.

The CurrentBranch type contains the following field:

from: the magnitude and angle of current flows at from bus ends

to: the magnitude and angle of current flows at to bus ends

impedance: the magnitude and angle of current flows through series impedances.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

power, current = analysisBranch(system, model)

analysisBranch(system::PowerSystem, model::DCPowerFlow)

The function returns the active powers associated with branches in the DC power flow framework.

Returns

The DCPowerBranch type contains the following fields:

from: the active power flows at the from ends

to: the active power flows at the to ends

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

power = analysisBranch(system, model)

Generator Function

JuliaGrid.analysisGenerator — Function

analysisGenerator(system::PowerSystem, model::ACPowerFlow)

The function return powers related to generators for the AC power flow analysis.

Returns

The PowerGenerator type contains the following variables:

active: the active power output of the generators

reactive: the reactive power output of the generators.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

power = analysisGenerator(system, model)

analysisGenerator(system::PowerSystem, model::DCPowerFlow)

The function returns powers related to generators in the DC power flow framework.

Returns

The DCPowerGenerator type contains the following field:

active: the active power output of the generators.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

power = analysisGenerator(system, model)