Postprocessing Analysis

For further information on this topic, please see the Power System Model section of the Manual.

API Index

analysisBus

analysisBranch

analysisGenerator

Bus Function

JuliaGrid.analysisBus — Function

The function returns the powers and currents associated with buses in the AC power flow framework.

analysisBus(system::PowerSystem, model::ACPowerFlow)

In particular, it computes the power injections, power injected by the generators, power associated with shunt elements, and current injections.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

power, current = analysisBus(system, model)

The function returns the active powers associated with buses in the DC power flow framework.

analysisBus(system::PowerSystem, model::DCPowerFlow)

In particular, it computes the active power injections and active power injected by generators.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

power = analysisBus(system, model)

Branch Function

JuliaGrid.analysisBranch — Function

The function returns the powers and currents associated with branches in the AC power flow framework.

analysisBranch(system::PowerSystem, model::ACPowerFlow)

In particular, it computes the power flows at from and to bus ends, power losses, reactive power injections, current flows at from and to bus ends, and current flow through series impedances.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

power, current = analysisBranch(system, model)

The function returns the active powers associated with branches in the DC power flow framework.

analysisBranch(system::PowerSystem, model::DCPowerFlow)

In particular, it computes the active power flows at from and to bus ends.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

power = analysisBranch(system, model)

Generator Function

JuliaGrid.analysisGenerator — Function

The function return powers related to generators for the AC power flow analysis.

analysisGenerator(system::PowerSystem, model::ACPowerFlow)

In particular, it computes the power output of the generators.

Example

system = powerSystem("case14.h5")
acModel!(system)

model = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, model)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, model)
end

power = analysisGenerator(system, model)

The function returns powers related to generators in the DC power flow framework.

analysisGenerator(system::PowerSystem, model::DCPowerFlow)

In particular, it computes the active power output of the generators.

Example

system = powerSystem("case14.h5")
dcModel!(system)

model = dcPowerFlow(system)
solve!(system, model)

power = analysisGenerator(system, model)