Power Flow

For further information on this topic, please see the AC Power Flow or DC Power Flow sections of the Manual. Below, we have provided a list of functions that can be utilized for power flow analysis.

To load power flow API functionalities into the current scope, utilize the following command:

using JuliaGrid

AC Power Flow

newtonRaphson

fastNewtonRaphsonBX

fastNewtonRaphsonXB

gaussSeidel

mismatch!

solve!

setInitialPoint!

reactiveLimit!

adjustAngle!

powerFlow!

DC Power Flow

dcPowerFlow

solve!

powerFlow!

AC Power Flow

JuliaGrid.newtonRaphson — Function

newtonRaphson(system::PowerSystem, [factorization::Factorization = LU])

The function sets up the Newton-Raphson method to solve the AC power flow.

Arguments

The function requires the PowerSystem type to establish the framework. Next, the Factorization argument, while optional, determines the method used to solve the linear system of equations within each iteration. It can take one of the following values:

LU: Utilizes LU factorization (default).

QR: Utilizes QR factorization.

Updates

If the AC model has not been created, the function automatically initiates an update within the ac field of the PowerSystem type. It also performs a check on bus types and rectifies any mistakes present.

Returns

The function returns an instance of the ACPowerFlow type, which includes the following fields:

voltage: The bus voltage magnitudes and angles.

power: The variable allocated to store the active and reactive powers.

current: The variable allocated to store the currents.

method: The Jacobian matrix, its factorization, mismatches, increments, and indices.

Examples

Set up the Newton-Raphson method utilizing LU factorization:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)

Set up the Newton-Raphson method utilizing QR factorization:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system, QR)

JuliaGrid.fastNewtonRaphsonBX — Function

fastNewtonRaphsonBX(system::PowerSystem, [factorization::Factorization = LU])

The function sets up the fast Newton-Raphson method of version BX to solve the AC power flow.

Arguments

The function requires the PowerSystem type to establish the framework. Next, the Factorization argument, while optional, determines the method used to solve the linear system of equations within each iteration. It can take one of the following values:

LU: Utilizes LU factorization (default).

QR: Utilizes QR factorization.

Updates

If the AC model has not been created, the function automatically initiates an update within the ac field of the PowerSystem type. It also performs a check on bus types and rectifies any mistakes present.

Returns

The function returns an instance of the ACPowerFlow type, which includes the following fields:

voltage: The bus voltage magnitudes and angles.

power: The variable allocated to store the active and reactive powers.

current: The variable allocated to store the currents.

method: The Jacobian matrices, their factorizations, mismatches, increments, and indices.

Examples

Set up the fast Newton-Raphson method utilizing LU factorization:

system = powerSystem("case14.h5")
acModel!(system)

analysis = fastNewtonRaphsonBX(system)

Set up the fast Newton-Raphson method utilizing QR factorization:

system = powerSystem("case14.h5")
acModel!(system)

analysis = fastNewtonRaphsonBX(system, QR)

JuliaGrid.fastNewtonRaphsonXB — Function

fastNewtonRaphsonXB(system::PowerSystem, [factorization::Factorization = LU])

The function sets up the fast Newton-Raphson method of version XB to solve the AC power flow.

Arguments

The function requires the PowerSystem type to establish the framework. Next, the Factorization argument, while optional, determines the method used to solve the linear system of equations within each iteration. It can take one of the following values:

LU: Utilizes LU factorization (default).

QR: Utilizes QR factorization.

Updates

If the AC model has not been created, the function automatically initiates an update within the ac field of the PowerSystem type. It also performs a check on bus types and rectifies any mistakes present.

Returns

The function returns an instance of the ACPowerFlow type, which includes the following fields:

voltage: The bus voltage magnitudes and angles.

power: The variable allocated to store the active and reactive powers.

current: The variable allocated to store the currents.

method: The Jacobian matrices, their factorizations, mismatches, increments, and indices.

Examples

Set up the fast Newton-Raphson method utilizing LU factorization:

system = powerSystem("case14.h5")
acModel!(system)

analysis = fastNewtonRaphsonXB(system)

Set up the fast Newton-Raphson method utilizing QR factorization:

system = powerSystem("case14.h5")
acModel!(system)

analysis = fastNewtonRaphsonXB(system, QR)

JuliaGrid.gaussSeidel — Function

gaussSeidel(system::PowerSystem)

The function sets up the Gauss-Seidel method to solve the AC power flow.

Argument

The function requires the PowerSystem type to establish the framework.

Updates

If the AC model has not been created, the function automatically initiates an update within the ac field of the PowerSystem type. It also performs a check on bus types and rectifies any mistakes present.

Returns

The function returns an instance of the ACPowerFlow type, which includes the following fields:

voltage: The bus voltage magnitudes and angles.

power: The variable allocated to store the active and reactive powers.

current: The variable allocated to store the currents.

method: The bus complex voltages and indices.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = gaussSeidel(system)

JuliaGrid.mismatch! — Method

mismatch!(system::PowerSystem, analysis::ACPowerFlow)

The function calculates both active and reactive power injection mismatches.

Updates

This function updates the mismatch variables in the Newton-Raphson and fast Newton-Raphson methods. It should be employed during the iteration loop before invoking the solve! function.

Returns

The function returns maximum absolute values of the active and reactive power injection mismatches, which can be utilized to terminate the iteration loop of the Newton-Raphson, fast Newton-Raphson, or Gauss-Seidel methods employed to solve the AC power flow problem.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
mismatch!(system, analysis)

JuliaGrid.solve! — Method

solve!(system::PowerSystem, analysis::ACPowerFlow)

The function employs the Newton-Raphson, fast Newton-Raphson, or Gauss-Seidel method to solve the AC power flow and calculate bus voltage magnitudes and angles.

After the mismatch! function is executed, the execution of this function will perform a single iteration of one of the methods.

Updates

The calculated voltages are stored in the voltage field of the ACPowerFlow type.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

JuliaGrid.setInitialPoint! — Method

setInitialPoint!(source::Union{PowerSystem, Analysis}, target::ACPowerFlow)

The function can reset the initial point of the AC power flow to values from the PowerSystem type. It can also initialize the AC power flow based on results from the Analysis type, whether from an AC or DC analysis.

The function assigns the bus voltage magnitudes and angles in the target argument, using data from the source argument. This allows users to initialize AC power flow as needed.

If source comes from a DC analysis, only the bus voltage angles are assigned in the target argument, while the bus voltage magnitudes remain unchanged.

Examples

Initialize the Newton-Raphson method with values from the Gauss-Seidel method:

system = powerSystem("case14.h5")
acModel!(system)

gs = gaussSeidel(system)
for i = 1:10
    solve!(system, gs)
end

nr = newtonRaphson(system)

setInitialPoint!(gs, nr)
for i = 1:10
    stopping = mismatch!(system, nr)
    if all(stopping .< 1e-10)
        break
    end
    solve!(system, nr)
end

Use wrapper functions and initialize Newton-Raphson with Gauss-Seidel values:

system = powerSystem("case14.h5")
acModel!(system)

gs = gaussSeidel(system)
powerFlow!(system, gs; iteration = 10)

nr = newtonRaphson(system)

setInitialPoint!(gs, nr)
powerFlow!(system, nr; tolerance = 1e-10)

JuliaGrid.reactiveLimit! — Function

reactiveLimit!(system::PowerSystem, analysis::ACPowerFlow)

The function verifies whether the generators in a power system exceed their reactive power limits. This is done by setting the reactive power of the generators to within the limits if they are violated after determining the bus voltage magnitudes and angles. If the limits are violated, the corresponding generator buses or the slack bus are converted to demand buses.

Updates

The function assigns values to the generator.output.active and bus.supply.active variables of the PowerSystem type.

Additionally, it examines the reactive powers of the generators and adjusts them to their maximum or minimum values if they exceed the specified threshold. This results in the modification of the variable generator.output.reactive of the PowerSystem type accordingly.

As a result of this adjustment, the bus.supply.reactive variable is also updated, and the bus types specified in bus.layout.type are modified. If the slack bus is converted, the bus.layout.slack field is correspondingly adjusted.

Returns

The function returns the variable to indicate which buses violate the limits, with -1 indicating a violation of the minimum limits and 1 indicating a violation of the maximum limits.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

violate = reactiveLimit!(system, analysis)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

JuliaGrid.adjustAngle! — Function

adjustAngle!(system::PowerSystem, analysis::ACPowerFlow; slack)

The function modifies the bus voltage angles based on a different slack bus than the one identified by the bus.layout.slack field.

For instance, if the reactive power of the generator exceeds the limit on the slack bus, the reactiveLimit! function will change that bus to the demand bus and designate the first generator bus in the sequence as the new slack bus. After obtaining the updated AC power flow solution based on the new slack bus, it is possible to adjust the voltage angles to align with the angle of the original slack bus.

Keyword

The slack keyword specifies the bus label for which we want to adjust voltage angles.

Updates

This function only updates the voltage.angle variable of the ACPowerFlow type.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

reactiveLimit!(system, analysis)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

adjustAngle!(system, analysis; slack = 1)

JuliaGrid.powerFlow! — Method

powerFlow!(system::PowerSystem, analysis::ACPowerFlow;
    iteration, tolerance, power, current, verbose)

The function serves as a wrapper for solving AC power flow and includes the functions:

mismatch!,

solve!,

power!,

current!.

It computes bus voltage magnitudes and angles using an iterative algorithm with the option to compute powers and currents.

Keywords

Users can use the following keywords:

iteration: Specifies the maximum number of iterations (default: 20).

tolerance: Defines the step size tolerance for the iteration stopping criterion (default: 1e-8).

power: Enables power computation upon convergence or reaching the iteration limit (default: false).

current: Enables current computation upon convergence or reaching the iteration limit (default: false).

verbose: Controls the output display, ranging from the default silent mode (0) to detailed output (3).

Examples

Solve the AC power flow using the Newton-Raphson method:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
powerFlow!(system, analysis; iteration = 30, tolerance = 1e-10, verbose = 3)

Solve the AC power flow using the Gauss-Seidel method:

system = powerSystem("case14.h5")
acModel!(system)

analysis = gaussSeidel(system)
powerFlow!(system, analysis; iteration = 200, power = true, current = true, verbose = 3)

DC Power Flow

JuliaGrid.dcPowerFlow — Function

dcPowerFlow(system::PowerSystem, [factorization::Factorization = LU])

The function sets up the framework to solve the DC power flow.

Arguments

The function requires the PowerSystem type to establish the framework. Next, the Factorization argument, while optional, determines the method used to solve the linear system of equations. It can take one of the following values:

LU: Utilizes LU factorization (default).

LDLt: Utilizes LDLt factorization.

QR: Utilizes QR factorization.

Updates

If the DC model was not created, the function will automatically initiate an update of the dc field within the PowerSystem composite type. Additionally, if the slack bus lacks an in-service generator, JuliaGrid considers it a mistake and defines a new slack bus as the first generator bus with an in-service generator in the bus type list.

Returns

The function returns an instance of the DCPowerFlow type, which includes the following fields:

voltage: The variable allocated to store the bus voltage angles.

power: The variable allocated to store the active powers.

method: The factorized nodal admittance matrix.

Examples

Set up the DC power flow utilizing LU factorization:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system)

Set up the DC power flow utilizing QR factorization:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system, QR)

JuliaGrid.solve! — Method

solve!(system::PowerSystem, analysis::DCPowerFlow)

The function solves the DC power flow model and calculates bus voltage angles.

Updates

The calculated voltage angles are stored in the voltage field of the DCPowerFlow type.

Example

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system)
solve!(system, analysis)

JuliaGrid.powerFlow! — Method

powerFlow!(system::PowerSystem, analysis::DCPowerFlow; power, verbose)

The function serves as a wrapper for solving DC power flow and includes the functions:

solve!,

power!.

It computes bus voltage angles and optionally calculates power values.

Keywords

Users can use the following keywords:

power: Enables the computation of powers (default: false).

verbose: Controls the output display, ranging from the default silent mode (0) to detailed output (3).

Example

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system)
powerFlow!(system, analysis; power = true, verbose = 1)