Power Flow

For further information on this topic, please see the AC Power Flow or DC Power Flow sections of the Manual. Below, we have provided a list of functions that can be utilized for power flow analysis.

Build AC Power Flow

newtonRaphson

fastNewtonRaphsonBX

fastNewtonRaphsonXB

gaussSeidel

Solve AC Power Flow

mismatch!

solve!

Additional AC Functions

reactiveLimit!

adjustAngle!

startingVoltage!

Build DC Power Flow

dcPowerFlow

Solve DC Power Flow

solve!

Build AC Power Flow

JuliaGrid.newtonRaphson — Function

newtonRaphson(system::PowerSystem, factorization::Factorization)

The function sets up the Newton-Raphson method to solve the AC power flow.

Arguments

The function requires the PowerSystem composite type to establish the framework. Moreover, the Factorization argument is optional and can be one of the following:

LU: utilizes LU factorization to solve the linear system of equations within each iteration;

QR: utilizes QR factorization to solve the linear system of equations within each iteration.

If the user does not provide the Factorization composite type, the default solving method will be LU factorization.

Updates

If the AC model has not been created, the function automatically initiates an update within the ac field of the PowerSystem composite type. It also performs a check on bus types and rectifies any mistakes present.

Returns

The function returns an instance of the NewtonRaphson subtype of the abstract ACPowerFlow type, which includes the following fields:

voltage: the bus voltage magnitudes and angles;

power: the variable allocated to store the active and reactive powers;

current: the variable allocated to store the currents;

method: the Jacobian matrix, its factorization, power injection mismatches, bus voltage increments, and indices.

Examples

Set up the Newton-Raphson method using LU factorization within each iteration:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)

Set up the Newton-Raphson method using QR factorization within each iteration:

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system, QR)

JuliaGrid.fastNewtonRaphsonBX — Function

fastNewtonRaphsonBX(system::PowerSystem, factorization::Factorization)

The function sets up the fast Newton-Raphson method of version BX to solve the AC power flow.

Arguments

The function requires the PowerSystem composite type to establish the framework. Moreover, the Factorization argument is optional and can be one of the following:

LU: utilizes LU factorization to factorize constant Jacobian matrices;

QR: utilizes QR factorization to factorize constant Jacobian matrices.

If the user does not provide the Factorization composite type, the default solving method will be LU factorization.

Updates

If the AC model has not been created, the function automatically initiates an update within the ac field of the PowerSystem composite type. It also performs a check on bus types and rectifies any mistakes present.

Returns

The function returns an instance of the FastNewtonRaphson subtype of the abstract ACPowerFlow type, which includes the following fields:

voltage: the bus voltage magnitudes and angles;

power: the variable allocated to store the active and reactive powers;

current: the variable allocated to store the currents;

method: contains Jacobian matrices, its factorization, power injection mismatches, bus voltage increments, and indices.

Examples

Set up the fast Newton-Raphson method utilizing LU factorization:

system = powerSystem("case14.h5")
acModel!(system)

analysis = fastNewtonRaphsonBX(system)

Set up the fast Newton-Raphson method utilizing QR factorization:

system = powerSystem("case14.h5")
acModel!(system)

analysis = fastNewtonRaphsonBX(system, QR)

JuliaGrid.fastNewtonRaphsonXB — Function

fastNewtonRaphsonXB(system::PowerSystem, factorization::Factorization)

The function sets up the fast Newton-Raphson method of version XB to solve the AC power flow.

Arguments

The function requires the PowerSystem composite type to establish the framework. Moreover, the Factorization argument is optional and can be one of the following:

LU: utilizes LU factorization to factorize constant Jacobian matrices;

QR: utilizes QR factorization to factorize constant Jacobian matrices.

If the user does not provide the Factorization composite type, the default solving method will be LU factorization.

Updates

If the AC model has not been created, the function automatically initiates an update within the ac field of the PowerSystem composite type. It also performs a check on bus types and rectifies any mistakes present.

Returns

The function returns an instance of the FastNewtonRaphson subtype of the abstract ACPowerFlow type, which includes the following fields:

voltage: the bus voltage magnitudes and angles;

power: the variable allocated to store the active and reactive powers;

current: the variable allocated to store the currents;

method: contains Jacobian matrices, its factorization, power injection mismatches, bus voltage increments, and indices.

Examples

Set up the fast Newton-Raphson method utilizing LU factorization:

system = powerSystem("case14.h5")
acModel!(system)

analysis = fastNewtonRaphsonXB(system)

Set up the fast Newton-Raphson method utilizing QR factorization:

system = powerSystem("case14.h5")
acModel!(system)

analysis = fastNewtonRaphsonXB(system, QR)

JuliaGrid.gaussSeidel — Function

gaussSeidel(system::PowerSystem)

The function accepts the PowerSystem composite type as input and uses it to set up the Gauss-Seidel method to solve AC power flow. Additionally, if the AC model was not created, the function will automatically initiate an update of the ac field within the PowerSystem composite type.

Returns

The function returns an instance of the GaussSeidel subtype of the abstract ACPowerFlow type, which includes the following fields:

voltage: the bus voltage magnitudes and angles;

power: the variable allocated to store the active and reactive powers;

current: the variable allocated to store the currents;

method: contains the bus complex voltages and indices.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = gaussSeidel(system)

Solve AC Power Flow

JuliaGrid.mismatch! — Method

mismatch!(system::PowerSystem, analysis::ACPowerFlow)

The function calculates both active and reactive power injection mismatches and returns their maximum absolute values, which can be utilized to terminate the iteration loop of methods employed to solve the AC power flow problem.

Updates

This function updates the mismatch variables in the Newton-Raphson and fast Newton-Raphson methods. It should be employed during the iteration loop before invoking the solve! function.

In contrast, the Gauss-Seidel method does not need mismatches to obtain bus voltages, but the maximum absolute values are commonly employed to stop the iteration loop. The function does not save any data and should be utilized during the iteration loop before invoking the solve! function.

Abstract type

The abstract type ACPowerFlow can have the following subtypes:

NewtonRaphson: computes the power mismatches within the Newton-Raphson method;

FastNewtonRaphson: computes the power mismatches within the fast Newton-Raphson method;

GaussSeidel: computes the power mismatches within the Gauss-Seidel method.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
mismatch!(system, analysis)

JuliaGrid.solve! — Method

solve!(system::PowerSystem, analysis::ACPowerFlow)

The function employs the Newton-Raphson, fast Newton-Raphson, or Gauss-Seidel method to solve the AC power flow problem and calculate bus voltage magnitudes and angles.

After the mismatch! function is called, this function should be executed to perform a single iteration of the method.

Updates

The calculated voltages are stored in the voltage field of the ACPowerFlow abstract type.

Abstract type

The abstract type ACPowerFlow can have the following subtypes:

NewtonRaphson: computes the bus voltages within the Newton-Raphson method;

FastNewtonRaphson: computes the bus voltages within the fast Newton-Raphson method;

GaussSeidel: computes the bus voltages within the Gauss-Seidel method.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

Additional AC Functions

JuliaGrid.reactiveLimit! — Function

reactiveLimit!(system::PowerSystem, analysis::ACPowerFlow)

The function verifies whether the generators in a power system exceed their reactive power limits. This is done by setting the reactive power of the generators to within the limits if they are violated, after determining the bus voltage magnitudes and angles. If the limits are violated, the corresponding generator buses or the slack bus are converted to demand buses.

Updates

The function assigns values to the generator.output.active and bus.supply.active variables of the PowerSystem type. Additionally, it examines the reactive powers of the generator and adjusts them to their maximum or minimum values if they exceed the specified threshold. Subsequently, the generator.output.reactive variable of the PowerSystem type is modified accordingly. As a result of this adjustment, the bus.supply.reactive variable of the PowerSystem type is also updated, and the bus types specified in bus.layout.type are modified. If the slack bus is converted, the bus.layout.slack field is correspondingly adjusted.

Returns

The function returns the violate variable to indicate which buses violate the limits, with -1 indicating a violation of the minimum limits and 1 indicating a violation of the maximum limits.

Abstract type

The abstract type ACPowerFlow can have the following subtypes:

NewtonRaphson: computes the bus voltages within the Newton-Raphson method,

FastNewtonRaphson: computes the bus voltages within the fast Newton-Raphson method,

GaussSeidel: computes the bus voltages within the Gauss-Seidel method.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

violate = reactiveLimit!(system, analysis)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

JuliaGrid.adjustAngle! — Function

adjustAngle!(system::PowerSystem, analysis::ACPowerFlow; slack)

The function modifies the bus voltage angles based on a different slack bus than the one identified by the bus.layout.slack field.

For instance, if the reactive power of the generator exceeds the limit on the slack bus, the reactiveLimit! function will change that bus to the demand bus and designate the first generator bus in the sequence as the new slack bus. After obtaining the updated AC power flow solution based on the new slack bus, it is possible to adjust the voltage angles to align with the angle of the original slack bus. The slack keyword specifies the bus label of the original slack bus.

Updates

This function only updates the voltage.angle variable of the ACPowerFlow abstract type.

Abstract type

The abstract type ACPowerFlow can have the following subtypes:

NewtonRaphson: computes the bus voltages within the Newton-Raphson method;

FastNewtonRaphson: computes the bus voltages within the fast Newton-Raphson method;

GaussSeidel: computes the bus voltages within the Gauss-Seidel method.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

reactiveLimit!(system, analysis)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

adjustAngle!(system, analysis; slack = 1)

JuliaGrid.startingVoltage! — Function

startingVoltage!(system::PowerSystem, analysis::ACPowerFlow)

The function extracts bus voltage magnitudes and angles from the PowerSystem composite type and assigns them to the ACPowerFlow abstract type, enabling users to initialize voltage values as required.

Updates

This function only updates the voltage field of the ACPowerFlow abstract type.

Abstract type

The abstract type ACPowerFlow can have the following subtypes:

NewtonRaphson: computes the bus voltages within the Newton-Raphson method;

FastNewtonRaphson: computes the bus voltages within the fast Newton-Raphson method;

GaussSeidel: computes the bus voltages within the Gauss-Seidel method.

Example

system = powerSystem("case14.h5")
acModel!(system)

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

updateBus!(system, analysis; label = 14, reactive = 0.13, magnitude = 1.2, angle = -0.17)

startingVoltage!(system, analysis)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end

Build DC Power Flow

JuliaGrid.dcPowerFlow — Function

dcPowerFlow(system::PowerSystem, factorization::Factorization)

The function sets up the framework to solve the DC power flow.

Arguments

The function requires the PowerSystem composite type to establish the framework. Moreover, the Factorization argument is optional and can be one of the following:

LU: solves the DC power flow problem in-place using LU factorization;

LDLt: solves the DC power flow problem using LDLt factorization;

QR: solves the DC power flow problem using QR factorization.

If the user does not provide the Factorization composite type, the default method for solving the DC power flow will be LU factorization.

Updates

If the DC model was not created, the function will automatically initiate an update of the dc field within the PowerSystem composite type. Additionally, if the slack bus lacks an in-service generator, JuliaGrid considers it a mistake and defines a new slack bus as the first generator bus with an in-service generator in the bus type list.

Returns

The function returns an instance of the DCPowerFlow composite type, which includes the following fields:

voltage: the variable allocated to store the bus voltage angles;

power: the variable allocated to store the active powers;

method: the factorized nodal matrix.

Examples

Establish the DC power flow framework that will be solved using the default LU factorization:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system)

Establish the DC power flow framework that will be solved using the QR factorization:

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system, QR)

Solve DC Power Flow

JuliaGrid.solve! — Method

solve!(system::PowerSystem, analysis::DCPowerFlow)

By computing the bus voltage angles, the function solves the DC power flow problem.

Updates

The resulting bus voltage angles are stored in the voltage field of the DCPowerFlow type.

Example

system = powerSystem("case14.h5")
dcModel!(system)

analysis = dcPowerFlow(system)
solve!(system, analysis)