Public Types

This section introduces the public types available in JuliaGrid. These types are designed to model and represent key components of power systems, measurements, and analyses.

To load public types into the current scope, utilize the following command:

using JuliaGrid

Power System Model

PowerSystem

Bus

Branch

Generator

BaseData

Model

Measurement Model

Measurement

Voltmeter

Ammeter

Wattmeter

Varmeter

PMU

Analysis

Analysis

AC

DC

LDLt

LU

QR

Power Flow

AcPowerFlow

NewtonRaphson

FastNewtonRaphson

GaussSeidel

DcPowerFlow

Optimal Power Flow

AcOptimalPowerFlow

DcOptimalPowerFlow

Observability Analysis

Island

PmuPlacement

State Estimation

AcStateEstimation

PmuStateEstimation

DcStateEstimation

GaussNewton

WLS

LAV

WlsMethod

Normal

Orthogonal

PetersWilkinson

Bad Data Analysis

ChiTest

ResidualTest

Power System Model

JuliaGrid.PowerSystem — Type

PowerSystem

A composite type constructed using the powerSystem function to store power system data.

Fields

bus::Bus: Bus-related data.

branch::Branch: Branch-related data.

generator::Generator: Generator-related data.

base::BaseData: Base power and base voltages.

model::Model: Data related to AC and DC analyses.

JuliaGrid.Bus — Type

Bus

A composite type used in PowerSystem to store bus-related data.

Fields

label::LabelDict: Bus labels.

demand::BusDemand: Active and reactive power demands.

supply::BusSupply: Active and reactive power supplies from generators.

shunt::BusShunt: Active and reactive power injections or demands from shunt elements.

voltage::BusVoltage: Initial voltages and voltage magnitude constraints.

layout::BusLayout: Bus layout, including bus types.

number::Int64: Total number of buses.

JuliaGrid.Branch — Type

Branch

A composite type used in PowerSystem to store branch-related data.

Fields

label::LabelDict: Branch labels.

parameter::BranchParameter: Data related to the π-model of the branches.

flow::BranchFlow: Branch flow constraints.

voltage::BranchVoltage: Voltage angle difference constraints.

layout::BranchLayout: Branch layout, including operational statuses.

number::Int64: Total number of branches.

JuliaGrid.Generator — Type

Generator

A composite type used in PowerSystem to store generator-related data.

Fields

label::LabelDict: Generator labels.

output::GeneratorOutput: Active and reactive power outputs.

capability::GeneratorCapability: Power output constraints.

ramping::GeneratorRamping: Ramp rate limits.

voltage::GeneratorVoltage: Voltage magnitude setpoints.

cost::GeneratorCost: Costs associated with active and reactive power outputs.

layout::GeneratorLayout: Generator layout, including operational statuses.

number::Int64: Total number of generators.

JuliaGrid.BaseData — Type

BaseData

A composite type used in PowerSystem to store base data.

Fields

power::BasePower: Base power.

voltage::BaseVoltage: Base voltages.

JuliaGrid.Model — Type

Model

A composite type used in PowerSystem to store vectors and matrices related to the power system's topology and parameters.

Fields

ac::AcModel: AC model, including the nodal admittance matrix and Y-parameters of two-port branches.

dc::DcModel: DC model, including the nodal admittance matrix and branch admittances.

Measurement Model

JuliaGrid.Measurement — Type

Measurement

A composite type built using the measurement function to store measurement data.

Fields

voltmeter::Voltmeter: Data related to bus voltage magnitude measurements.

ammeter::Ammeter: Data related to branch current magnitude measurements.

wattmeter::Wattmeter: Data related to active power injection and active power flow measurements.

varmeter::Varmeter: Data related to reactive power injection and reactive power flow measurements.

pmu::PMU: Data related to bus voltage and branch current phasor measurements.

system::PowerSystem: The reference to the power system.

JuliaGrid.Voltmeter — Type

Voltmeter

A composite type used in Measurement to store voltmeter-related data.

Fields

label::LabelDict: Voltmeter labels.

magnitude::GaussMeter: Bus voltage magnitude measurements.

layout::VoltmeterLayout: Placement indices and indicators.

number::Int64: Total number of voltmeters.

JuliaGrid.Ammeter — Type

Ammeter

A composite type used in Measurement to store ammeter-related data.

Fields

label::LabelDict: Ammeter labels.

magnitude::GaussMeter: Branch current magnitude measurements.

layout::AmmeterLayout: Placement indices and indicators.

number::Int64: Total number of ammeters.

JuliaGrid.Wattmeter — Type

Wattmeter

A composite type used in Measurement to store wattmeter-related data.

Fields

label::LabelDict: Wattmeter labels.

active::GaussMeter: Active power injection and active power flow measurements.

layout::PowermeterLayout: Placement indices and indicators.

number::Int64: Total number of wattmeters.

JuliaGrid.Varmeter — Type

Varmeter

A composite type used in Measurement to store varmeter-related data.

Fields

label::LabelDict: Varmeter labels.

reactive::GaussMeter: Reactive power injection and reactive power flow measurements.

layout::PowermeterLayout: Placement indices and indicators.

number::Int64: Total number of varmeters.

JuliaGrid.PMU — Type

PMU

A composite type used in Measurement to store PMU-related data.

Fields

label::LabelDict: PMU labels.

magnitude::GaussMeter: Bus voltage and branch current magnitude measurements.

angle::GaussMeter: Bus voltage and branch current angle measurements.

layout::PmuLayout: Placement indices and indicators.

number::Int64: Total number of PMUs.

Analysis

JuliaGrid.Analysis — Type

Analysis

An abstract type used for representing both AC and DC analyses in JuliaGrid.

JuliaGrid.AC — Type

AC <: Analysis

An abstract type representing AC analyses in JuliaGrid.

JuliaGrid.DC — Type

DC <: Analysis

An abstract type representing DC analyses in JuliaGrid.

JuliaGrid.LDLt — Type

LDLt

An abstract type used for representing LDLt factorization in JuliaGrid.

JuliaGrid.LU — Type

LU

An abstract type used for representing LU factorization in JuliaGrid.

JuliaGrid.QR — Type

QR

An abstract type used for representing QR factorization in JuliaGrid.

Power Flow

JuliaGrid.AcPowerFlow — Type

AcPowerFlow{T <: Union{NewtonRaphson, FastNewtonRaphson, GaussSeidel}} <: AC

A composite type representing an AC power flow model, where the type parameter T specifies the numerical method used to solve the power flow. Supported methods include NewtonRaphson, FastNewtonRaphson, and GaussSeidel.

Fields

voltage::Polar: Bus voltages represented in polar form.

power::AcPower: Active and reactive powers at the buses, branches, and generators.

current::AcCurrent: Currents at the buses and branches.

method::T: Vectors and matrices associated with the method used to solve the AC power flow.

system::PowerSystem: The reference to the power system.

JuliaGrid.NewtonRaphson — Type

NewtonRaphson

A composite type built using the newtonRaphson function to define the AC power flow model, which will be solved using the Newton-Raphson method.

Fields

jacobian::SparseMatrixCSC{Float64, Int64}: Jacobian matrix.

mismatch::Vector{Float64}: Vector of mismatches.

increment::Vector{Float64}: Vector of state variable increments.

factorization::Factorization: Factorization of the Jacobian matrix.

pq::Vector{Int64}: Indices related to demand buses.

pvpq::Vector{Int64}: Indices related to demand and generator buses.

signature::Signature: Tracks modifications in key variables.

iteration::Int64: The iteration counter.

JuliaGrid.FastNewtonRaphson — Type

FastNewtonRaphson

A composite type built using the fastNewtonRaphsonBX and fastNewtonRaphsonXB functions to define the AC power flow model, which will be solved using the fast Newton-Raphson method.

Fields

active:FastNewtonRaphsonModel: Jacobian, mismatches, and incrementes for active power equations.

reactive:FastNewtonRaphsonModel: Jacobian, mismatches, and incrementes for active power equations.

pq::Vector{Int64}: Indices related to demand buses.

pvpq::Vector{Int64}: Indices related to demand and generator buses.

signature::Signature: Tracks modifications in key variables.

bx::Bool: Version of the method, either BX or XB.

iteration::Int64: The iteration counter.

JuliaGrid.GaussSeidel — Type

GaussSeidel

A composite type built using the gaussSeidel function to define the AC power flow model, which will be solved using the Gauss-Seidel method.

Fields

voltage::Vector{ComplexF64}: Vector of complex voltage values.

pq::Vector{Int64}: Indices related to demand buses.

pv::Vector{Int64}: Indices related to generator buses.

signature::Signature: Tracks modifications in key variables.

iteration::Int64: The iteration counter.

JuliaGrid.DcPowerFlow — Type

DcPowerFlow <: DC

A composite type built using the dcPowerFlow function to define the DC power flow model.

Fields

voltage::Angle: Bus voltage angles.

power::DcPower: Active powers at the buses, branches, and generators.

method::DcPowerFlowMethod: Factorization of the nodal admittance matrix.

system::PowerSystem: The reference to the power system.

Optimal Power Flow

JuliaGrid.AcOptimalPowerFlow — Type

AcOptimalPowerFlow <: AC

A composite type built using the acOptimalPowerFlow function to define the AC optimal power flow model.

Fields

voltage::Polar: Bus voltages represented in polar form.

power::AcPower: Active and reactive powers at buses, branches, and generators.

current::AcCurrent: Currents at buses and branches.

method::AcOptimalPowerFlowMethod: The JuMP model, including variables, constraints, and objective.

system::PowerSystem: The reference to the power system.

JuliaGrid.DcOptimalPowerFlow — Type

DcOptimalPowerFlow <: DC

A composite type built using the dcOptimalPowerFlow function to define the DC optimal power flow model.

Fields

voltage::Angle: Bus voltage angles.

power::DcPower: Active powers at buses, branches, and generators.

method::DcOptimalPowerFlowMethod: The JuMP model, including variables, constraints, and objective.

system::PowerSystem: The reference to the power system.

Observability Analysis

JuliaGrid.Island — Type

Island

A composite type built using the islandTopologicalFlow and islandTopological functions, which holds data about observable islands.

Fields

island::Vector{Vector{Int64}}: List of observable islands, represented by a vector of bus indices.

bus::Vector{Int64}: Positions of buses in relation to each island.

tie::TieData: Tie data associated with buses and branches.

JuliaGrid.PmuPlacement — Type

PmuPlacement

A composite type built using the pmuPlacement function, which stores data on optimal PMU placement.

Fields

bus::LabelDict: Phasor measurement placement at buses.

from::LabelDict: Phasor measurement placement at from-buses.

to::LabelDict: Phasor measurement placement at to-buses.

State Estimation

JuliaGrid.AcStateEstimation — Type

AcStateEstimation{T <: Union{GaussNewton, LAV}} <: AC

A composite type representing an AC state estimation model, where the type parameter T specifies the estimation method. Supported methods include GaussNewton and LAV. The model is constructed using either the gaussNewton or acLavStateEstimation function.

Fields

voltage::Polar: Bus voltages represented in polar form.

power::AcPower: Active and reactive powers at the buses and branches.

current::AcCurrent: Currents at the buses and branches.

method::T: The estimation model associated with the method used to solve the AC state estimation.

system::PowerSystem: The reference to the power system.

monitoring::Measurement: The reference to the measurement model.

JuliaGrid.PmuStateEstimation — Type

PmuStateEstimation{T <: Union{WLS, LAV}} <: AC

A composite type representing a PMU state estimation model, where the type parameter T specifies the estimation method. Supported methods include WLS and LAV. The model is constructed using either the pmuStateEstimation or pmuLavStateEstimation function.

Fields

voltage::Polar: Bus voltages represented in polar form.

power::AcPower: Active and reactive powers at the buses and branches.

current::AcCurrent: Currents at the buses and branches.

method::T: The estimation model associated with the method used to solve the PMU state estimation.

system::PowerSystem: The reference to the power system.

monitoring::Measurement: The reference to the measurement model.

JuliaGrid.DcStateEstimation — Type

DcStateEstimation{T <: Union{WLS, LAV}} <: DC

A composite type representing a DC state estimation model, where the type parameter T specifies the estimation method. Supported methods include WLS and LAV. The model is constructed using either the dcStateEstimation or dcLavStateEstimation function.

Fields

voltage::Angle: Bus voltage angles.

power::DcPower: Active powers at the buses and generators.

method::T: The estimation model associated with the method used to solve the DC state estimation.

system::PowerSystem: The reference to the power system.

monitoring::Measurement: The reference to the measurement model.

JuliaGrid.GaussNewton — Type

GaussNewton{T <: WlsMethod}

A composite type built using the gaussNewton function to define the AC state estimation model, which will be solved using the Gauss-Newton method.

Fields

jacobian::SparseMatrixCSC{Float64, Int64}: Jacobian matrix.

precision::SparseMatrixCSC{Float64, Int64}: Precision matrix.

mean::Vector{Float64}: Mean vector.

residual::Vector{Float64}: Residual vector.

increment::Vector{Float64}: Increment vector.

factorization::Factorization: Factorization of the Jacobian matrix.

type::Vector{Int8}: Indicators of measurement types.

index::Vector{Int64}: Indices of buses and branches where measurements are located.

range::Vector{Int64}: Range of measurement devices.

signature::Signature: Tracks modifications in key variables.

objective::Float64: Value of the objective function.

iteration::Int64: The iteration counter.

JuliaGrid.WLS — Type

WLS{T <: WlsMethod}

A composite type representing a linear weighted least-squares state estimation model.

Fields

coefficient::SparseMatrixCSC{Float64, Int64}: Coefficient matrix.

precision::SparseMatrixCSC{Float64, Int64}: Precision matrix.

mean::Vector{Float64}: Mean vector.

factorization::Factorization: Factorization of the coefficient matrix.

index::OrderedDict{Int64, Int64}: Indices if needed.

number::Int64: Number of measurement devices.

inservice::Int64: Number of equations related to in-service measurement devices.

signature::Signature: Tracks modifications in key variables.

JuliaGrid.LAV — Type

LAV

A composite type representing a least absolute value state estimation model.

Fields

jump::JuMP.Model: The JuMP optimization model.

variable::LavVariableRef: References to state variables and positive and negative deviations.

residual::Dict{Int64, ConstraintRef}: References to residual constraints.

index::OrderedDict{Int64, Int64}: Mapping of indices, if needed.

range::Vector{Int64}: Range of measurement devices.

number::Int64: Total number of measurement devices.

JuliaGrid.WlsMethod — Type

WlsMethod

An abstract type representing a method for solving WLS state estimation.

JuliaGrid.Normal — Type

Normal <: WlsMethod

An abstract type representing weighted least-squares state estimation, where the normal equation is solved by factorizing the gain matrix and performing forward/backward substitutions on the right-hand-side vector. It is used as a type parameter in GaussNewton and WLS models.

JuliaGrid.Orthogonal — Type

Orthogonal <: WlsMethod

An abstract type representing weighted least-squares state estimation, where the normal equation is solved using an orthogonal method. It is used as a type parameter in GaussNewton and WLS models.

JuliaGrid.PetersWilkinson — Type

PetersWilkinson <: WlsMethod

An abstract type representing weighted least-squares state estimation, where the normal equation is solved using an Peters and Wilkinson method. It is used as a type parameter in GaussNewton and WLS models.

Bad Data Analysis

JuliaGrid.ChiTest — Type

ChiTest

A composite type constructed using the chiTest function, which stores results from the Chi-squared bad data detection test.

Fields

detect::Bool: Flag indicating bad data detection.

threshold::Float64: Chi-squared critical value.

objective::Float64: Objective function value from WLS state estimation.

JuliaGrid.ResidualTest — Type

ResidualTest

A composite type built using residualTest! function, which stores results from the bad data processing.

Fields

detect::Bool: Flag indicating bad data detection.

maxNormalizedResidual::Float64: The maximum value of the normalized residual.

label::IntStr: The label of the measurement suspected to be an outlier.

index::Int64: The index of the outlier measurement within the model.