Setup and Print

For further information on this topic, please see the Power System Model or Measurement Model sections of the Manual. Please note that when using macros, they modify variables within the current scope. Print functions can be used to print results to the REPL, or users can redirect output to print results to a text file, for example.

Base Units

@base

Input Units

@power

@voltage

@current

@parameter

Default Settings

@default

Print Data

printBusData

printBranchData

printGeneratorData

Print Summary

printBusSummary

printBranchSummary

printGeneratorSummary

To load power system model API functionalities into the current scope, utilize the following command:

using JuliaGrid

Base Units

JuliaGrid.@base — Macro

@base(system::PowerSystem, power, voltage)

By default, the units for base power and base voltages are set to volt-ampere (VA) and volt (V), but you can modify the prefixes using the macro.

Prefixes must be specified according to the SI prefixes and should be included with the unit of power (VA) or unit of voltage (V). Keep in mind that the macro must be used after creating the composite type PowerSystem.

Example

system = powerSystem("case14.h5")
@base(system, MVA, kV)

Input Units

JuliaGrid.@power — Macro

@power(active, reactive, apparent)

JuliaGrid stores all data related with powers in per-units, and these cannot be altered. However, the power units of the built-in functions used to add or modified power system elements can be modified using the macro.

Prefixes must be specified according to the SI prefixes and should be included with the unit of active power (W), reactive power (VAr), or apparent power (VA). Also, it is a possible to combine SI units with/without prefixes with per-units (pu).

Changing the unit of active power is reflected in the following quantities:

addBus!, updateBus!, @bus: active, conductance;

addBranch!, updateBranch!, @branch: if type = 2: longTerm, shortTerm, emergency;

addGenerator!, updateGenerator!, @generator: active, minActive, maxActive, lowActive, upActive, loadFollowing, reserve10min, reserve30min;

cost!: if active: piecewise, polynomial;

addWattmeter!, updateWattmeter!: active, variance;

@wattmeter: , varianceBus, varianceFrom, varianceTo.

Changing the unit of reactive power unit is reflected in the following quantities:

addBus!, updateBus!, @bus: reactive, susceptance;

addGenerator!, updateGenerator!, @generator: reactive, minReactive, maxReactive, minLowReactive, maxLowReactive, minUpReactive, maxUpReactive, reactiveTimescale;

cost!: if reactive: piecewise, polynomial;

addVarmeter!, updateVarmeter!: reactive, variance;

@varmeter: varianceBus, varianceFrom, varianceTo.

Changing the unit of apparent power unit is reflected in the following quantities:

addBranch!, updateBranch!, @branch: if type = 1 or type = 3: longTerm, shortTerm, emergency.

Example

@power(MW, kVAr, VA)

JuliaGrid.@voltage — Macro

@voltage(magnitude, angle, base)

JuliaGrid stores all data related with voltages in per-units and radians, and these cannot be altered. However, the voltage magnitude and angle units of the built-in functions used to add or modified power system elements can be modified using the macro.

The prefixes must adhere to the SI prefixes and should be specified along with the unit of voltage, either magnitude (V) or base (V). Alternatively, the unit of voltage magnitude can be expressed in per-unit (pu). The unit of voltage angle should be in radians (rad) or degrees (deg).

Changing the unit of voltage magnitude is reflected in the following quantities:

addBus!, updateBus!, @bus: magnitude, minMagnitude, maxMagnitude;

addGenerator!, updateGenerator!, @generator: magnitude;

addVoltmeter!, updateVoltmeter!, @voltmeter: magnitude, variance;

addPmu!, updatePmu!: if bus: magnitude, varianceMagnitude;

@pmu: varianceMagnitudeBus.

Changing the unit of voltage angle is reflected in the following quantities:

addBus!, updateBus!, @bus: angle;

addBranch!, updateBranch!, @branch: shiftAngle, minDiffAngle, maxDiffAngle;

addPmu!, updatePmu!: if bus: angle, varianceAngle;

@pmu: varianceAngleBus.

Changing the unit prefix of voltage base is reflected in the following quantity:

addBus!, updateBus!, @bus: base.

Example

@voltage(pu, deg, kV)

JuliaGrid.@current — Macro

@current(magnitude, angle)

JuliaGrid stores all data related with currents in per-units and radians, and these cannot be altered. However, the current magnitude and angle units of the built-in functions used to add or modified measurement devices can be modified using the macro.

The prefixes must adhere to the SI prefixes and should be specified along with the unit of current magnitude (V). Alternatively, the unit of current magnitude can be expressed in per-unit (pu). The unit of current angle should be in radians (rad) or degrees (deg).

Changing the unit of current magnitude is reflected in the following quantities:

addAmmeter!, updateAmmeter!: magnitude, variance;

@ammeter: varianceFrom, varianceTo;

addPmu!, updatePmu!: if from or to: magnitude, varianceMagnitude;

@pmu: varianceMagnitudeFrom, varianceMagnitudeTo.

Changing the unit of current angle is reflected in the following quantities:

addPmu!, updatePmu!: if from or to: angle, varianceAngle;

@pmu: varianceAngleFrom, varianceAngleTo.

Example

@current(pu, deg)

JuliaGrid.@parameter — Macro

@parameter(impedance, admittance)

JuliaGrid stores all data related with impedances and admittancies in per-units, and these cannot be altered. However, units of impedance and admittance of the built-in functions used to add or modified power system elements can be modified using the macro.

Prefixes must be specified according to the SI prefixes and should be included with the unit of impedance (Ω) or unit of admittance (S). The second option is to define the units in per-unit (pu).

In the case where impedance and admittance are being used in SI units (Ω and S) and these units are related to the transformer, the assignment must be based on the primary side of the transformer.

Changing the units of impedance is reflected in the following quantities in specific functions:

addBranch!, updateBranch!, @branch: resistance, reactance.

Changing the units of admittance is reflected in the following quantities:

addBranch!, updateBranch!, @branch: conductance, susceptance.

Example

@parameter(Ω, pu)

Default Settings

JuliaGrid.@default — Macro

@default(mode)

The macro is designed to reset various settings to their default values.

The mode argument can take on the following values:

unit: Resets all units to their default settings.

power: Sets active, reactive, and apparent power to per-units.

voltage: Sets voltage magnitude to per-unit and voltage angle to radian.

parameter: Sets impedance and admittance to per-units.

template: Resets bus, branch, generator, voltmeter, ammeter, wattmeter, varmeter, and pmu templates to their default settings.

bus: Resets the bus template to its default settings.

branch: Resets the branch template to its default settings.

generator: Resets the generator template to its default settings.

voltmeter: Resets the voltmeter template to its default settings.

ammeter: Resets the ammeter template to its default settings.

wattmeter: Resets the wattmeter template to its default settings.

varmeter: Resets the varmeter template to its default settings.

pmu: Resets the pmu template to its default settings.

Example

@default(unit)

Print Data

JuliaGrid.printBusData — Function

printBusData(system::PowerSystem, analysis::Analysis, io::IO)

The function prints voltages, powers, and currents related to buses. Optionally, an IO may be passed as the last argument to redirect output.

Julia 1.10

The function printBusData requires Julia 1.10 or later.

Example

system = powerSystem("case14.h5")

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
power!(system, analysis)

printBusData(system, analysis)

JuliaGrid.printBranchData — Function

printBranchData(system::PowerSystem, analysis::Analysis, io::IO)

The function prints powers and currents related to branches. Optionally, an IO may be passed as the last argument to redirect output.

Julia 1.10

The function printBranchData requires Julia 1.10 or later.

Example

system = powerSystem("case14.h5")

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
power!(system, analysis)

printBranchData(system, analysis)

JuliaGrid.printGeneratorData — Function

printGeneratorData(system::PowerSystem, analysis::Analysis, io::IO)

The function prints powers related to generators. Optionally, an IO may be passed as the last argument to redirect output.

Julia 1.10

The function printGeneratorData requires Julia 1.10 or later.

Example

system = powerSystem("case14.h5")

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
power!(system, analysis)

printGeneratorData(system, analysis)

Print Summary

JuliaGrid.printBusSummary — Function

printBusSummary(system::PowerSystem, analysis::Analysis, io::IO)

The function prints a summary of the electrical quantities related to buses. Optionally, an IO may be passed as the last argument to redirect output.

Julia 1.10

The function printBusSummary requires Julia 1.10 or later.

Example

system = powerSystem("case14.h5")

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
power!(system, analysis)

printBusSummary(system, analysis)

JuliaGrid.printBranchSummary — Function

printBranchSummary(system::PowerSystem, analysis::Analysis, io::IO)

The function prints a summary of the electrical quantities related to branches. Optionally, an IO may be passed as the last argument to redirect output.

Julia 1.10

The function printBranchSummary requires Julia 1.10 or later.

Example

system = powerSystem("case14.h5")

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
power!(system, analysis)

printBranchSummary(system, analysis)

JuliaGrid.printGeneratorSummary — Function

printGeneratorSummary(system::PowerSystem, analysis::Analysis, io::IO)

The function prints a summary of the electrical quantities related to generators. Optionally, an IO may be passed as the last argument to redirect output.

Julia 1.10

The function printGeneratorSummary requires Julia 1.10 or later.

Example

system = powerSystem("case14.h5")

analysis = newtonRaphson(system)
for i = 1:10
    stopping = mismatch!(system, analysis)
    if all(stopping .< 1e-8)
        break
    end
    solve!(system, analysis)
end
power!(system, analysis)

printGeneratorSummary(system, analysis)