State Estimation

For further information on this topic, please see the AC State Estimation or DC State Estimation sections of the Manual. Below, we have provided a list of functions that can be utilized for power flow analysis.

Build DC State Estimation

dcStateEstimation

Solve DC State Estimation

solve!

Bad Data Analysis

badData!

Build DC State Estimation

JuliaGrid.dcStateEstimation — Function

dcStateEstimation(system::PowerSystem, device::Measurement, method)

The function sets up the framework to solve the DC state estimation.

Arguments

This function requires the PowerSystem and Measurement composite types to establish the framework.

Additionally, the inclusion of the method parameter is not obligatory. It offers a range of techniques for resolving DC state estimation. To tackle the weighted least-squares (WLS) issue, users have the option to employ standard factorization methods like LU or LDLt, but only when the gain matrix is symmetrical. For a more resilient solution in situations involving ill-conditioned data, particularly when significant differences in variances exist, choosing the QR factorization method is recommended.

Moreover, instead of solving the WLS state estimation problem, users can utilize the least absolute value (LAV) method to find an estimator. This can be achieved by selecting one of the optimization solvers, where typically Ipopt.Optimizer suffices for common scenarios.

If the user does not provide the method, the default method for solving the DC estimation will be LU factorization.

Updates

If the DC model was not created, the function will automatically initiate an update of the dc field within the PowerSystem composite type. Additionally, if the slack bus lacks an in-service generator, JuliaGrid considers it a mistake and defines a new slack bus as the first generator bus with an in-service generator in the bus type list.

Returns

The function returns an instance of the DCStateEstimation abstract type, which includes the following fields:

voltage: the variable allocated to store the bus voltage angles;

power: the variable allocated to store the active powers;

method: the system model vectors and matrices, or alternatively, the optimization model;

bad: the variable linked to identifying bad data within the measurement set.

Examples

Establish the DC state estimation WLS framework that will be solved using the default LU factorization:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device)

Establish the DC state estimation WLS framework that will be solved using the QR factorization:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device, QR)

Establish the DC state estimation LAV framework:

using Ipopt

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device, Ipopt.Optimizer)

Solve DC State Estimation

JuliaGrid.solve! — Method

solve!(system::PowerSystem, analysis::DCStateEstimation)

By computing the bus voltage angles, the function solves the DC state estimation problem.

Updates

The resulting bus voltage angles are stored in the voltage field of the DCStateEstimation type.

Example

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device)
solve!(system, analysis)

Bad Data Analysis

JuliaGrid.badData! — Function

badData!(system::PowerSystem, device::Measurement, analysis::DCStateEstimation; threshold)

The function conducts bad data detection and identification using the largest normalized residual test, subsequently removing measurement outliers from the measurement set. It can be executed after obtaining estimation solutions.

Arguments

This function necessitates the composite types PowerSystem, Measurement, and DCStateEstimation to detect and identify bad data.

Keyword

The keyword threshold establishes the identification threshold. If the largest normalized residual surpasses this threshold, the measurement is flagged as bad data. The default threshold value is set to threshold = 3.0.

Updates

In case bad data is detected, the function removes measurements from the jacobian matrix, weight, and mean vectors within the DCStateEstimation type. Additionally, it marks the respective measurement within the Measurement type as out-of-service.

Furthermore, the variable bad within the DCStateEstimation type stores information regarding bad data detection and identification:

detect: returns true after the function's execution if bad data is detected;

maxNormalizedResidual: denotes the value of the largest normalized residual;

index: represents the index of the bad data associated with the DCStateEstimation type;

label: signifies the label of the bad data.

Example

Obtaining the solution after detecting and removing bad data:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device)
solve!(system, analysis)

badData!(system, device, analysis; threshold = 4.0)
solve!(system, analysis)

Obtaining the solution while bad data is detected:

system = powerSystem("case14.h5")
device = measurement("measurement14.h5")

analysis = dcStateEstimation(system, device)

while analysis.bad.detect
    solve!(system, analysis)
    badData!(system, device, analysis; threshold = 4.0)
end