Power Flow Analysis

Once the bus voltages are obtained, it is possible to calculate powers and currents related to buses and branches, and powers related to generators:

bus!()

branch!()

generator!().

Note that the JuliaGrid stores complex currents in the polar coordinate system, while complex powers are stored in the rectangle coordinate system.

Bus

JuliaGrid.bus! — Function

The function computes powers and currents related to buses.

bus!(system::PowerSystem, result::Result)

The function updates the field result.bus.

AC Power Flow Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:10
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

bus!(system, result)

DC Power Flow Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)
bus!(system, result)

Branch

JuliaGrid.branch! — Function

The function computes powers and currents related to branches for the AC power flow analysis. For the DC power flow analysis, the function computes only active powers.

branch!(system::PowerSystem, result::Result)

The function updates the field result.branch.

AC Power Flow Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:10
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

branch!(system, result)

DC Power Flow Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)
branch!(system, result)

Generator

JuliaGrid.generator! — Function

The function computes powers related to generators.

generator!(system::PowerSystem, result::Result)

The function updates the field result.generator.

AC Power Flow Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:10
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

generator!(system, result)

DC Power Flow Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)
generator!(system, result)