Power Flow Analysis

The power flow analysis requires the main composite type PowerSystem with fields bus, branch, generator. In addition, depending on whether AC or DC power flow analysis is used, acModel or dcModel is required.

JuliaGrid stores results in the main composite type Result with fields:

bus

branch

generator

algorithm

Once the main composite type PowerSystem is created, it is possible to create main composite type Result. The composite type Result in the DC power flow analysis is created when determining the bus voltages using the function dcPowerFlow(). In contrast, the AC power flow analysis first requires the initialization of the iterative method, during which the composite type Result is created:

newtonRaphson()

fastNewtonRaphsonBX()

fastNewtonRaphsonXB()

gaussSeidel().

The calculation of the bus voltages, depending on the type of analysis and the selected method, can be performed using one of the functions:

newtonRaphson!()

fastNewtonRaphson!()

gaussSeidel!()

dcPowerFlow()

Note that the methods for solving AC power flow problem should be called inside a loop, thus simulating an iterative process.

Then, it is possible to calculate other quantities of interest using functions:

bus!()

branch!()

generator!()

Gauss-Seidel Method

Functions receive the composite type PowerSystem.

JuliaGrid.newtonRaphson — Function

The function initializes the Newthon-Raphson method.

newtonRaphson(system::PowerSystem)

The function affects the field algorithm of the type Result.

JuliaGrid.newtonRaphson! — Function

The function solves the AC power flow problem using the Newthon-Raphson method.

newtonRaphson!(system::PowerSystem, result::Result)

The function affects fields bus.voltage and algorithm of the type Result.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.stopping
for i = 1:10
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

<!– –-

Power Flow Solution

The functions receive the composite types PowerSystem and Result.

JuliaGrid.gaussSeidel! — Function

The function solves the AC power flow problem using the Gauss-Seidel method.

gaussSeidel!(system::PowerSystem, result::Result)

The function affects fields bus.voltage and algorithm of the type Result.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = gaussSeidel(system)
for i = 1:1000
    gaussSeidel!(system, result)
    if result.algorithm.stopping.active < 1e-8 && result.algorithm.stopping.reactive < 1e-8
        break
    end
end

Missing docstring.

Missing docstring for newtonRaphson!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fastNewtonRaphson!. Check Documenter's build log for details.

JuliaGrid.dcPowerFlow — Function

The function solves the DC power flow problemnby determining the bus voltage angles, and returns the composite type Result.

dcPowerFlow(system::PowerSystem)

The function affects field result.bus.voltage.angle and algorithm of the type Result.

Example

result = dcPowerFlow(system)

–>