Power Flow Analysis

The power flow analysis requires the main composite type PowerSystem with fields bus, branch, generator. In addition, depending on whether AC or DC power flow analysis is used, acModel or dcModel is required.

JuliaGrid stores results in the main composite type Result with fields:

bus

branch

generator

algorithm

Once the main composite type PowerSystem is created, it is possible to create main composite type Result. The composite type Result in the DC power flow analysis is created when determining the bus voltages using the function dcPowerFlow(). In contrast, the AC power flow analysis first requires the initialization of the iterative method, during which the composite type Result is created:

newtonRaphson()

fastNewtonRaphsonBX()

fastNewtonRaphsonXB()

gaussSeidel().

The calculation of the bus voltages, depending on the type of analysis and the selected method, can be performed using one of the functions:

newtonRaphson!()

fastNewtonRaphson!()

gaussSeidel!()

dcPowerFlow()

Note that the methods for solving AC power flow problem should be called inside a loop, thus simulating an iterative process.

Then, it is possible to calculate other quantities of interest using functions:

bus!()

branch!()

generator!()

Newton-Raphson Method

Functions receive the composite type PowerSystem.

JuliaGrid.newtonRaphson — Function

The function initializes the Newthon-Raphson method, and returns the composite type Result.

newtonRaphson(system::PowerSystem)

The function affects the field algorithm of the composite type Result.

JuliaGrid.newtonRaphson! — Function

The function solves the AC power flow problem using the Newthon-Raphson method.

newtonRaphson!(system::PowerSystem, result::Result)

The function affects fields bus.voltage and algorithm of the composite type Result.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.stopping
for i = 1:10
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

Fast Newton-Raphson Method

Functions receive the composite type PowerSystem.

JuliaGrid.fastNewtonRaphsonBX — Function

The function initializes the fast Newthon-Raphson BX method, and returns the composite type Result.

fastNewtonRaphsonBX(system::PowerSystem)

The function affects the field algorithm of the composite type Result.

JuliaGrid.fastNewtonRaphsonXB — Function

The function initializes the fast Newthon-Raphson XB method, and returns the composite type Result.

fastNewtonRaphsonXB(system::PowerSystem)

The function affects the field algorithm of the composite type Result.

Missing docstring.

Missing docstring for fastNewtonRaphson!. Check Documenter's build log for details.

Gauss-Seidel Method

Functions receive the composite type PowerSystem.

JuliaGrid.gaussSeidel — Function

The function initializes the Gauss-Seidel method, and returns the composite type Result.

gaussSeidel(system::PowerSystem)

The function affects the field algorithm of the composite type Result.

JuliaGrid.gaussSeidel! — Function

The function solves the AC power flow problem using the Gauss-Seidel method.

gaussSeidel!(system::PowerSystem, result::Result)

The function affects fields bus.voltage and algorithm of the type Result.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = gaussSeidel(system)
stopping = result.algorithm.stopping
for i = 1:1000
    gaussSeidel!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

DC Power Flow Solution

Function receives the composite type PowerSystem, and returns the composite type Result.

JuliaGrid.dcPowerFlow — Function

The function solves the DC power flow problem determining the bus voltage angles, and returns the composite type Result.

dcPowerFlow(system::PowerSystem)

The function affects field result.bus.voltage.angle and algorithm of the type Result.

Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)