Power Flow Solution

The solution of the power flow implies the calculation of the bus voltages. To obtain a solution, the framework requires the composite type PowerSystem with fields bus, branch, and generator. In addition, depending on whether AC or DC power flow is used, acModel or dcModel is required.

JuliaGrid stores results in the composite type Result with fields:

bus

branch

generator

algorithm.

Once the composite type PowerSystem is created, it is possible to create the composite type Result. The composite type Result in the DC power flow is created when determining the bus voltage angles using the function dcPowerFlow(). In contrast, the AC power flow requires the initialization of the iterative method, during which the composite type Result is created:

newtonRaphson()

fastNewtonRaphsonBX()

fastNewtonRaphsonXB()

gaussSeidel().

The calculation of the bus voltages, depending on the type of power flow and the selected method, can be performed using one of the functions:

newtonRaphson!()

fastNewtonRaphson!()

gaussSeidel!()

dcPowerFlow().

Note that methods for solving the AC power flow problem should be called inside a loop, thus simulating an iterative process.

In addition, JuliaGrid has an additional set of functions for power flow analysis, which includes determining the powers and currents related to buses, branches or generators:

bus!()

branch!()

generator!().

JuliaGrid also provides the function that checks reactive power limits of the generators, once the solution of the AC power flow is obtained:

reactivePowerLimit!()

Newton-Raphson Method

JuliaGrid.newtonRaphson — Function

The function receives the composite type PowerSystem, initializes the Newton-Raphson method, and returns the composite type Result.

newtonRaphson(system::PowerSystem)

The function updates the field algorithm of the composite type Result.

JuliaGrid.newtonRaphson! — Function

The function receives the composite type PowerSystem and Result. Using the Newton-Raphson method, the function computes bus voltage magnitudes and angles.

newtonRaphson!(system::PowerSystem, result::Result)

The function updates fields result.bus.voltage and result.algorithm.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = newtonRaphson(system)
stopping = result.algorithm.iteration.stopping
for i = 1:10
    newtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

Fast Newton-Raphson Method

JuliaGrid.fastNewtonRaphsonBX — Function

The function receives the composite type PowerSystem, initializes the fast Newton-Raphson method of version BX, and returns the composite type Result.

fastNewtonRaphsonBX(system::PowerSystem)

The function updates the field algorithm of the composite type Result.

JuliaGrid.fastNewtonRaphsonXB — Function

The function receives the composite type PowerSystem, initializes the fast Newton-Raphson method of version XB, and returns the composite type Result.

fastNewtonRaphsonXB(system::PowerSystem)

The function updates the field algorithm of the composite type Result.

JuliaGrid.fastNewtonRaphson! — Function

The function receives the composite type PowerSystem and Result. Using the fast Newton-Raphson method, the function computes bus voltage magnitudes and angles.

fastNewtonRaphson!(system::PowerSystem, result::Result)

The function updates fields result.bus.voltage and result.algorithm.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = fastNewtonRaphsonBX(system)
stopping = result.algorithm.iteration.stopping
for i = 1:100
    fastNewtonRaphson!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

Gauss-Seidel Method

JuliaGrid.gaussSeidel — Function

The function receives the composite type PowerSystem, initializes the Gauss-Seidel method, and returns the composite type Result.

gaussSeidel(system::PowerSystem)

The function updates the field algorithm of the composite type Result.

JuliaGrid.gaussSeidel! — Function

The function receives the composite type PowerSystem and Result. Using the Gauss-Seidel method, the function computes bus voltage magnitudes and angles.

gaussSeidel!(system::PowerSystem, result::Result)

The function updates fields result.bus.voltage and result.algorithm.

Example

system = powerSystem("case14.h5")
acModel!(system)

result = gaussSeidel(system)
stopping = result.algorithm.iteration.stopping
for i = 1:1000
    gaussSeidel!(system, result)
    if stopping.active < 1e-8 && stopping.reactive < 1e-8
        break
    end
end

DC Power Flow Solution

JuliaGrid.dcPowerFlow — Function

The function receives the composite type PowerSystem, solves the DC power flow problem by determining bus voltage angles, and returns the composite type Result.

dcPowerFlow(system::PowerSystem)

The function updates fields result.bus.voltage.angle and result.algorithm.

Example

system = powerSystem("case14.h5")
dcModel!(system)

result = dcPowerFlow(system)