Build Power System Model

The composite type PowerSystem with fields bus, branch, generator, acModel, dcModel, and basePower can be created using a function:

powerSystem().

Once the model is created, it is possible to add buses, branches and generators using the functions:

addBus!()

addBranch!()

addGenerator!().

In addition, it is possible to manipulate the parameters of buses, branches and generators using functions:

shuntBus!()

statusBranch!()

parameterBranch!()

statusGenerator!()

outputGenerator!().

The execution of these functions will automatically trigger the update of all fields affected by these functions. Changing other parameters of the power system can be done by changing variables by accessing their values in fields bus, branch, generator, and basePower of the composite type PowerSystem.

At any moment after the composite type PowerSystem is formed, it is possible to save the data in HDF5 format that describe the power system:

savePowerSystem().

To create vectors and matrices obtained based on the power system topology and parameters, the following functions are used:

acModel!()

dcModel!().

Note that, functions can be executed at any time, and when all power system buses are defined. More precisely, once the field acModel and dcModel are formed, using function addBranch!(), will automatically trigger the update of these fields. In contrast, adding a new bus, using addBus!(), requires executing the functions acModel!() and dcModel!() again. Further, the execution of the functions related to parameter manipulation of buses and branches will automatically trigger the update of the fields acModel and dcModel.

Build Model

The function builds the composite type PowerSystem and populates fields bus, branch, generator and basePower.

JuliaGrid.powerSystem — Function

The path to the HDF5 file with the .h5 extension should be passed to the function:

powerSystem("pathToExternalData/name.h5")

Similarly, the path to the Matpower file with the .m extension should be passed to the same function:

powerSystem("pathToExternalData/name.m")

Ignoring the function argument initializes the composite type PowerSystem, which enables building the model from scratch:

powerSystem()

Once the composite type PowerSystem is created it is possible to add new buses, branches or generators, and also change the parameters of the existing ones.

Bus Functions

Functions receive the composite type PowerSystem and arguments by keyword to set or change bus parameters and update the field bus.

JuliaGrid.addBus! — Function

The function adds a new bus and updates the field bus. Names, descriptions and units of keywords are given in the table bus group.

addBus!(system::PowerSystem; label, slackLabel, area, lossZone, active, reactive,
    conductance, susceptance, magnitude, angle, minMagnitude, maxMagnitude, base)

The keyword label is mandatory. Default keyword values are set to zero, except for keywords lossZone = 1, area = 1, magnitude = 1.0, minMagnitude = 0.9, and maxMagnitude = 1.1.

The slack bus, using the keyword slackLabel, can be specified in each function call with the label of the bus being defined or already existing. If the bus is not defined as slack, the function addBus!() automatically defines the bus as the demand bus (PQ). If a generator is connected to a bus, using the function addGenerator!(), the bus becomes a generator bus (PV).

Example

system = powerSystem()

addBus!(system; label = 1, slackLabel = 1, active = 0.25, reactive = -0.04)

JuliaGrid.shuntBus! — Function

The function allows changing conductance and susceptance parameters of the shunt element connected to the bus.

shuntBus!(system::PowerSystem; label, conductance, susceptance)

The keyword label should correspond to the already defined bus label. Keywords conductance or susceptance can be omitted, then the value of the omitted parameter remains unchanged. The function also updates the field acModel, if field exist.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
shuntBus!(system; label = 1, conductance = 0.04)

Branch Functions

Functions receive the composite type PowerSystem and arguments by keyword to set or change branch parameters. Further, functions update the field branch, but also fields acModel and dcModel. More precisely, once acModel and dcModel are created, the execution of functions will automatically trigger the update of these fields.

JuliaGrid.addBranch! — Function

The function adds a new branch. Names, descriptions and units of keywords are given in the table branch group. A branch can be added between already defined buses.

addBranch!(system::PowerSystem; label, from, to, status, resistance, reactance,
    susceptance, turnsRatio, shiftAngle, longTerm, shortTerm, emergency,
    minAngleDifference, maxAngleDifference)

The keywords label, from, to, and one of the parameters resistance or reactance are mandatory. Default keyword values are set to zero, except for keywords status = 1, minAngleDifference = -2*pi, maxAngleDifference = 2*pi.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
addBus!(system; label = 2, active = 0.15, reactive = 0.08)

addBranch!(system; label = 1, from = 1, to = 2, resistance = 0.05, reactance = 0.12)

JuliaGrid.statusBranch! — Function

The function allows changing the operating status of the branch, from in-service to out-of-service, and vice versa.

statusBranch!(system::PowerSystem; label, status)

The keywords label should correspond to the already defined branch label.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
addBus!(system; label = 2, active = 0.15, reactive = 0.08)

addBranch!(system; label = 1, from = 1, to = 2, resistance = 0.05, reactance = 0.12)
statusBranch!(system; label = 1, status = 0)

JuliaGrid.parameterBranch! — Function

The function allows changing resistance, reactance, susceptance, turnsRatio and shiftAngle parameters of the branch.

parameterBranch!(system::PowerSystem; label, resistance, reactance, susceptance,
    turnsRatio, shiftAngle)

The keywords label should correspond to the already defined branch label. Keywords resistance, reactance, susceptance, turnsRatio or shiftAngle can be omitted, and then the value of the omitted parameter remains unchanged.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
addBus!(system; label = 2, active = 0.15, reactive = 0.08)

addBranch!(system; label = 1, from = 1, to = 2, resistance = 0.05, reactance = 0.12)
parameterBranch!(system; label = 1, susceptance = 0.062)

Generator Functions

Functions receive the composite type PowerSystem and arguments by keyword to set or change generator parameters. Further, functions update fields generator and bus.

JuliaGrid.addGenerator! — Function

The function adds a new generator. Names, descriptions and units of keywords are given in the table generator group. A generator can be added to an already-defined bus.

addGenerator!(system::PowerSystem; label, bus, area, status, active, reactive, magnitude,
    minActive, maxActive, minReactive, maxReactive, lowerActive, minReactiveLower,
    maxReactiveLower, upperActive, minReactiveUpper, maxReactiveUpper, loadFollowing,
    reserve10minute, reserve30minute, reactiveTimescale, activeModel, activeStartup,
    activeShutdown, activeDataPoint, activeCoefficient, reactiveModel, reactiveStartup,
    reactiveShutdown, reactiveDataPoint, reactiveCoefficient)

The keywords label and bus are mandatory. Default keyword values are set to zero, except for keywords status = 1, magnitude = 1.0, maxActive = Inf, minReactive = -Inf, maxReactive = Inf, activeModel = 2, activeDataPoint = 3, reactiveModel = 2, and reactiveDataPoint = 3.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)

addGenerator!(system; label = 1, bus = 1, active = 0.5 reactive = 0.1)

JuliaGrid.statusGenerator! — Function

The function allows changing the operating status of the generator, from in-service to out-of-service, and vice versa.

statusGenerator!(system::PowerSystem; label, status)

The keywords label should correspond to the already defined generator label.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)

addGenerator!(system; label = 1, bus = 1, active = 0.5 reactive = 0.1)
statusGenerator!(system; label = 1, status = 0)

JuliaGrid.outputGenerator! — Function

The function allows changing active and reactive output power of the generator.

outputGenerator!(system::PowerSystem; label, active, reactive)

The keywords label should correspond to the already defined generator label. Keywords active or reactive can be omitted, then the value of the omitted parameter remains unchanged.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)

addGenerator!(system; label = 1, bus = 1, active = 0.5 reactive = 0.1)
outputGenerator!(system; label = 1, active = 0.85)

Save Model

The function receives the composite type PowerSystem and arguments by keyword to set the path, reference and note of the data set.

JuliaGrid.savePowerSystem — Function

The function saves power system data in the HDF5 file using fields bus, branch, generator and basePower of the composite type PowerSystem.

savePowerSystem(system::PowerSystem; path, reference, note)

The keyword path::String is mandatory in the form "path/name.h5", while keywords reference::String and note::String are optional.

Example

system = powerSystem("case14.m")
savePowerSystem(system; path = "C:/case14.h5")

Build AC and DC Model

Functions receive the composite type PowerSystem and form vectors and matrices related to AC or DC simulations.

JuliaGrid.acModel! — Function

We advise the reader to read the section in-depth AC Model, which explains all the data involved in the field acModel.

acModel!(system::PowerSystem)

The function updates the field acModel. Once formed, the field will be automatically updated when using functions addBranch!(), shuntBus!(), statusBranch!(), parameterBranch!().

Example

system = powerSystem("case14.h5")
acModel!(system)

JuliaGrid.dcModel! — Function

We advise the reader to read the section in-depth DC Model, which explains all the data involved in the field dcModel.

dcModel!(system::PowerSystem)

The function updates the field dcModel. Once formed, the field will be automatically updated when using functions addBranch!(), statusBranch!(), parameterBranch!().

Example

system = powerSystem("case14.h5")
dcModel!(system)