Power System Model

The JuliaGrid supports the composite type PowerSystem to preserve power system data, with the following fields: bus, branch, generator, base, acModel, and dcModel. The fields bus, branch, and generator hold data related to buses, branches, and generators, respectively. The base field stores base values for power and voltages, with the default being three-phase power measured in volt-amperes (VA) for the base power and line-to-line voltages measured in volts (V) for base voltages. The macro command can be used to change the default unit settings for the base quantities of the composite type PowerSystem before it is created:

@unit

Finally, the acModel and dcModel fields store vectors and matrices calculated based on the power system's topology and parameters.

The composite type PowerSystem can be created using a function:

powerSystem().

JuliaGrid supports three modes for populating the PowerSystem type: using built-in functions, using HDF5 file format, and using Matpower case files. It is recommended to use the HDF5 format for large-scale systems. To facilitate this, JuliaGrid has the function:

savePowerSystem().

This function allows to save power systems that were either loaded from Matpower case files or created using built-in functions in the HDF5 format.

Once the PowerSystem type is created, you can add buses, branches, and generators using the following functions:

addBus!()

addBranch!()

addGenerator!().

In addition, it is possible to manipulate the parameters of buses, branches, and generators using the following functions:

shuntBus!()

statusBranch!()

parameterBranch!()

statusGenerator!()

outputGenerator!().

Executing these functions will automatically update all fields affected by them. You can also change other parameters of the power system by accessing and modifying the values in the bus, branch, generator, and base fields of the PowerSystem composite type. The input electrical quantities should be entered in per-units or radians, but this default setting can be altered using the macro @unit.

To create vectors and matrices based on the power system topology and parameters, you can use the following functions:

acModel!()

dcModel!().

Note that these functions can be executed at any time once all power system buses are defined. Specifically, using the addBranch!() function to add a new branch will automatically update the acModel and dcModel fields. However, adding a new bus using addBus!() requires executing the acModel!() and dcModel!() functions again. In addition, executing functions related to parameter manipulation of buses and branches will also automatically update the acModel and dcModel fields.

Build Model

JuliaGrid.powerSystem — Function

The function builds the composite type PowerSystem and populates bus, branch, generator and base fields. In general, once the composite type PowerSystem has been created, it is possible to add new buses, branches, or generators, or modify the parameters of existing ones.

To use the powerSystem() function, the path to the HDF5 file with the .h5 extension can be passed as arguments. For example:

powerSystem("pathToExternalData/name.h5")

Similarly, the path to the Matpower file with the .m extension can be passed to the same function:

powerSystem("pathToExternalData/name.m")

Alternatively, the PowerSystem composite type can be initialized by calling the function without any arguments:

powerSystem()

This allows the model to be built from scratch and modified as needed.

Example

system = powerSystem("case14.h5")

It is important to note that the prefixes of the base power and base voltage units can be adjusted using the macro @unit.

Example

@unit(base, MVA, kV)
system = powerSystem("case14.h5")

Save Model

JuliaGrid.savePowerSystem — Function

The function savePowerSystem, saves the power system's data in the HDF5 file using the fields bus, branch, generator, and base from the composite type PowerSystem.

savePowerSystem(system::PowerSystem; path, reference, note)

The location and file name of the HDF5 file is specified by the mandatory keyword path::String in the format of "path/name.h5". Additional information can be provided by the optional keywords reference::String and note::String.

Example

system = powerSystem("case14.m")
savePowerSystem(system; path = "D:/case14.h5")

Bus

JuliaGrid.addBus! — Function

The function adds a new bus to the PowerSystem type and updates its bus field.

addBus!(system::PowerSystem; label, active, reactive, conductance, susceptance,
    magnitude, angle, minMagnitude, maxMagnitude, base, area, lossZone)

The bus is defined with the following parameters:

label::Int64: a unique label for the bus

active::Float64: the active power demand at the bus

reactive::Float64: the reactive power demand at the bus

conductance::Float64: the active power demanded of the shunt element

susceptance::Float64: the reactive power injected of the shunt element

magnitude::Float64: the initial value of the voltage magnitude

angle::Float64: the initial value of the voltage angle

minMagnitude::Float64: the minimum voltage magnitude value

maxMagnitude::Float64: the maximum voltage magnitude value

base::Float64: the base value of the voltage magnitude

area::Int64: the area number

lossZone::Int64: the loss zone

The function automatically defines the bus as a demand bus, but it will be converted to a generator bus by using the addGenerator!() function or to a slack bus by using the slackBus!() function.

The input units are in per-units and radians by default, but they can be modified using the @unit macro.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)

JuliaGrid.slackBus! — Function

The function is used to set a slack bus, and it can also be used to dynamically change the slack bus. Every time the function is executed, the previous slack bus becomes a demand or generator bus, depending on whether the bus has a generator.

slackBus!(system::PowerSystem; label)

The label::Int64 keyword argument should correspond to the already defined bus label.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
slackBus!(system; label = 1)

JuliaGrid.shuntBus! — Function

The function allows changing conductance and susceptance parameters of the shunt element connected to the bus.

shuntBus!(system::PowerSystem; label, conductance, susceptance)

The keyword label::Int64 should correspond to the already defined bus label. Keywords conductance::Float64 or susceptance::Float64 can be omitted, then the value of the omitted parameter remains unchanged.

The usefulness of the function is that its execution automatically updates the field acModel. That is, when changing these parameters, it is not necessary to create this model from scratch.

The input units are in per-units by default, but they can be modified using the @unit macro.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
shuntBus!(system; label = 1, conductance = 0.04)