Power System Model

The JuliaGrid supports the composite type PowerSystem to preserve power system data, with the following fields: bus, branch, generator, base, acModel, and dcModel. The fields bus, branch, and generator hold data related to buses, branches, and generators, respectively. The base field stores base values for power and voltages, with the default being three-phase power measured in volt-amperes (VA) for the base power and line-to-line voltages measured in volts (V) for base voltages. The macro command @base can be used to change the default unit settings for the base quantities of the composite type PowerSystem before it is created. Finally, the acModel and dcModel fields store vectors and matrices calculated based on the power system's topology and parameters.

The composite type PowerSystem can be created using a function:

powerSystem().

JuliaGrid supports three modes for populating the PowerSystem type: using built-in functions, using HDF5 file format, and using Matpower case files. It is recommended to use the HDF5 format for large-scale systems. To facilitate this, JuliaGrid has the function:

savePowerSystem().

This function allows to save power systems that were either loaded from Matpower case files or created using built-in functions in the HDF5 format.

Once the PowerSystem type is created, you can add buses, branches, and generators using the following functions:

addBus!()

addBranch!()

addGenerator!().

In addition, it is possible to manipulate the parameters of buses, branches, and generators using the following functions:

slackBus!()

shuntBus!()

statusBranch!()

parameterBranch!()

statusGenerator!()

outputGenerator!().

Executing these functions will automatically update all fields affected by them. You can also change other parameters of the power system by accessing and modifying the values in the bus, branch, generator, and base fields of the PowerSystem composite type. The input electrical quantities should be entered in per-units or radians, but this default setting can be altered using the following macros @power, @voltage, and @parameter.

To create vectors and matrices based on the power system topology and parameters, you can use the following functions:

acModel!()

dcModel!().

Note that these functions can be executed at any time once all power system buses are defined. Specifically, using the addBranch!() function to add a new branch will automatically update the acModel and dcModel fields. However, adding a new bus using addBus!() requires executing the acModel!() and dcModel!() functions again. In addition, executing functions related to parameter manipulation of buses and branches will also automatically update the acModel and dcModel fields.

Build Model

JuliaGrid.powerSystem — Function

The function builds the composite type PowerSystem and populates bus, branch, generator and base fields. The function can be used by passing the path to the HDF5 file with the .h5 extension or a Matpower file with the .m extension as an argument.

powerSystem("pathToExternalData/name.extension")

In general, once the composite type PowerSystem has been created, it is possible to add new buses, branches, or generators, or modify the parameters of existing ones.

Units

JuliaGrid stores all data in per-unit (pu) and radian (rad) format which are fixed, the exceptions are base values in volt-ampere (VA) and volt (V) which can be changed using the macro @base.

Example

system = powerSystem("case14.h5")

Alternatively, the PowerSystem composite type can be initialized by calling the function without any arguments.

powerSystem()

This allows the model to be built from scratch and modified as needed.

Save Model

JuliaGrid.savePowerSystem — Function

The function saves the power system's data in the HDF5 file using the fields bus, branch, generator, and base from the PowerSystem composite type.

savePowerSystem(system::PowerSystem; path, reference, note)

The location and file name of the HDF5 file is specified by the mandatory keyword path::String in the format of "path/name.h5". Additional information can be provided by the optional keywords reference::String and note::String.

Example

system = powerSystem("case14.m")
savePowerSystem(system; path = "D:/case14.h5")

Bus Functions

JuliaGrid.addBus! — Function

The function adds a new bus to the PowerSystem type, updating its bus field. It automatically sets the bus as a demand bus, but it can be changed to a generator bus using the addGenerator!() function or to a slack bus using the slackBus!() function.

addBus!(system::PowerSystem; label, active, reactive, conductance, susceptance, 
    magnitude, angle, minMagnitude, maxMagnitude, base, area, lossZone)

The bus is defined with the following parameters:

label: a unique label for the bus

active (pu or W): the active power demand at the bus

reactive (pu or VAr): the reactive power demand at the bus

conductance (pu or W): the active power demanded of the shunt element

susceptance (pu or VAr): the reactive power injected of the shunt element

magnitude (pu or V): the initial value of the voltage magnitude

angle (rad or deg): the initial value of the voltage angle given in rad or deg

minMagnitude (pu or V): the minimum voltage magnitude value

maxMagnitude (pu or V): the maximum voltage magnitude value

base (V): the base value of the voltage magnitude

area: the area number

lossZone: the loss zone

Units

The input units are in per-unit (pu) and radian (rad) by default as shown, except for the keyword base which is given by default in volt (V). The unit and prefix settings can be modified using the macros @base, @power, @voltage, and @parameter.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)

JuliaGrid.slackBus! — Function

The function is used to set a slack bus, and it can also be used to dynamically change the slack bus.

slackBus!(system::PowerSystem; label)

Every time the function is executed, the previous slack bus becomes a demand or generator bus, depending on whether the bus has a generator. The bus label should be specified using the label keyword argument and should match an already defined bus label.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
slackBus!(system; label = 1)

JuliaGrid.shuntBus! — Function

This function enables the modification of the conductance and susceptance parameters of a shunt element connected to a bus.

shuntBus!(system::PowerSystem; label, conductance, susceptance)

The label keyword must match an existing bus label. If either conductance or susceptance is left out, the corresponding value will remain unchanged. Additionally, this function automatically updates the acModel field, eliminating the need to rebuild the model from scratch when making changes to these parameters.

Units

The input units are in per-units by default, but they can be modified using the @power macro.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
shuntBus!(system; label = 1, conductance = 0.04)

Branch Functions

JuliaGrid.addBranch! — Function

The function adds a new branch to the PowerSystem type and updates its branch field. A branch can be added between already defined buses.

addBranch!(system::PowerSystem; label, from, to, status, resistance, reactance, 
    susceptance, turnsRatio, shiftAngle, longTerm, shortTerm, emergency, 
    minDiffAngle, maxDiffAngle)

The branch is defined with the following parameters:

label: unique branch label

from: from bus label, corresponds to the bus label

to: to bus label, corresponds to the bus label

status: operating status of the branch, in-service = 1, out-of-service = 0

resistance (pu or Ω): branch resistance

reactance (pu or Ω): branch reactance

susceptance (pu or S): total line charging susceptance

turnsRatio: transformer off-nominal turns ratio, equal to zero for a line

shiftAngle (rad or deg): transformer phase shift angle, where positive value defines delay

longTerm (pu or VA): short-term rating (equal to zero for unlimited)

shortTerm (pu or VA): long-term rating (equal to zero for unlimited)

emergency (pu or VA): emergency rating (equal to zero for unlimited)

minDiffAngle (rad or deg): minimum voltage angle difference value between from and to bus

maxDiffAngle (rad or deg): maximum voltage angle difference value between from and to bus

Units

The input units are in per-units and radians by default, but they can be modified using the following macros @power, @voltage, and @parameter.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
addBus!(system; label = 2, active = 0.15, reactive = 0.08)

addBranch!(system; label = 1, from = 1, to = 2, resistance = 0.05, reactance = 0.12)

JuliaGrid.statusBranch! — Function

The function enables the switching of the operational status of a branch, identified by its label, within the PowerSystem system between in-service and out-of-service.

statusBranch!(system::PowerSystem; label, status)

This function updates the acModel and dcModel fields automatically when the operating status of a branch is changed, thus eliminating the need to rebuild the model from scratch.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
addBus!(system; label = 2, active = 0.15, reactive = 0.08)

addBranch!(system; label = 1, from = 1, to = 2, resistance = 0.05, reactance = 0.12)
statusBranch!(system; label = 1, status = 0)

JuliaGrid.parameterBranch! — Function

This function enables the alteration of the resistance, reactance, susceptance, turnsRatio and shiftAngle parameters of a branch, identified by its label, within the PowerSystem.

parameterBranch!(system::PowerSystem; label, resistance, reactance, susceptance, 
    turnsRatio, shiftAngle)

If any of these parameters are omitted, their current values will be retained. Additionally, this function updates the acModel and dcModel fields automatically, removing the need to rebuild the model from scratch.

Units

The input units are in per-units by default, but they can be modified using the following macros @parameter and @voltage.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)
addBus!(system; label = 2, active = 0.15, reactive = 0.08)

addBranch!(system; label = 1, from = 1, to = 2, resistance = 0.05, reactance = 0.12)
parameterBranch!(system; label = 1, susceptance = 0.062)

Generator Functions

JuliaGrid.addGenerator! — Function

The function is used to add a new generator to the PowerSystem type and update its generator field. The generator can be added to an already defined bus.

addGenerator!(system::PowerSystem; label, bus, status, active, reactive, magnitude,
    minActive, maxActive, minReactive, maxReactive, lowActive, minLowReactive, 
    maxLowReactive, upActive, minUpReactive, maxUpReactive,
    loadFollowing, reactiveTimescale, reserve10min, reserve30min, area)

The generator is defined with the following parameters:

label: a unique label for the generator

bus: the label of the bus to which the generator is connected

status: the operating status of the generator, in-service = 1, out-of-service = 0

active (pu or W): output active power

reactive (pu or VAr): output reactive power

magnitude (pu or V): voltage magnitude setpoint

minActive (pu or W): minimum allowed output active power value

maxActive (pu or W): maximum allowed output active power value

minReactive (pu or VAr): minimum allowed output reactive power value

maxReactive (pu or VAr): maximum allowed output reactive power value

lowActive (pu or W): lower allowed active power output value of PQ capability curve

minLowReactive (pu or VAr): minimum allowed reactive power output value at lowActive value

maxLowReactive (pu or VAr): maximum allowed reactive power output value at lowActive value

upActive (pu or W): upper allowed active power output value of PQ capability curve

minUpReactive (pu or VAr): minimum allowed reactive power output value at upActive value

maxUpReactive (pu or VAr): maximum allowed reactive power output value at upActive value

loadFollowing (pu or W per min): ramp rate for load following/AG

reserve10min (pu or W): ramp rate for 10-minute reserves

reserve30min (pu or W): ramp rate for 30-minute reserves

reactiveTimescale (pu or VAr per min): ramp rate for reactive power, two seconds timescale

area: area participation factor

Units

The input units are realted to per-units by default, but they can be modified using macros @power and @voltage.

Example

system = powerSystem()

addBus!(system; label = 1, active = 0.25, reactive = -0.04)

addGenerator!(system; label = 1, bus = 1, active = 0.5, reactive = 0.1)

Units

JuliaGrid.@base — Macro

By default, the units for base power and base voltages are set to volt-ampere (VA) and volt (V), but you can modify the prefixes using the macro:

@base(power, voltage)

Prefixes must be specified according to the SI prefixes and should be included with the unit of power (VA) or unit of voltage (V). Keep in mind that the macro must be used before creating the composite type PowerSystem.

Example

@base(MVA, kV)
system = powerSystem("case14.h5")

JuliaGrid.@power — Macro

JuliaGrid stores all data related with powers in per-units, and these cannot be altered. However, the power units of the built-in functions used to add or modified power system elements can be modified using the macro:

@power(active, reactive, apparent)

Prefixes must be specified according to the SI prefixes and should be included with the unit of active power (W), reactive power (VAr), or apparent power (VA). Also it is a possible to combine SI units with/without prefixes with per-units (pu).

Changing the unit of active power is reflected in the following quantities:

addBus!(): active, conductance

shuntBus!(): conductance

addGenerator!(): active, minActive, maxActive, lowActive, upActive, loadFollowing, reserve10min, reserve30min

addActiveCost!(): piecewise, polynomial.

Changing the unit of reactive power unit is reflected in the following quantities:

addBus!(): reactive, susceptance

shuntBus!(): susceptance

addGenerator!(): reactive, minReactive, maxReactive, minLowReactive, maxLowReactive, minUpReactive, maxUpReactive, reactiveTimescale

addReactiveCost!(): piecewise, polynomial.

Changing the unit of apparent power unit is reflected in the following quantities:

addBranch!(): longTerm, shortTerm, emergency.

Example

@power(MW, kVAr, VA)

JuliaGrid.@voltage — Macro

JuliaGrid stores all data related with voltages in per-units and radians, and these cannot be altered. However, the voltage magnitude and angles units of the built-in functions used to add or modified power system elements can be modified using the macro:

@voltage(magnitude, angle)

Prefixes must be specified according to the SI prefixes and should be included with the unit of voltage magnitude (V). The second option is to define the unit of voltage magnitude in per-unit (pu). The unit of the voltage angle should be given in radian (rad) or degree (deg).

Changing the unit of voltage magnitude is reflected in the following quantities:

addBus!(): magnitude, minMagnitude, maxMagnitude

addGenerator!(): magnitude.

Changing the unit of voltage angle is reflected in the following quantities:

addBus!(): angle, susceptance

addBranch!(): shiftAngle, minDiffAngle, maxDiffAngle

parameterBranch!(): shiftAngle

Example

@voltage(kV, deg)

JuliaGrid.@parameter — Macro

JuliaGrid stores all data related with impedances and admittancies in per-units, and these cannot be altered. However, units of impedance and admittance of the built-in functions used to add or modified power system elements can be modified using the macro:

@parameter(impedance, admittance)

Prefixes must be specified according to the SI prefixes and should be included with the unit of impedance (Ω) or unit of admittance (S). The second option is to define the units in per-unit (pu).

In the case where impedance and admittance are being used in SI units (Ω and S) and these units are related to the transformer, the assignment must be based on the primary side of the transformer.

Changing the units of impedance is reflected in the following quantities in specific functions:

addBranch!(): resistance, reactance

parameterBranch!(): resistance, reactance

Changing the units of admittance is reflected in the following quantities:

addBranch!(): susceptance

parameterBranch!(): susceptance

Example

@voltage(Ω, pu)